### 2016-07-24 #44 @whedon
[REVIEW]: Elektra: universal framework to access configuration parameters
**Submitting author:** @markus2330 (<a href="http://orcid.org/0000-0002-1493-9065">Markus Raab</a>)
**Repository:** <a href="https://github.com/ElektraInitiative/libelektra" target ="_blank">https://github.com/ElektraInitiative/libelektra</a>
**Version:** 0.8.17
