### 2016-07-25 af52d23
Shell recorder: fix grep -z
BSD systems do not have the grep -z option, so all newlines must be removed from the input
