### 2016-07-25 281a88e
Shell Recorder: missed one mktemp
