### 2016-07-25 37943f1
Shell recorder: seems that mac os x grep does not have a --text
