+++ tests/shell/shell_recorder/shell_recorder.sh
@@ -121,7 +121,7 @@ execute()
     if [ ! -z "$STDERRCMP" ];
     then
         nbTest=$(( nbTest + 1 ))
-        echo "$STDERR" | sed -e ':a' -e 'N' -e '$!ba' -e 's/\n//g' | grep -Eq --text "$STDERRCMP"
+        echo "$STDERR" | sed -e ':a' -e 'N' -e '$!ba' -e 's/\n//g' | grep -Eqa "$STDERRCMP"
         if [ "$?" -ne "0" ];
         then
             echo "STDERR doesn't match $STDERRCMP"
@@ -138,7 +138,7 @@ execute()
     if [ ! -z "$STDOUTCMP" ];
     then
         nbTest=$(( nbTest + 1 ))
-        echo "$STDOUT" | sed -e ':a' -e 'N' -e '$!ba' -e 's/\n//g' | grep -Eq --text "$STDOUTCMP"
+        echo "$STDOUT" | sed -e ':a' -e 'N' -e '$!ba' -e 's/\n//g' | grep -Eqa "$STDOUTCMP"
         if [ "$?" -ne "0" ];
         then
             echo "STDOUT doesn't match $STDOUTCMP"
@@ -155,7 +155,7 @@ execute()
     if [ ! -z "$WARNINGSCMP" ];
     then
         nbTest=$(( nbTest + 1 ))
-        echo "$WARNINGS" | sed -e ':a' -e 'N' -e '$!ba' -e 's/\n//g' | grep -Eq --text "($WARNINGSCMP)"
+        echo "$WARNINGS" | sed -e ':a' -e 'N' -e '$!ba' -e 's/\n//g' | grep -Eqa "($WARNINGSCMP)"
         if [ "$?" -ne "0" ];
         then
             echo "WARNINGS doesn't match $WARNINGSCMP"
@@ -174,7 +174,7 @@ execute()
     if [ ! -z "$ERRORSCMP" ];
     then
         nbTest=$(( nbTest + 1 ))
-        echo "$ERRORS" | sed -e ':a' -e 'N' -e '$!ba' -e 's/\n//g' | grep -Eq --text "($ERRORSCMP)"
+        echo "$ERRORS" | sed -e ':a' -e 'N' -e '$!ba' -e 's/\n//g' | grep -Eaq "($ERRORSCMP)"
         if [ "$?" -ne "0" ];
         then
             echo "ERRORS doesn't match $ERRORSCMP"
@@ -189,7 +189,7 @@ execute()
     if [ ! -z "$DIFFCMP" ];
     then
         nbTest=$(( nbTest + 1 ))
-        echo "$DIFF" | sed -e ':a' -e 'N' -e '$!ba' -e 's/\n//g' | grep -Eq --text "($DIFFCMP)"
+        echo "$DIFF" | sed -e ':a' -e 'N' -e '$!ba' -e 's/\n//g' | grep -Eqa "($DIFFCMP)"
         if [ "$?" -ne "0" ];
         then
             echo "Changes to $DBFile don't match $DIFFCMP"
