### 2016-07-25 086b1ac
Build: Only generate `kdberrors.h` once
Before this commit multiple targets would generate `kdberrors.h` in
parallel. Therefore it was possible that one build process would
overwrite `kdberrors.h` – with exactly the same content as before –
while another process was reading the file.

This commit fixes #674.
