+++ src/libs/getenv/src/CMakeLists.txt
@@ -8,6 +8,7 @@ file (GLOB_RECURSE SRC_FILES *.cpp)
 set (SOURCES ${SRC_FILES} ${HDR_FILES})
 
 add_library (elektragetenv SHARED ${SOURCES})
+add_dependencies (elektragetenv kdberrors_generated)
 
 target_link_libraries (elektragetenv ${CMAKE_DL_LIBS})
 target_link_libraries (elektragetenv elektra-kdb elektra-meta) # must be shared!
