+++ benchmarks/CMakeLists.txt
@@ -6,6 +6,7 @@ macro (do_benchmark source)
 	include_directories ("${CMAKE_CURRENT_SOURCE_DIR}")
 	set (SOURCES ${HDR_FILES} benchmarks.c benchmarks.h ${source}.c)
 	add_executable (${source} ${SOURCES})
+	add_dependencies (${source} kdberrors_generated)
 
 	target_link_elektra(${source} elektra-kdb elektra-meta)
 
