### 2016-07-25 637e86c
OPMPHM: vstack, vheap add links.
