### 2016-07-25 de342b1
OPMPHM: add group for doxygen + add trie to group
