### 2016-07-25 ca42b3d
crypto-plugin: fix unused parameter in elektraCryptoOpenSSLInit
See #840 for discussion
