### 2016-07-26 f19fb2b
Shell Recorder: Remove useless whitespace
