### 2016-07-26 80793d2
Shell Recorder: Support paths containing spaces
Before this change shell recorder scripts would fail, if `kdb` was
located in a path containing spaces in its name.
