### 2016-07-26 b047c8a
Revert "Shell recorder: seems that mac os x grep does not have a --text"
This reverts commit 37943f15a8188161ca8d2c024e35a2cd52a5e694.
