### 2016-07-26 a37f9f8
Revert "Shell recorder: fix grep -z"
This reverts commit af52d236ac657e6224ed1c0e9c8cb03094fd5c5b.
