### 2016-07-27 ab772e8
Shell Recorder: Improve portability of code
Before this change we used the GNU grep options `-z` and `-P`. These
options are not portable since BSD's grep does not offer the
functionality provided by them.

This commit fixes #858.
