### 2016-07-27 698deb7
Shell Recorder: File: mktemp feature
setting File to mktemp provides a fresh database temp file for ever run
