### 2016-07-27 9d51be1
Shell recorder: If File: empty use temp file
