+++ tests/shell/shell_recorder/shell_recorder.sh
@@ -215,7 +215,7 @@ run_script()
                 ;;
             File:)
                 DBFile=$(echo $line|cut -d ' ' -f2)
-                if [ "$DBFile" = "mktemp" ]; then
+                if [ "$DBFile" = "File:" ]; then
 			DBFile=$(mktemp -t elektraenv.XXXXXXXXX 2>/dev/null || mktemp -t 'elektraenv')
                 fi
                 ;;
