### 2016-07-27 2ccdad8
Shell Recorder: mktemp
