### 2016-07-27 60fc7c7
cv: add testcase + deactivate
