### 2016-07-27 7b91d97
ruby: first version of ruby bindings
This is still a VERY early stage
