### 2016-07-29 eaa953a
doc: draft dod
