### 2016-07-29 1c054a7
Shell Recorder: Remove useless whitespace
