### 2016-07-29 2537ab5
Shell Recorder: Only replace newline symbols
The shell recorder tests, after commit ab772e8c, worked only because
all the test outputs do not contain the `NULL` symbol.
