+++ tests/shell/shell_recorder/shell_recorder.sh
@@ -30,8 +30,8 @@ else
         KDBCOMMAND="kdb"
 fi
 
-replace_newline_null () {
-	tr '\n' '⏎' | tr '\000' '\n'
+replace_newline_return () {
+	tr '\n' '⏎'
 }
 
 execute()
@@ -125,7 +125,7 @@ execute()
     if [ ! -z "$STDERRCMP" ];
     then
         nbTest=$(( nbTest + 1 ))
-        echo "$STDERR" | replace_newline_null | grep -Eq --text "$STDERRCMP"
+        echo "$STDERR" | replace_newline_return | grep -Eq --text "$STDERRCMP"
         if [ "$?" -ne "0" ];
         then
             echo "STDERR doesn't match $STDERRCMP"
@@ -142,7 +142,7 @@ execute()
     if [ ! -z "$STDOUTCMP" ];
     then
         nbTest=$(( nbTest + 1 ))
-        echo "$STDOUT" | replace_newline_null | grep -Eq --text "$STDOUTCMP"
+        echo "$STDOUT" | replace_newline_return | grep -Eq --text "$STDOUTCMP"
         if [ "$?" -ne "0" ];
         then
             echo "STDOUT doesn't match $STDOUTCMP"
@@ -159,7 +159,7 @@ execute()
     if [ ! -z "$WARNINGSCMP" ];
     then
         nbTest=$(( nbTest + 1 ))
-        echo "$WARNINGS" | replace_newline_null | grep -Eq --text "($WARNINGSCMP)"
+        echo "$WARNINGS" | replace_newline_return | grep -Eq --text "($WARNINGSCMP)"
         if [ "$?" -ne "0" ];
         then
             echo "WARNINGS doesn't match $WARNINGSCMP"
@@ -178,7 +178,7 @@ execute()
     if [ ! -z "$ERRORSCMP" ];
     then
         nbTest=$(( nbTest + 1 ))
-        echo "$ERRORS" | replace_newline_null | grep -Eq --text "($ERRORSCMP)"
+        echo "$ERRORS" | replace_newline_return | grep -Eq --text "($ERRORSCMP)"
         if [ "$?" -ne "0" ];
         then
             echo "ERRORS doesn't match $ERRORSCMP"
@@ -193,7 +193,7 @@ execute()
     if [ ! -z "$DIFFCMP" ];
     then
         nbTest=$(( nbTest + 1 ))
-        echo "$DIFF" | replace_newline_null | grep -Eq --text "($DIFFCMP)"
+        echo "$DIFF" | replace_newline_return | grep -Eq --text "($DIFFCMP)"
         if [ "$?" -ne "0" ];
         then
             echo "Changes to $DBFile don't match $DIFFCMP"
