### 2016-07-29 f7087c2
Shell Recorder: Extend documentation
