### 2016-07-30 fa59012
Shell Recorder: Use `awk` instead of `tr`
The GNU version of `tr` does not handle multi-byte characters such as
`⏎` correctly. This means that Shell Recorder tests that match line
ending characters (`⏎`) did not work in GNU/Linux before this change.
