### 2016-07-30 35a87d7
crypto-plugin: add password generating function
Preparation for generating the master password.
A PRNG is used to generate byte sequences which can be used as master passwords
