+++ src/plugins/crypto/openssl_operations.h
@@ -11,10 +11,12 @@
 #define ELEKTRA_PLUGIN_LIBCRYPTO_OPERATIONS_H
 
 #include <kdb.h>
+#include <kdbtypes.h>
 
 #define ELEKTRA_CRYPTO_SSL_KEYSIZE (32)
 #define ELEKTRA_CRYPTO_SSL_BLOCKSIZE (16)
 
+char * elektraCryptoOpenSSLCreateRandomString (const kdb_unsigned_short_t length);
 int elektraCryptoOpenSSLInit (Key * errorKey);
 int elektraCryptoOpenSSLHandleCreate (elektraCryptoHandle ** handle, KeySet * config, Key * errorKey);
 void elektraCryptoOpenSSLHandleDestroy (elektraCryptoHandle * handle);
