+++ src/plugins/crypto/botan_operations.h
@@ -11,11 +11,13 @@
 #define ELEKTRA_PLUGIN_BOTAN_OPERATIONS_H
 
 #include <kdb.h>
+#include <kdbtypes.h>
 
 #define ELEKTRA_CRYPTO_BOTAN_KEYSIZE (32)
 #define ELEKTRA_CRYPTO_BOTAN_BLOCKSIZE (16)
 #define ELEKTRA_CRYPTO_BOTAN_ALGORITHM "AES-256/CBC"
 
+char * elektraCryptoBotanCreateRandomString (const kdb_unsigned_short_t length);
 int elektraCryptoBotanInit (Key * errorKey);
 int elektraCryptoBotanEncrypt (KeySet * pluginConfig, Key * k, Key * errorKey);
 int elektraCryptoBotanDecrypt (KeySet * pluginConfig, Key * k, Key * errorKey);
