### 2016-07-31 44260dc
getenv: further context example
