### 2016-08-01 40ad721
logger: first draft
