### 2016-08-01 27956f0
first logger draft
