### 2016-08-01 c8e790e
crypto-plugin: fix password generation in crypto_gcrypt variant
