### 2016-08-01 b7f4385
crypto-plugin: add checkconf function
Prepares the random password generation in the checkconf function.
The module for calling the GPG binary is still missing.
