+++ src/plugins/crypto/crypto.h
@@ -20,6 +20,7 @@ enum ElektraCryptoHeaderFlags
 	ELEKTRA_CRYPTO_FLAG_NULL = 2
 };
 
+#define ELEKTRA_CRYPTO_PARAM_MASTER_PWD "/crypto/masterpassword"
 #define ELEKTRA_CRYPTO_PARAM_KEY_PATH ("/crypto/key")
 #define ELEKTRA_CRYPTO_PARAM_IV_PATH ("/crypto/iv")
 #define ELEKTRA_CRYPTO_PARAM_SHUTDOWN ("/shutdown")
@@ -62,6 +63,7 @@ int CRYPTO_PLUGIN_FUNCTION (open) (Plugin * handle, Key * errorKey);
 int CRYPTO_PLUGIN_FUNCTION (close) (Plugin * handle, Key * errorKey);
 int CRYPTO_PLUGIN_FUNCTION (get) (Plugin * handle, KeySet * ks, Key * parentKey);
 int CRYPTO_PLUGIN_FUNCTION (set) (Plugin * handle, KeySet * ks, Key * parentKey);
+int CRYPTO_PLUGIN_FUNCTION (checkconf) (Key * errorKey, KeySet * conf);
 int CRYPTO_PLUGIN_FUNCTION (error) (Plugin * handle, KeySet * ks, Key * parentKey);
 
 Plugin * ELEKTRA_PLUGIN_EXPORT (crypto);
