### 2016-08-02 d174be2
crypto-plugin: add error specification
Add error specification for random password generation fault.
