### 2016-08-02 25931d5
crypto-plugin: make password length configurable
The length of the random master password can be provided in /crypto/masterpasswordlength within the plugin or backend config.
