+++ src/plugins/passwd/hasfgetpwent.c
@@ -1,0 +1,7 @@
+#include <pwd.h>
+
+int main (int argc, char const * argv[])
+{
+	fgetpwent (0);
+	return 0;
+}
