+++ src/plugins/passwd/passwd.c
@@ -1,0 +1,215 @@
+/**
+ * @file
+ *
+ * @brief Source for passwd plugin
+ *
+ * @copyright BSD License (see doc/COPYING or http://www.libelektra.org)
+ *
+ */
+
+#include "passwd.h"
+
+#include <kdbhelper.h>
+
+#include <pwd.h>
+#include <stdio.h>
+#include <stdlib.h>
+#include <string.h>
+
+#define ID_MAX_CHARACTERS 11
+
+typedef enum {
+	NAME,
+	UID,
+} SortBy;
+
+static KeySet * pwentToKS (struct passwd * pwd, Key * parentKey, SortBy index)
+{
+	KeySet * ks = ksNew (0, KS_END);
+	Key * append = keyNew (keyName (parentKey), KEY_END);
+	char id[ID_MAX_CHARACTERS];
+	if (index == UID)
+	{
+		snprintf (id, sizeof (id), "%u", pwd->pw_uid);
+		keyAddBaseName (append, id);
+		keySetBinary (append, 0, 0);
+		ksAppendKey (ks, keyDup (append));
+		keyAddBaseName (append, "name");
+		keySetString (append, pwd->pw_name);
+	}
+	else
+	{
+		keyAddBaseName (append, pwd->pw_name);
+		keySetBinary (append, 0, 0);
+		ksAppendKey (ks, keyDup (append));
+		snprintf (id, sizeof (id), "%u", pwd->pw_uid);
+		keyAddBaseName (append, "uid");
+		keySetString (append, id);
+	}
+	ksAppendKey (ks, keyDup (append));
+	keySetString (append, 0);
+	keySetBaseName (append, "shell");
+	keySetString (append, pwd->pw_shell);
+	ksAppendKey (ks, keyDup (append));
+	keySetString (append, 0);
+	keySetBaseName (append, "home");
+	keySetString (append, pwd->pw_dir);
+	ksAppendKey (ks, keyDup (append));
+	keySetString (append, 0);
+	keySetBaseName (append, "gid");
+	snprintf (id, sizeof (id), "%u", pwd->pw_gid);
+	keySetString (append, id);
+	ksAppendKey (ks, keyDup (append));
+	keySetString (append, 0);
+	keySetBaseName (append, "passwd");
+	keySetString (append, pwd->pw_passwd);
+	ksAppendKey (ks, keyDup (append));
+	keySetString (append, 0);
+	keySetBaseName (append, "gecos");
+	keySetString (append, pwd->pw_gecos);
+	ksAppendKey (ks, keyDup (append));
+	keyDel (append);
+	return ks;
+}
+
+int elektraPasswdGet (Plugin * handle ELEKTRA_UNUSED, KeySet * returned ELEKTRA_UNUSED, Key * parentKey ELEKTRA_UNUSED)
+{
+	if (!elektraStrCmp (keyName (parentKey), "system/elektra/modules/passwd"))
+	{
+		KeySet * contract =
+			ksNew (30, keyNew ("system/elektra/modules/passwd", KEY_VALUE, "passwd plugin waits for your orders", KEY_END),
+			       keyNew ("system/elektra/modules/passwd/exports", KEY_END),
+			       keyNew ("system/elektra/modules/passwd/exports/get", KEY_FUNC, elektraPasswdGet, KEY_END),
+			       keyNew ("system/elektra/modules/passwd/exports/set", KEY_FUNC, elektraPasswdSet, KEY_END),
+#include ELEKTRA_README (passwd)
[…]
