### 2016-08-03 c89568c
passwd: error messages + fixed shell tests
