+++ src/plugins/passwd/testmod_passwd.c
@@ -14,31 +14,6 @@
 
 #include <tests_plugin.h>
 
-static void test_basics ()
-{
-	printf ("test basics\n");
-
-	Key * parentKey = keyNew ("user/tests/passwd", KEY_END);
-	KeySet * conf = ksNew (0, KS_END);
-	PLUGIN_OPEN ("passwd");
-
-	KeySet * ks = ksNew (0, KS_END);
-
-	succeed_if (plugin->kdbOpen (plugin, parentKey) == 1, "call to kdbOpen was not successful");
-
-	succeed_if (plugin->kdbGet (plugin, ks, parentKey) == 1, "call to kdbGet was not successful");
-
-	succeed_if (plugin->kdbSet (plugin, ks, parentKey) == 1, "call to kdbSet was not successful");
-
-	succeed_if (plugin->kdbError (plugin, ks, parentKey) == 1, "call to kdbError was not successful");
-
-	succeed_if (plugin->kdbClose (plugin, parentKey) == 1, "call to kdbClose was not successful");
-
-	keyDel (parentKey);
-	ksDel (ks);
-	PLUGIN_CLOSE ();
-}
-
 
 int main (int argc, char ** argv)
 {
@@ -46,8 +21,6 @@ int main (int argc, char ** argv)
 	printf ("==================\n\n");
 
 	init (argc, argv);
-
-	test_basics ();
 
 	printf ("\ntestmod_passwd RESULTS: %d test(s) done. %d error(s).\n", nbTest, nbError);
 
