### 2016-08-03 d9a7727
global plugins: make positions explicit (init, once, foreach, deinit)
