+++ src/libs/elektra/kdb.c
@@ -450,6 +450,8 @@ int kdbClose (KDB * handle, Key * errorKey)
  */
 static int elektraGetCheckUpdateNeeded (Split * split, Key * parentKey)
 {
+	// GLOBAL: getresolver [init]
+	// GLOBAL: getresolver [max once]
 	int updateNeededOccurred = 0;
 	for (size_t i = 0; i < split->size; i++)
 	{
@@ -459,7 +461,7 @@ static int elektraGetCheckUpdateNeeded (Split * split, Key * parentKey)
 
 		if (backend->getplugins[RESOLVER_PLUGIN] && backend->getplugins[RESOLVER_PLUGIN]->kdbGet)
 		{
-			// GLOBAL: getresolver
+			// GLOBAL: getresolver [foreach]
 			ksRewind (split->keysets[i]);
 			keySetName (parentKey, keyName (split->parents[i]));
 			keySetString (parentKey, "");
@@ -490,6 +492,7 @@ static int elektraGetCheckUpdateNeeded (Split * split, Key * parentKey)
 			return -1;
 		}
 	}
+	// GLOBAL: getresolver [deinit]
 	return updateNeededOccurred;
 }
 
@@ -568,6 +571,13 @@ static int elektraGetDoUpdateWithGlobalHooks (KDB * handle, Split * split, KeySe
 	const int bypassedSplits = 1;
 	int pgs_done = 0;
 	int pgc_done = 0;
+
+	// GLOBAL: getstorage [init]
+	// GLOBAL: getstorage [max once]
+
+	// GLOBAL: postgetstorage [init]
+	// GLOBAL: postgetstorage [max once]
+
 	for (size_t i = 0; i < split->size - bypassedSplits; i++)
 	{
 		if (!test_bit (split->syncbits[i], SPLIT_FLAG_SYNC))
@@ -593,6 +603,10 @@ static int elektraGetDoUpdateWithGlobalHooks (KDB * handle, Split * split, KeySe
 		for (int p = start; p < end; ++p)
 		{
 			int ret = 0;
+
+			// GLOBAL: getstorage [foreach]
+			// GLOBAL: postgetstorage [foreach]
+
 			if (!pgs_done && (p == (STORAGE_PLUGIN + 1)) && handle->globalPlugins[POSTGETSTORAGE])
 			{
 				pgs_done = 1;
@@ -629,10 +643,14 @@ static int elektraGetDoUpdateWithGlobalHooks (KDB * handle, Split * split, KeySe
 			{
 				// Ohh, an error occurred,
 				// lets stop the process.
+				// GLOBAL: getstorage [deinit]
+				// GLOBAL: postgetstorage [deinit]
 				return -1;
 			}
 		}
 	}
+	// GLOBAL: getstorage [deinit]
+	// GLOBAL: postgetstorage [deinit]
 	return 0;
 }
 
@@ -773,7 +791,10 @@ int kdbGet (KDB * handle, KeySet * ks, Key * parentKey)
 		ELEKTRA_SET_ERROR (37, parentKey, "handle or ks null pointer");
 		goto error;
 	}
-	// GLOBAL: pregetstorage
+	// GLOBAL: pregetstorage [init]
+	// GLOBAL: pregetstorage [max once]
+	// GLOBAL: pregetstorage [foreach]
+	// GLOBAL: pregetstorage [deinit]
 	if (handle->globalPlugins[PREGETSTORAGE])
 	{
 		handle->globalPlugins[PREGETSTORAGE]->kdbGet (handle->globalPlugins[PREGETSTORAGE], ks, parentKey);
@@ -813,7 +834,6 @@ int kdbGet (KDB * handle, KeySet * ks, Key * parentKey)
 		goto error;
 	}
 
[…]
