### 2016-08-04 2d4306e
crypto-plugin: began work on the gpg module
Implemented a fork into the gpg process
