### 2016-08-04 ce08aa4
crypto-plugin: began work on the gpg module
Implemented a fork into the gpg process
