+++ src/plugins/crypto/CMakeLists.txt
@@ -129,6 +129,8 @@ add_plugin (crypto_gcrypt
 	SOURCES
 		rand_helper.c
 		rand_helper.h
+		gpg.c
+		gpg.h
 		gcrypt_operations.h
 		gcrypt_operations.c
 		crypto.h
@@ -150,6 +152,8 @@ add_plugin (crypto_openssl
 	SOURCES
 		rand_helper.c
 		rand_helper.h
+		gpg.c
+		gpg.h
 		openssl_operations.h
 		openssl_operations.c
 		crypto.h
@@ -171,6 +175,8 @@ add_plugin (crypto_botan
 	SOURCES
 		rand_helper.c
 		rand_helper.h
+		gpg.c
+		gpg.h
 		botan_operations.h
 		botan_operations.cpp
 		crypto.h
