### 2016-08-05 2b872a0
crypto-plugin: progress on gpg module
