### 2016-08-05 ac5cf09
crypto-plugin: master password encryption using GPG
