+++ src/plugins/crypto/gpg.h
@@ -16,6 +16,7 @@
 #define ELEKTRA_CRYPTO_DEFAULT_GPG_BIN "/usr/bin/gpg2"
 #define ELEKTRA_CRYPTO_MESSAGE_MAX_LEN (1024)
 
-int elektraCryptoGpgCall (Key * errorKey, Key * msgKey, char * argv[], size_t argc);
+int elektraCryptoGpgEncryptMasterPassword (KeySet * conf, Key * errorKey, Key * msgKey);
+int elektraCryptoGpgCall (KeySet * conf, Key * errorKey, Key * msgKey, char * argv[], size_t argc);
 
 #endif
