### 2016-08-05 445127e
crypto-plugin: master password encryption using GPG
