### 2016-08-05 7a94ace
crypto-plugin: fix possible memory leak
