### 2016-08-08 3484145
passwd: unit tests
