### 2016-08-08 50b35bb
keyRel2: improve api robustness
