### 2016-08-08 97951c6
intercept: generate temporary file if no filename is given
