### 2016-08-08 29694d4
blockresolver: initial commit
