+++ src/plugins/blockresolver/blockresolver.c
@@ -1,0 +1,339 @@
+/**
+ * @file
+ *
+ * @brief Source for blockresolver plugin
+ *
+ * @copyright BSD License (see doc/COPYING or http://www.libelektra.org)
+ *
+ */
+
+#include "blockresolver.h"
+
+#include <kdbhelper.h>
+#include <stdio.h>
+#include <stdlib.h>
+#include <string.h>
+#include <sys/stat.h>
+#include <sys/time.h>
+#include <time.h>
+#include <unistd.h>
+
+#define TV_MAX_DIGITS 26
+#define BUFSIZE_MAX 1024
+
+typedef struct
+{
+	char * tmpFile;
+	char * realFile;
+	char * identifier;
+	time_t mtime;
+	long startPos;
+	long endPos;
+	unsigned short getPass;
+	unsigned short setPass;
+} BlockData;
+
+int elektraBlockresolverCheckFile (const char * filename ELEKTRA_UNUSED)
+{
+	return 1;
+}
+
+static const char * genTempFilename ()
+{
+	struct timeval tv;
+	gettimeofday (&tv, 0);
+	const char * tmpFilePrefix = "/tmp/elektra_blockresolver_";
+	char * tmpFile = elektraMalloc (strlen (tmpFilePrefix) + TV_MAX_DIGITS + 2);
+	snprintf (tmpFile, strlen (tmpFilePrefix) + TV_MAX_DIGITS + 2, "%s%lu:%lu", tmpFilePrefix, tv.tv_sec, tv.tv_usec);
+	return tmpFile;
+}
+
+static int initData (Plugin * handle)
+{
+	BlockData * data = elektraPluginGetData (handle);
+	if (!data)
+	{
+		data = elektraCalloc (sizeof (BlockData));
+		elektraPluginSetData (handle, data);
+		KeySet * config = elektraPluginGetConfig (handle);
+		ksRewind (config);
+		Key * key = ksLookupByName (config, "/identifier", KDB_O_NONE);
+		if (!key) return -1;
+		data->identifier = (char *)keyString (key);
+		key = ksLookupByName (config, "/path", KDB_O_NONE);
+		if (!key) return -1;
+		data->realFile = (char *)keyString (key);
+		struct stat buf;
+		if (stat (data->realFile, &buf)) return -1;
+		data->mtime = buf.st_mtime;
+		data->tmpFile = (char *)genTempFilename ();
+		data->startPos = -1;
+		data->endPos = -1;
+		data->getPass = 0;
+		data->setPass = 0;
+	}
+	return 0;
+}
+int elektraBlockresolverOpen (Plugin * handle ELEKTRA_UNUSED, Key * errorKey ELEKTRA_UNUSED)
+{
+	// plugin initialization logic
+	// this function is optional
+	return 1; // success
+}
+
+int elektraBlockresolverClose (Plugin * handle ELEKTRA_UNUSED, Key * errorKey ELEKTRA_UNUSED)
+{
+	// free all plugin resources and shut it down
+	// this function is optional
+	BlockData * data = elektraPluginGetData (handle);
+	if (data)
+	{
+		if (data->tmpFile)
+		{
+			unlink (data->tmpFile);
+			elektraFree (data->tmpFile);
+		}
+		elektraFree (data);
+	}
+	elektraPluginSetData (handle, 0);
+	return 1; // success
+}
+
+static long getBlockStart (FILE * fp, const char * identifier)
+{
+	long position = -1;
+	char buffer[BUFSIZE_MAX];
+	fseek (fp, 0, SEEK_SET);
+	while (fgets (buffer, sizeof (buffer), fp))
+	{
[…]
