+++ src/plugins/blockresolver/README.md
@@ -1,0 +1,23 @@
+- infos = Information about the blockresolver plugin is in keys below
+- infos/author = Name <name@libelektra.org>
+- infos/licence = BSD
+- infos/needs =
+- infos/provides = resolver
+- infos/recommends =
+- infos/placements = rollback getresolver setresolver commit
+- infos/status = recommended productive maintained reviewed conformant compatible coverage specific unittest shelltest tested nodep libc configurable final preview memleak experimental difficult unfinished old nodoc concept 
+- infos/metadata =
+- infos/description =
+
+## Usage ##
+
+Copy this blockresolver if you want to start a new
+plugin written in C.
+
+You can use scripts/copy-blockresolver
+to automatically rename everything to your
+plugin name:
+
+	cd src/plugins
+	../../scripts/copy-blockresolver yourplugin
+
