+++ src/plugins/blockresolver/CMakeLists.txt
@@ -1,0 +1,8 @@
+include (LibAddMacros)
+
+add_plugin (blockresolver
+	SOURCES
+		blockresolver.h
+		blockresolver.c
+	ADD_TEST
+	)
