+++ src/plugins/passwd/testmod_passwd.c
@@ -73,11 +73,11 @@ int main (int argc, char ** argv)
 
 	init (argc, argv);
 
-#ifdef HAS_FGETPWENT
+#if HAS_FGETPWENT
 	test_read ();
 #endif
 	test_write ();
-#ifdef HAS_FGETPWENT
+#if HAS_FGETPWENT
 	test_read_write ();
 #endif
 
