### 2016-08-09 8d96bb9
mozilla prefs: basic read support
