+++ src/plugins/prefs/prefs.c
@@ -1,0 +1,226 @@
+/**
+ * @file
+ *
+ * @brief Source for prefs plugin
+ *
+ * @copyright BSD License (see doc/COPYING or http://www.libelektra.org)
+ *
+ */
+
+#include "prefs.h"
+
+#include <kdbhelper.h>
+
+#include <stdio.h>
+#include <stdlib.h>
+#include <string.h>
+
+typedef enum {
+	PREF = 0,
+	USER,
+	LOCK,
+	STICKY,
+	PREF_END,
+} PrefType;
+typedef enum {
+	VAR = 0,
+	CONST,
+	DATA_END,
+} DataType;
+
+const char * data[] = { "var", "const" };
+
+const char * function[] = { "pref", "user_pref", "lock_pref", "sticky_pref" };
+const char * prefix[] = { "pref", "user", "lock", "sticky" };
+
+int elektraPrefsOpen (Plugin * handle ELEKTRA_UNUSED, Key * errorKey ELEKTRA_UNUSED)
+{
+	// plugin initialization logic
+	// this function is optional
+
+	return 1; // success
+}
+
+int elektraPrefsClose (Plugin * handle ELEKTRA_UNUSED, Key * errorKey ELEKTRA_UNUSED)
+{
+	// free all plugin resources and shut it down
+	// this function is optional
+
+	return 1; // success
+}
+
+static Key * prefToKey (Key * parentKey, PrefType type, const char * pref)
+{
+	Key * key = keyNew (keyName (parentKey), KEY_END);
+	keyAddBaseName (key, "preferences");
+	keyAddBaseName (key, prefix[type]);
+	char * localString = strdup (pref);
+	char * cPtr = strstr (localString, ",");
+	*cPtr = '\0';
+	char * sPtr = localString;
+	++sPtr;
+	*sPtr++ = '\0';
+	char * ePtr = cPtr - 1;
+	*ePtr = '\0';
+	size_t keyLen = ePtr - sPtr;
+	char * prefKey = elektraMalloc (keyLen + 1);
+	snprintf (prefKey, keyLen + 1, "%s", sPtr);
+	char * tPtr = strtok (prefKey, ".");
+	if (tPtr) keyAddBaseName (key, tPtr);
+	while ((tPtr = strtok (NULL, ".")) != NULL)
+	{
+		keyAddBaseName (key, tPtr);
+	}
+	elektraFree (prefKey);
+	sPtr = cPtr + 2;
+	ePtr = strrchr (sPtr, ')');
+	*ePtr = '\0';
+	size_t argLen = ePtr - sPtr;
+	char * prefArg = elektraMalloc (argLen + 1);
+	snprintf (prefArg, argLen + 1, "%s", sPtr);
+	if (!strcmp (prefArg, "true") || !(strcmp (prefArg, "false")))
+	{
+		keySetMeta (key, "pref/type", "boolean");
+		keySetString (key, prefArg);
+	}
+	else if (prefArg[0] == '"' && prefArg[strlen (prefArg) - 1] == '"')
+	{
+		// TODO: else if list
+		keySetMeta (key, "pref/type", "string");
+		*prefArg = '\0';
+		*(prefArg + (strlen (prefArg + 1))) = '\0';
+		keySetString (key, (prefArg + 1));
+	}
+	else
+	{
+		keySetMeta (key, "pref/type", "integer");
+		keySetString (key, prefArg);
+	}
+	elektraFree (prefArg);
+	elektraFree (localString);
+	return key;
+}
+
+static Key * varToKey (Key * parentKey, DataType d, const char * string)
+{
+	Key * key = keyNew (keyString (parentKey), KEY_END);
+	keyAddBaseName (parentKey, "variables");
+	return key;
+}
+
[…]
