+++ src/plugins/prefs/README.md
@@ -1,0 +1,23 @@
+- infos = Information about the prefs plugin is in keys below
+- infos/author = Name <name@libelektra.org>
+- infos/licence = BSD
+- infos/needs =
+- infos/provides = storage
+- infos/recommends =
+- infos/placements = getstorage setstorage
+- infos/status = recommended productive maintained reviewed conformant compatible coverage specific unittest shelltest tested nodep libc configurable final preview memleak experimental difficult unfinished old nodoc concept 
+- infos/metadata =
+- infos/description =
+
+## Usage ##
+
+Copy this prefs if you want to start a new
+plugin written in C.
+
+You can use scripts/copy-prefs
+to automatically rename everything to your
+plugin name:
+
+	cd src/plugins
+	../../scripts/copy-prefs yourplugin
+
