+++ src/plugins/prefs/CMakeLists.txt
@@ -1,0 +1,8 @@
+include (LibAddMacros)
+
+add_plugin (prefs
+	SOURCES
+		prefs.h
+		prefs.c
+	ADD_TEST
+	)
