### 2016-08-09 a1fa515
logger: fix log to file
