### 2016-08-09 a42c637
doc: add draft decision
