### 2016-08-10 680c0aa
doc: add link global_validation.md
