### 2016-08-10 58161b4
tests: make regex stricter+fix output
