+++ tests/shell/check_plugins.sh
@@ -25,9 +25,9 @@ do
 	echo "Do in-source check for plugin $PLUGIN"
 
 	README="$PLUGINS_DIR/README.md"
-	WC=`grep "^- \\[$PLUGIN\\]($PLUGIN.*)" "$README"  | wc -l`
+	WC=`grep "^- \\[$PLUGIN\\]($PLUGIN/)" "$README"  | wc -l`
 	[ $WC -eq 1 ]
-	succeed_if "Your plugin $PLUGIN does not have an entry ^- [$PLUGIN]($PLUGIN)$ in $README"
+	succeed_if "Your plugin $PLUGIN does not have an entry ^- [$PLUGIN]($PLUGIN/) in $README"
 
 	#CACHE=@CMAKE_SOURCE_DIR@/cmake/ElektraCache.cmake
 	#WC=`grep "^	$PLUGIN$" "$CACHE"  | wc -l`
