### 2016-08-10 2a380e4
add desktop plugin
+ mark info plugins that they provide info
