+++ src/plugins/desktop/CMakeLists.txt
@@ -1,0 +1,8 @@
+include (LibAddMacros)
+
+add_plugin (desktop
+	SOURCES
+		desktop.h
+		desktop.c
+	ADD_TEST
+	)
