### 2016-08-10 1d86229
kdb: do not crash if set is missing
