### 2016-08-10 ca1a0c2
desktop: remove not needed functions
