+++ src/plugins/desktop/desktop.c
@@ -16,22 +16,6 @@
 #include <kdbhelper.h>
 
 
-int elektraDesktopOpen (Plugin * handle ELEKTRA_UNUSED, Key * errorKey ELEKTRA_UNUSED)
-{
-	// plugin initialization logic
-	// this function is optional
-
-	return 1; // success
-}
-
-int elektraDesktopClose (Plugin * handle ELEKTRA_UNUSED, Key * errorKey ELEKTRA_UNUSED)
-{
-	// free all plugin resources and shut it down
-	// this function is optional
-
-	return 1; // success
-}
-
 int elektraDesktopGet (Plugin * handle ELEKTRA_UNUSED, KeySet * returned ELEKTRA_UNUSED, Key * parentKey ELEKTRA_UNUSED)
 {
 	if (!elektraStrCmp (keyName (parentKey), "system/elektra/modules/desktop"))
@@ -39,12 +23,7 @@ int elektraDesktopGet (Plugin * handle ELEKTRA_UNUSED, KeySet * returned ELEKTRA
 		KeySet * contract =
 			ksNew (30, keyNew ("system/elektra/modules/desktop", KEY_VALUE, "desktop plugin waits for your orders", KEY_END),
 			       keyNew ("system/elektra/modules/desktop/exports", KEY_END),
-			       keyNew ("system/elektra/modules/desktop/exports/open", KEY_FUNC, elektraDesktopOpen, KEY_END),
-			       keyNew ("system/elektra/modules/desktop/exports/close", KEY_FUNC, elektraDesktopClose, KEY_END),
 			       keyNew ("system/elektra/modules/desktop/exports/get", KEY_FUNC, elektraDesktopGet, KEY_END),
-			       keyNew ("system/elektra/modules/desktop/exports/set", KEY_FUNC, elektraDesktopSet, KEY_END),
-			       keyNew ("system/elektra/modules/desktop/exports/error", KEY_FUNC, elektraDesktopError, KEY_END),
-			       keyNew ("system/elektra/modules/desktop/exports/checkconf", KEY_FUNC, elektraDesktopCheckConfig, KEY_END),
 #include ELEKTRA_README (desktop)
 			       keyNew ("system/elektra/modules/desktop/infos/version", KEY_VALUE, PLUGINVERSION, KEY_END), KS_END);
 		ksAppend (returned, contract);
@@ -85,43 +64,11 @@ int elektraDesktopGet (Plugin * handle ELEKTRA_UNUSED, KeySet * returned ELEKTRA
 	return 1; // success
 }
 
-int elektraDesktopSet (Plugin * handle ELEKTRA_UNUSED, KeySet * returned ELEKTRA_UNUSED, Key * parentKey ELEKTRA_UNUSED)
-{
-	// get all keys
-	// this function is optional
-
-	return 1; // success
-}
-
-int elektraDesktopError (Plugin * handle ELEKTRA_UNUSED, KeySet * returned ELEKTRA_UNUSED, Key * parentKey ELEKTRA_UNUSED)
-{
-	// set all keys
-	// this function is optional
-
-	return 1; // success
-}
-
-int elektraDesktopCheckConfig (Key * errorKey ELEKTRA_UNUSED, KeySet * conf ELEKTRA_UNUSED)
-{
-	// validate plugin configuration
-	// this function is optional
-
-	// the return codes have the following meaning:
-	// 0: The configuration was OK and has not been changed
-	// 1: The configuration has been changed and now it is OK
[…]
