### 2016-08-10 54d0f92
desktop: remove test
