+++ src/plugins/desktop/CMakeLists.txt
@@ -4,5 +4,4 @@ add_plugin (desktop
 	SOURCES
 		desktop.h
 		desktop.c
-	ADD_TEST
 	)
