### 2016-08-10 6974b2d
github: add templates
