+++ .github/ISSUE_TEMPLATE.md
@@ -1,0 +1,12 @@
+## Describe what you wanted to do
+
+## Describe what you expected
+
+## Describe what actually happened
+
+## System Information
+
+- Elektra Version: master
+- Versions of other relevant software?
+
+## Further Log Files and Output
