+++ .github/PULL_REQUEST_TEMPLATE.md
@@ -1,0 +1,13 @@
+## Purpose
+
+Describe the purpose of the pull request.
+(Long descriptions should be in documentation, not here)
+
+## Checklist
+
+- [ ] on my machine nothing got broken
+- [ ] I added unit tests
+- [ ] commit messages are fine with references to issues
+- [ ] affected documentation is fixed
+- [ ] I added code comments, logging, and assertions
+- [ ] meta data is updated (e.g. README.md of plugins)
