+++ .github/CONTRIBUTING.md
@@ -1,0 +1,24 @@
+## Documentation
+
+Any typo fixes, spelling fixes, fixes in documentation,
+or README.md can be done without any further considerations!
+Simply do the fixes in your cloned repo and create a pull request.
+
+
+## Code
+
+Before you issue a pull request that modifies code:
+
+- You should read the [coding document](/doc/CODING.md).
+- You should read the [design document](/doc/DESIGN.md).
+- Make sure you fulfilled the [definition of done](/.github/PULL_REQUEST_TEMPLATE.md).
+
+
+## Architecture
+
+Before you start making fundamental changes:
+
+- Propose it by creating a [github issue](https://github.com/ElektraInitiative/libelektra/issues/new)
+  for discussions.
+- Create a [decisions](/doc/decisions/README.md) describing you want to do
+  to keep the discussion more efficient.
