+++ .github/ISSUE_TEMPLATE.md
@@ -1,3 +1,7 @@
+Please remove this template if you have
+a question or proposal and do not want
+to report an issue.
+
 ## Describe what you wanted to do
 
 ## Describe what you expected
