+++ .github/PULL_REQUEST_TEMPLATE.md
@@ -5,9 +5,15 @@ Describe the purpose of the pull request.
 
 ## Checklist
 
+Please only check relevant points.
+For docu fixes, spell checking and similar nothing
+needs to be checked.
+
+- [ ] commit messages are fine with references to issues
 - [ ] on my machine nothing got broken
 - [ ] I added unit tests
-- [ ] commit messages are fine with references to issues
 - [ ] affected documentation is fixed
 - [ ] I added code comments, logging, and assertions
 - [ ] meta data is updated (e.g. README.md of plugins)
+
+@markus2330 please review my pull request
