### 2016-08-10 b61104a
doc: logger
