### 2016-08-11 2c86cad
Documentation: Fix spelling mistake
