### 2016-08-11 6f01626
fix double return
