### 2016-08-11 2ce2aa3
doc: fix text
