+++ .github/PULL_REQUEST_TEMPLATE.md
@@ -9,8 +9,8 @@ Please only check relevant points.
 For docu fixes, spell checking and similar nothing
 needs to be checked.
 
-- [ ] commit messages are fine with references to issues
-- [ ] on my machine nothing got broken
+- [ ] commit messages are fine (with references to issues)
+- [ ] I ran all tests and everything went fine
 - [ ] I added unit tests
 - [ ] affected documentation is fixed
 - [ ] I added code comments, logging, and assertions
