+++ .github/CONTRIBUTING.md
@@ -1,11 +1,11 @@
-## Documentation
+# Documentation
 
 Any typo fixes, spelling fixes, fixes in documentation,
 or README.md can be done without any further considerations!
 Simply do the fixes in your cloned repo and create a pull request.
 
 
-## Issues
+# Issues
 
 If you have any questions you can ask them using
 [github issue](https://github.com/ElektraInitiative/libelektra/issues/new).
@@ -13,7 +13,7 @@ Simply ignore the template in this case.
 
 
 
-## Code
+# Code
 
 Before you issue a pull request that modifies code:
 
@@ -22,7 +22,7 @@ Before you issue a pull request that modifies code:
 - Make sure you fulfilled the [definition of done](/.github/PULL_REQUEST_TEMPLATE.md).
 
 
-## Architecture
+# Architecture
 
 Before you start making fundamental changes:
 
