### 2016-08-11 e7adec9
OPMPHM: add PR feedback + add clear to vstack and vheap
