### 2016-08-12 08de66a
OPMPHM: rename opmphm to kdbopmphm
