### 2016-08-13 d57a94d
global plugins: introduce positions and subpositions
Simplify and revamp mount.c to be able to mount the newly defined global positions.

Use a macro to generate positions and subpositions. This makes the
system more flexible to add/remove positions later.

Use a function to execute global plugins and keep boilerplate code
minimal.

Add very simple and basic shell recorder test to test if mounting and
executing the global plugins still works.
