+++ src/libs/elektra/mount.c
@@ -241,8 +241,78 @@ int mountDefault (KDB * kdb, KeySet * modules, int inFallback, Key * errorKey)
 	return 0;
 }
 
+KeySet * elektraMountGlobalsGetConfig (Key * cur, KeySet * global)
+{
+	// putting together the plugins configuration KeySet.
+	Key * sysConfigCutKey = keyDup (cur);
+	keyAddBaseName (sysConfigCutKey, "system");
+	Key * usrConfigCutKey = keyDup (cur);
+	keyAddBaseName (usrConfigCutKey, "user");
+	KeySet * sysConfigKS = ksCut (global, sysConfigCutKey);
+	KeySet * usrConfigKS = ksCut (global, usrConfigCutKey);
+	KeySet * renamedSysConfig = elektraRenameKeys (sysConfigKS, "system");
+	KeySet * renamedUsrConfig = elektraRenameKeys (usrConfigKS, "user");
+	ksDel (sysConfigKS);
+	ksDel (usrConfigKS);
+	keyDel (usrConfigCutKey);
+	keyDel (sysConfigCutKey);
+	KeySet * config = ksNew (0, KS_END);
+	ksAppend (config, renamedSysConfig);
+	ksAppend (config, renamedUsrConfig);
+	ksDel (renamedSysConfig);
+	ksDel (renamedUsrConfig);
+
+	return config;
+}
+
+Key * elektraMountGlobalsFindPlugin (KeySet * referencePlugins, Key * cur)
+{
+	Key * refKey;
+	Key * searchKey = keyNew ("/", KEY_END);
+	keyAddBaseName (searchKey, keyString (cur));
+	refKey = ksLookup (referencePlugins, searchKey, 0);
+	keyDel (searchKey);
+	return refKey;
+}
+
+Plugin * elektraMountGlobalsLoadPlugin (KeySet * referencePlugins, Key * cur, KeySet * global,
+										KeySet * modules, Key * errorKey)
+{
+	Plugin * plugin;
+	Key * refKey = elektraMountGlobalsFindPlugin (referencePlugins, cur);
+
+	if (refKey)
+	{
+		// plugin already loaded, just reference it
+		plugin = *(Plugin **)keyValue (refKey);
+		plugin->refcounter += 1;
+	}
+	else
+	{
+		KeySet * config = elektraMountGlobalsGetConfig(cur, global);
+		const char * pluginName = keyString (cur);
+		// loading the new plugin
+		plugin = elektraPluginOpen (pluginName, modules, ksDup (config), errorKey);
+		if (!plugin)
+		{
+			ELEKTRA_ADD_WARNING (64, errorKey, pluginName);
+			return NULL;
+		}
+
+		// saving the plugin reference to avoid having to load the plugin multiple times
+		refKey = keyNew ("/", KEY_BINARY, KEY_SIZE, sizeof (Plugin *), KEY_VALUE, &plugin, KEY_END);
+		keyAddBaseName (refKey, keyString (cur));
+		ksAppendKey (referencePlugins, refKey);
+		keyDel (refKey);
+		ksDel (config);
+	}
+
+	return plugin;
+}
+
 int mountGlobals (KDB * kdb, KeySet * keys, KeySet * modules, Key * errorKey)
 {
+	int retval = 0;
 	Key * root = ksLookupByName (keys, "system/elektra/globalplugins", 0);
 	if (!root)
 	{
@@ -274,93 +344,75 @@ int mountGlobals (KDB * kdb, KeySet * keys, KeySet * modules, Key * errorKey)
[…]
