+++ src/libs/elektra/kdb.c
@@ -145,6 +145,18 @@ void elektraRemoveMetaData (Key * key, const char * searchfor)
 }
 
 /**
+ * @internal
+ * Helper function to execute global plugins
+ */
+void elektraExecGlobalPlugin (KDB * handle, KeySet * ks, Key * parentKey, int position, int subPosition)
+{
+	if (handle && handle->globalPlugins[position][subPosition])
+	{
+		handle->globalPlugins[position][subPosition]->kdbGet (handle->globalPlugins[position][subPosition], ks, parentKey);
+	}
+}
+
+/**
  * @brief Bootstrap, first phase with fallback
  * @internal
  *
@@ -415,9 +427,12 @@ int kdbClose (KDB * handle, Key * errorKey)
 		handle->initBackend = 0;
 	}
 
-	for (int i = 0; i < NR_GLOBAL_PLUGINS; ++i)
+	for (int i = 0; i < NR_GLOBAL_POSITIONS; ++i)
 	{
-		elektraPluginClose (handle->globalPlugins[i], errorKey);
+		for (int j = 0; j < NR_GLOBAL_SUBPOSITIONS; ++j)
+		{
+			elektraPluginClose (handle->globalPlugins[i][j], errorKey);
+		}
 	}
 
 	if (handle->modules)
@@ -607,20 +622,20 @@ static int elektraGetDoUpdateWithGlobalHooks (KDB * handle, Split * split, KeySe
 			// GLOBAL: getstorage [foreach]
 			// GLOBAL: postgetstorage [foreach]
 
-			if (!pgs_done && (p == (STORAGE_PLUGIN + 1)) && handle->globalPlugins[POSTGETSTORAGE])
+			if (!pgs_done && (p == (STORAGE_PLUGIN + 1)) && handle->globalPlugins[POSTGETSTORAGE][MAXONCE])
 			{
 				pgs_done = 1;
 				keySetName (parentKey, keyName (initialParent));
 				ksRewind (ks);
-				handle->globalPlugins[POSTGETSTORAGE]->kdbGet (handle->globalPlugins[POSTGETSTORAGE], ks, parentKey);
+				handle->globalPlugins[POSTGETSTORAGE][MAXONCE]->kdbGet (handle->globalPlugins[POSTGETSTORAGE][MAXONCE], ks, parentKey);
 				keySetName (parentKey, keyName (split->parents[i]));
 			}
-			else if (!pgc_done && (p == (NR_OF_PLUGINS - 1)) && handle->globalPlugins[POSTGETCLEANUP])
+			else if (!pgc_done && (p == (NR_OF_PLUGINS - 1)) && handle->globalPlugins[POSTGETCLEANUP][MAXONCE])
 			{
 				pgc_done = 1;
 				keySetName (parentKey, keyName (initialParent));
 				ksRewind (ks);
-				handle->globalPlugins[POSTGETCLEANUP]->kdbGet (handle->globalPlugins[POSTGETCLEANUP], ks, parentKey);
+				handle->globalPlugins[POSTGETCLEANUP][MAXONCE]->kdbGet (handle->globalPlugins[POSTGETCLEANUP][MAXONCE], ks, parentKey);
 				keySetName (parentKey, keyName (split->parents[i]));
 			}
 
@@ -795,10 +810,9 @@ int kdbGet (KDB * handle, KeySet * ks, Key * parentKey)
 	// GLOBAL: pregetstorage [max once]
 	// GLOBAL: pregetstorage [foreach]
 	// GLOBAL: pregetstorage [deinit]
-	if (handle->globalPlugins[PREGETSTORAGE])
-	{
[…]
