+++ src/include/kdbprivate.h
@@ -244,18 +244,49 @@ struct _KeySet
  * Helper for identifying global plugin positions
  */
 
+#define FOREACH_POSITION(POSITION) \
+	POSITION(PREROLLBACK) \
+	POSITION(ROLLBACK) \
+	POSITION(POSTROLLBACK) \
+	POSITION(GETRESOLVER) \
+	POSITION(PREGETSTORAGE) \
+	POSITION(GETSTORAGE) \
+	POSITION(POSTGETSTORAGE) \
+	POSITION(SETRESOLVER) \
+	POSITION(POSTGETCLEANUP) \
+	POSITION(PRESETSTORAGE) \
+	POSITION(SETSTORAGE) \
+	POSITION(PRESETCLEANUP) \
+	POSITION(PRECOMMIT) \
+	POSITION(COMMIT) \
+	POSITION(POSTCOMMIT) \
+	POSITION(NR_GLOBAL_POSITIONS)
+
+#define FOREACH_SUBPOSITION(SUBPOSITION) \
+	SUBPOSITION(INIT) \
+	SUBPOSITION(MAXONCE) \
+	SUBPOSITION(FOREACH) \
+	SUBPOSITION(DEINIT) \
+	SUBPOSITION(NR_GLOBAL_SUBPOSITIONS)
+
+#define GENERATE_ENUM(ENUM) ENUM,
+#define GENERATE_STRING(STRING) #STRING,
+
 typedef enum {
-	PREROLLBACK = 0,
-	POSTROLLBACK,
-	PREGETSTORAGE,
-	POSTGETSTORAGE,
-	POSTGETCLEANUP,
-	PRESETSTORAGE,
-	PRESETCLEANUP,
-	PRECOMMIT,
-	POSTCOMMIT,
-	NR_GLOBAL_PLUGINS
+	FOREACH_POSITION(GENERATE_ENUM)
 } GlobalpluginPositions;
+
+typedef enum {
+	FOREACH_SUBPOSITION(GENERATE_ENUM)
+} GlobalpluginSubPositions;
+
+static const char *GlobalpluginPositionsStr[] = {
+	FOREACH_POSITION(GENERATE_STRING)
+};
+
+static const char *GlobalpluginSubPositionsStr[] = {
+	FOREACH_SUBPOSITION(GENERATE_STRING)
+};
 
 /**
  * The access point to the key database.
@@ -289,7 +320,7 @@ struct _KDB
 
 	Backend * initBackend; /*!< The init backend for bootstrapping.*/
 
-	Plugin * globalPlugins[NR_GLOBAL_PLUGINS];
+	Plugin * globalPlugins[NR_GLOBAL_POSITIONS][NR_GLOBAL_SUBPOSITIONS];
 };
 
 
