### 2016-08-15 c78a8c5
remove spec namespace from globalhook keyset
