### 2016-08-15 c822dc9
passwd: fixed typos
