### 2016-08-15 f282308
blockresolver: added errors and warnings
