### 2016-08-15 5a93fbd
crypto-plugin: first prototype for GPG key/IV derivation with gcrypt variant
The prototype calls GPG to encrypt and decrypt the master password r.
The master password r is used to derivate a cryptographic key and an initializtion vector IV.
