+++ src/plugins/crypto/gpg.h
@@ -17,6 +17,7 @@
 #define ELEKTRA_CRYPTO_MESSAGE_MAX_LEN (1024)
 
 int elektraCryptoGpgEncryptMasterPassword (KeySet * conf, Key * errorKey, Key * msgKey);
+int elektraCryptoGpgDecryptMasterPassword (KeySet * conf, Key * errorKey, Key * msgKey);
 int elektraCryptoGpgCall (KeySet * conf, Key * errorKey, Key * msgKey, char * argv[], size_t argc);
 
 #endif
