### 2016-08-15 caef8af
crypto-plugin: move meta-data check for encryption flag to crypto.c
