### 2016-08-15 0c066ae
crypto-plugin: fix compile error in gcrypt_operations.c
