### 2016-08-17 4e70db7
crypto-plugin: add error descriptions to gcrypt_operations.c
