### 2016-08-17 8d4dfa1
crypto-plugin: fix compiler warning
