### 2016-08-17 0373939
crypto-plugin: fix compiler warning
