### 2016-08-17 cdd5c4d
crypto-plugin: fix KDF in gcrypt_operations.c
Hashing algorithm for the PBKDF2 was missing. Now set to SHA512
