### 2016-08-18 2ad9774
doc: make clear global-mount overwrites previous mounts
see #886
