+++ src/tools/kdb/globalmount.hpp
@@ -43,7 +43,7 @@ public:
 
 	virtual std::string getShortHelpText () override
 	{
-		return "Globally mount given plugins.";
+		return "Globally mount given plugins, overwriting previous ones.";
 	}
 
 	virtual std::string getLongHelpText () override
