### 2016-08-18 75dbf32
doc: split up docu of libraries
