### 2016-08-18 e67fe30
doc: write about exported symbols
