### 2016-08-18 f73ad42
crypto-plugin: working prototype for GPG key/IV derivation with OpenSSL
