### 2016-08-18 d2b5e0c
crypto-plugin: working prototype for GPG key/IV derivation with OpenSSL
