+++ src/plugins/crypto/openssl_operations.c
@@ -1,7 +1,7 @@
 /**
  * @file
  *
- * @brief cryptographic interface using the gcrypt library
+ * @brief cryptographic interface using the libcrypto library (part of the OpenSSL project)
  *
  * @copyright BSD License (see doc/COPYING or http://www.libelektra.org)
  *
@@ -9,6 +9,7 @@
 
 #include "crypto.h"
 
+#include "gpg.h"
 #include "openssl_operations.h"
 #include "rand_helper.h"
 
@@ -25,35 +26,133 @@
 static pthread_mutex_t mutex_ssl = PTHREAD_MUTEX_INITIALIZER;
 
 /**
- * @brief read the cryptographic key from the given keyset.
- * @retval NULL on error
- * @retval address of the (Elektra) key in the given keyset holding the cryptographic key to be used.
+ * @brief derive the cryptographic key and IV for a given (Elektra) Key k
+ * @param config KeySet holding the plugin/backend configuration
+ * @param errorKey holds an error description in case of failure
+ * @param k the (Elektra)-Key to be encrypted
+ * @param cKey (Elektra)-Key holding the cryptographic material
+ * @param cIv (Elektra)-Key holding the initialization vector
+ * @retval -1 on failure. errorKey holds the error description.
+ * @retval 1 on success
  */
-static Key * elektraCryptoReadParamKey (KeySet * config, Key * errorKey)
+static int getKeyIvForEncryption (KeySet * config, Key * errorKey, Key * k, Key * cKey, Key * cIv)
 {
-	Key * key = ksLookupByName (config, ELEKTRA_CRYPTO_PARAM_KEY_PATH, 0);
-	if (key == NULL)
+	kdb_octet_t salt[ELEKTRA_CRYPTO_DEFAULT_SALT_LEN];
+	const size_t keyBufferSize = ELEKTRA_CRYPTO_SSL_KEYSIZE + ELEKTRA_CRYPTO_SSL_BLOCKSIZE;
+	kdb_octet_t keyBuffer[keyBufferSize];
+
+	// generate the salt
+	pthread_mutex_lock (&mutex_ssl);
+	if (!RAND_bytes (salt, ELEKTRA_CRYPTO_DEFAULT_SALT_LEN - 1))
 	{
-		ELEKTRA_SET_ERRORF (ELEKTRA_ERROR_CRYPTO_CONFIG_FAULT, errorKey, "missing %s in configuration",
-				    ELEKTRA_CRYPTO_PARAM_KEY_PATH);
+		ELEKTRA_SET_ERRORF (ELEKTRA_ERROR_CRYPTO_INTERNAL_ERROR, errorKey, "failed to generate random salt with error code %lu",
+				    ERR_get_error ());
+		pthread_mutex_unlock (&mutex_ssl);
+		return -1;
 	}
-	return key;
+	pthread_mutex_unlock (&mutex_ssl);
+	elektraCryptoNormalizeRandomString (salt, sizeof (salt));
+	keySetMeta (k, ELEKTRA_CRYPTO_META_SALT, (char *)salt);
+
+	// read iteration count
+	// TODO make iteration count configurable
+	const kdb_unsigned_long_t iterations = ELEKTRA_CRYPTO_DEFAULT_ITERATION_COUNT;
+
+	// receive master password from the configuration
+	Key * master = ksLookupByName (config, ELEKTRA_CRYPTO_PARAM_MASTER_PWD, 0);
[…]
