### 2016-08-18 86867e2
crypto-plugin: working prototype for GPG key/IV derivation with OpenSSL
