+++ src/plugins/crypto/openssl_operations.h
@@ -18,7 +18,8 @@
 
 char * elektraCryptoOpenSSLCreateRandomString (const kdb_unsigned_short_t length);
 int elektraCryptoOpenSSLInit (Key * errorKey);
-int elektraCryptoOpenSSLHandleCreate (elektraCryptoHandle ** handle, KeySet * config, Key * errorKey);
+int elektraCryptoOpenSSLHandleCreate (elektraCryptoHandle ** handle, KeySet * config, Key * errorKey, Key * k,
+				      const enum ElektraCryptoOperation op);
 void elektraCryptoOpenSSLHandleDestroy (elektraCryptoHandle * handle);
 int elektraCryptoOpenSSLEncrypt (elektraCryptoHandle * handle, Key * k, Key * errorKey);
 int elektraCryptoOpenSSLDecrypt (elektraCryptoHandle * handle, Key * k, Key * errorKey);
