### 2016-08-18 2038758
crypto-plugin: working prototype for GPG key/IV derivation with Botan
