### 2016-08-18 d7db39f
crypto-plugin: working prototype for GPG key/IV derivation with Botan
