+++ src/plugins/crypto/botan_operations.cpp
@@ -11,6 +11,7 @@
 #include <botan/filters.h>
 #include <botan/init.h>
 #include <botan/lookup.h>
+#include <botan/pbkdf2.h>
 #include <botan/pipe.h>
 #include <botan/symkey.h>
 #include <kdbplugin.h>
@@ -22,44 +23,134 @@ extern "C" {
 
 #include "botan_operations.h"
 #include "crypto.h"
+#include "gpg.h"
 #include "rand_helper.h"
 #include <kdberrors.h>
 #include <string.h>
 
 /**
- * @brief read the cryptographic key from the given keyset.
- * @retval NULL on error
- * @retval a SymmetricKey to be used for the cryptographic operation
+ * @brief derive the cryptographic key and IV for a given (Elektra) Key k
+ * @param config KeySet holding the plugin/backend configuration
+ * @param errorKey holds an error description in case of failure
+ * @param k the (Elektra)-Key to be encrypted
+ * @param cKey pointer to the SymmetricKey
+ * @param cIv pointer to the InitializationVector
+ * @retval -1 on failure. errorKey holds the error description.
+ * @retval 1 on success
  */
-static SymmetricKey * getSymmKey (KeySet * config, Key * errorKey)
+static int getKeyIvForEncryption (KeySet * config, Key * errorKey, Key * k, SymmetricKey ** cKey, InitializationVector ** cIv)
 {
-	Key * key = ksLookupByName (config, ELEKTRA_CRYPTO_PARAM_KEY_PATH, 0);
-	if (key == NULL)
+	byte salt[ELEKTRA_CRYPTO_DEFAULT_SALT_LEN];
+	const size_t requiredKeyBytes = ELEKTRA_CRYPTO_BOTAN_KEYSIZE + ELEKTRA_CRYPTO_BOTAN_BLOCKSIZE;
+	Key * msg = NULL;
+
+	try
 	{
-		ELEKTRA_SET_ERRORF (ELEKTRA_ERROR_CRYPTO_CONFIG_FAULT, errorKey, "missing %s in configuration",
-				    ELEKTRA_CRYPTO_PARAM_KEY_PATH);
-		return NULL;
+		// generate the salt
+		AutoSeeded_RNG rng;
+		rng.randomize (salt, ELEKTRA_CRYPTO_DEFAULT_SALT_LEN - 1);
+		elektraCryptoNormalizeRandomString (salt, sizeof (salt));
+		keySetMeta (k, ELEKTRA_CRYPTO_META_SALT, (char *)salt);
+
+		// read iteration count
+		// TODO make iteration count configurable
+		const kdb_unsigned_long_t iterations = ELEKTRA_CRYPTO_DEFAULT_ITERATION_COUNT;
+
+		// receive master password from the configuration
+		Key * master = ksLookupByName (config, ELEKTRA_CRYPTO_PARAM_MASTER_PWD, 0);
+		if (!master)
+		{
+			ELEKTRA_SET_ERRORF (ELEKTRA_ERROR_CRYPTO_CONFIG_FAULT, errorKey, "missing %s", ELEKTRA_CRYPTO_PARAM_MASTER_PWD);
+			return -1;
+		}
+		msg = keyDup (master);
+		if (elektraCryptoGpgDecryptMasterPassword (config, errorKey, msg) != 1)
+		{
+			keyDel (msg);
+			return -1;
+		}
+
+		// generate/derive the cryptographic key and the IV
+		PBKDF * pbkdf = get_pbkdf ("PBKDF2(SHA-256)");
[…]
