### 2016-08-18 605d639
crypto-plugin: increase default lengths for parameters
Set default iteration count for PBKDF2 to 15000 and salt length to 17 (=16 Bytes)
