+++ src/plugins/crypto/crypto.h
@@ -27,8 +27,8 @@ enum ElektraCryptoOperation
 };
 
 #define ELEKTRA_CRYPTO_DEFAULT_MASTER_PWD_LENGTH (30)
-#define ELEKTRA_CRYPTO_DEFAULT_ITERATION_COUNT (2048)
-#define ELEKTRA_CRYPTO_DEFAULT_SALT_LEN (12)
+#define ELEKTRA_CRYPTO_DEFAULT_ITERATION_COUNT (15000)
+#define ELEKTRA_CRYPTO_DEFAULT_SALT_LEN (17)
 #define ELEKTRA_CRYPTO_PARAM_MASTER_PWD_LEN "/crypto/masterpasswordlength"
 #define ELEKTRA_CRYPTO_PARAM_MASTER_PWD "/crypto/masterpassword"
 #define ELEKTRA_CRYPTO_PARAM_KEY_PATH "/crypto/key"
