### 2016-08-19 36ccb6c
crypto-plugin: make path to gpg binary configurable
The GPG module reads from '/crypto/gpg' and uses a default if not provided.
