### 2016-08-19 024e5ae
crypto-plugin: make path to gpg binary configurable
The GPG module reads from '/crypto/gpg' and uses a default if not provided.
