### 2016-08-19 dde5db9
crypto-plugin: improve gpg path configuration
