### 2016-08-19 d313bc4
crypto-plugin: verify master password in checkconf
