### 2016-08-19 8bc80ca
crypto-plugin: make iteration count configurable
The iteration count (required by the KDF) can be set as 'crypto/iterations' within the plugin config
