+++ src/plugins/crypto/rand_helper.h
@@ -10,8 +10,10 @@
 #ifndef ELEKTRA_PLUGIN_CRYPTO_RAND_HELPER_H
 #define ELEKTRA_PLUGIN_CRYPTO_RAND_HELPER_H
 
+#include <kdb.h>
 #include <kdbtypes.h>
 
+kdb_unsigned_long_t elektraCryptoGetIterationCount (KeySet * config);
 void elektraCryptoNormalizeRandomString (kdb_octet_t * buffer, const kdb_unsigned_short_t length);
 
 #endif
