+++ src/plugins/crypto/botan_operations.cpp
@@ -53,8 +53,7 @@ static int getKeyIvForEncryption (KeySet * config, Key * errorKey, Key * k, Symm
 		keySetMeta (k, ELEKTRA_CRYPTO_META_SALT, (char *)salt);
 
 		// read iteration count
-		// TODO make iteration count configurable
-		const kdb_unsigned_long_t iterations = ELEKTRA_CRYPTO_DEFAULT_ITERATION_COUNT;
+		const kdb_unsigned_long_t iterations = elektraCryptoGetIterationCount (config);
 
 		// receive master password from the configuration
 		Key * master = ksLookupByName (config, ELEKTRA_CRYPTO_PARAM_MASTER_PWD, 0);
@@ -113,8 +112,7 @@ static int getKeyIvForDecryption (KeySet * config, Key * errorKey, Key * k, Symm
 	}
 
 	// get the iteration count
-	// TODO make iteration count configurable
-	const kdb_unsigned_long_t iterations = ELEKTRA_CRYPTO_DEFAULT_ITERATION_COUNT;
+	const kdb_unsigned_long_t iterations = elektraCryptoGetIterationCount (config);
 
 	// receive master password from the configuration
 	Key * master = ksLookupByName (config, ELEKTRA_CRYPTO_PARAM_MASTER_PWD, 0);
