### 2016-08-19 48579c5
crypto-plugin: rename rand_helper.[c,h] to helper.[c,h]
