+++ src/plugins/crypto/helper.h
@@ -1,14 +1,14 @@
 /**
  * @file
  *
- * @brief helper functions for dealing with random values
+ * @brief helper functions for the crypto plugin
  *
  * @copyright BSD License (see doc/COPYING or http://www.libelektra.org)
  *
  */
 
-#ifndef ELEKTRA_PLUGIN_CRYPTO_RAND_HELPER_H
-#define ELEKTRA_PLUGIN_CRYPTO_RAND_HELPER_H
+#ifndef ELEKTRA_PLUGIN_CRYPTO_HELPER_H
+#define ELEKTRA_PLUGIN_CRYPTO_HELPER_H
 
 #include <kdb.h>
 #include <kdbtypes.h>
