### 2016-08-19 ba82094
crypto-plugin: rename rand_helper.[c,h] to helper.[c,h]
