### 2016-08-19 ffdc0a2
crypto-plugin: make GPG keys configurable
GPG recipient keys will be read from /crypto/key or below
