+++ src/plugins/crypto/gpg.h
@@ -14,8 +14,8 @@
 #include <kdbtypes.h>
 
 #define ELEKTRA_CRYPTO_PARAM_GPG_BIN "/crypto/gpg"
+#define ELEKTRA_CRYPTO_PARAM_GPG_KEY "/crypto/key"
 #define ELEKTRA_CRYPTO_DEFAULT_GPG_BIN "/usr/bin/gpg2"
-#define ELEKTRA_CRYPTO_MESSAGE_MAX_LEN (1024)
 
 int elektraCryptoGpgEncryptMasterPassword (KeySet * conf, Key * errorKey, Key * msgKey);
 int elektraCryptoGpgDecryptMasterPassword (KeySet * conf, Key * errorKey, Key * msgKey);
