### 2016-08-19 2ccabdc
crypto-plugin: update plugin documentation
