### 2016-08-20 55ac142
mozilla prefs: fixed read + added write support
