+++ src/plugins/prefs/prefs.c
@@ -11,6 +11,7 @@
 
 #include <kdbhelper.h>
 
+#include <ctype.h>
 #include <stdio.h>
 #include <stdlib.h>
 #include <string.h>
@@ -49,6 +50,22 @@ int elektraPrefsClose (Plugin * handle ELEKTRA_UNUSED, Key * errorKey ELEKTRA_UN
 	return 1; // success
 }
 
+static inline void lskip (char ** p)
+{
+	while (**p && isspace (**p))
+		++(*p);
+}
+
+static inline void rstrip (const char * s, char ** p)
+{
+	while ((*p > s) && **p && isspace (**p))
+	{
+		**p = '\0';
+		--(*p);
+	}
+}
+
+
 static Key * prefToKey (Key * parentKey, PrefType type, const char * pref)
 {
 	Key * key = keyNew (keyName (parentKey), KEY_END);
@@ -61,7 +78,7 @@ static Key * prefToKey (Key * parentKey, PrefType type, const char * pref)
 	++sPtr;
 	*sPtr++ = '\0';
 	char * ePtr = cPtr - 1;
-	*ePtr = '\0';
+	rstrip (sPtr, &ePtr);
 	size_t keyLen = ePtr - sPtr;
 	char * prefKey = elektraMalloc (keyLen + 1);
 	snprintf (prefKey, keyLen + 1, "%s", sPtr);
@@ -72,28 +89,30 @@ static Key * prefToKey (Key * parentKey, PrefType type, const char * pref)
 		keyAddBaseName (key, tPtr);
 	}
 	elektraFree (prefKey);
-	sPtr = cPtr + 2;
+	sPtr = cPtr + 1;
+	lskip (&sPtr);
 	ePtr = strrchr (sPtr, ')');
-	*ePtr = '\0';
-	size_t argLen = ePtr - sPtr;
+	*ePtr-- = '\0';
+	rstrip (sPtr, &ePtr);
+	size_t argLen = ePtr - sPtr + 1;
 	char * prefArg = elektraMalloc (argLen + 1);
 	snprintf (prefArg, argLen + 1, "%s", sPtr);
 	if (!strcmp (prefArg, "true") || !(strcmp (prefArg, "false")))
 	{
-		keySetMeta (key, "pref/type", "boolean");
+		keySetMeta (key, "type", "boolean");
 		keySetString (key, prefArg);
 	}
 	else if (prefArg[0] == '"' && prefArg[strlen (prefArg) - 1] == '"')
 	{
 		// TODO: else if list
-		keySetMeta (key, "pref/type", "string");
+		keySetMeta (key, "type", "string");
 		*prefArg = '\0';
 		*(prefArg + (strlen (prefArg + 1))) = '\0';
 		keySetString (key, (prefArg + 1));
 	}
 	else
 	{
-		keySetMeta (key, "pref/type", "integer");
+		keySetMeta (key, "type", "integer");
 		keySetString (key, prefArg);
 	}
 	elektraFree (prefArg);
@@ -108,7 +127,7 @@ static Key * varToKey (Key * parentKey, DataType d, const char * string)
 	return key;
 }
 
-int elektraPrefsGet (Plugin * handle ELEKTRA_UNUSED, KeySet * returned ELEKTRA_UNUSED, Key * parentKey ELEKTRA_UNUSED)
+int elektraPrefsGet (Plugin * handle ELEKTRA_UNUSED, KeySet * returned, Key * parentKey)
 {
 	if (!elektraStrCmp (keyName (parentKey), "system/elektra/modules/prefs"))
 	{
@@ -148,7 +167,7 @@ int elektraPrefsGet (Plugin * handle ELEKTRA_UNUSED, KeySet * returned ELEKTRA_U
 	{
[…]
