### 2016-08-20 9b86402
intercept: access + xstat(64)
