### 2016-08-20 5a75e30
mozilla prefs: autoconfig file
