+++ src/plugins/prefs/autoconfig/prefs.js
@@ -1,0 +1,6 @@
+lockPref("a.lock.1", "lock1");
+lockPref("a.lock.2", "lock2");
+pref("a.default.1", 1);
+pref("a.default.2", 2);
+user_pref("a.user.f", false);
+user_pref("a.user.t", true);
