+++ src/plugins/prefs/autoconfig/firefox.cfg
@@ -1,0 +1,47 @@
+//
+
+const configName = getPref("elektra.config.file");
+const serverPort = getPref("elektra.config.reload_trigger_port");
+
+function reload()
+{
+    let infile = Components.classes["@mozilla.org/file/local;1"].createInstance(Components.interfaces.nsILocalFile);
+    infile.initWithPath(configName);
+    var outfile = Components.classes["@mozilla.org/file/directory_service;1"].getService(Components.interfaces.nsIProperties).get("ProfD", Components.interfaces.nsIFile);
+    outfile.append("prefs.js");
+
+    const nsIPrefService = Components.interfaces.nsIPrefService;
+    const PrefServiceContractID = "@mozilla.org/preferences-service;1"
+    let prefService = Components.classes[PrefServiceContractID].getService(nsIPrefService);
+
+    prefService.readUserPrefs(infile);
+    prefService.savePrefFile(outfile); 	  	//save user prefs to prefs.js
+    prefService.readUserPrefs(outfile);   	//don't own my input file!!! 
+}
+
+var listener = {
+onSocketAccepted: function(serverSocket, clientSocket) {
+		      var stream = clientSocket.openInputStream(0,0,0);
+		      var sin = Components.classes["@mozilla.org/scriptableinputstream;1"].createInstance(Components.interfaces.nsIScriptableInputStream);
+		      try{            
+			  sin.init(stream);
+			  var bytes = sin.available();
+			  var request = "";
+			  request = sin.read(bytes);
+			  if (request.valueOf() == "reload"){
+			      reload();
+			  }
+		      } finally {
+			  sin.close();
+			  stream.close();
+		      }
+
+		  }
+}
+
+reload(); 	//read once on startup
+
+var serverSocket = Components.classes["@mozilla.org/network/server-socket;1"].createInstance(Components.interfaces.nsIServerSocket);
+serverSocket.init(serverPort, true, 1); 	//only listen on loopback
+
+serverSocket.asyncListen(listener);
