+++ src/plugins/prefs/autoconfig/preload.ini
@@ -1,0 +1,4 @@
+[open]
+\/tmp\/myPrefs.js = /tmp/elektra_prefs.js
+\/tmp\/myPrefs.js/generate = system/prefs/gen
+\/tmp\/myPrefs.js/generate/plugin = prefs
