### 2016-08-20 40a79e6
crypto-plugin: improve docu
