+++ doc/METADATA.ini
@@ -618,5 +618,7 @@ description= used to mark a key for encryption. The crypto plugin is supposed to
 status= implemented
 usedby/plugins= crypto
 type= string
-description= used to store the salt thas is required for deriving the cryptographic key as well as the initialization vector (IV).
+description= used to store the salt that seeds the Key Derivation Function, which is used for deriving the cryptographic key and the initialization vector (IV).
+ The salt is re-generated on every encryption call (called by kdb set) so no key and IV combination is ever used twice.
+ No special protection is required for the salt. It can be persisted as plain text.
 
