### 2016-08-20 3efa7ac
crypto-plugin: rename plugin configuration keys
Rename /crypto/key to /gpg/key.
Rename /crypto/gpg to /gpg/bin.
