+++ src/plugins/crypto/gpg.h
@@ -13,8 +13,8 @@
 #include <kdb.h>
 #include <kdbtypes.h>
 
-#define ELEKTRA_CRYPTO_PARAM_GPG_BIN "/crypto/gpg"
-#define ELEKTRA_CRYPTO_PARAM_GPG_KEY "/crypto/key"
+#define ELEKTRA_CRYPTO_PARAM_GPG_BIN "/gpg/bin"
+#define ELEKTRA_CRYPTO_PARAM_GPG_KEY "/gpg/key"
 #define ELEKTRA_CRYPTO_DEFAULT_GPG_BIN "/usr/bin/gpg2"
 
 int elektraCryptoGpgEncryptMasterPassword (KeySet * conf, Key * errorKey, Key * msgKey);
