### 2016-08-20 daf8337
crypto-plugin: fix spelling in README.md
