### 2016-08-20 ed39515
crypto-plugin: correct boolean value of meta-key /crypto/encrypt
Considering false as default and '1' as true for the meta-key /crypto/encrypt.

See doc/decisions/bool.md for further details.
