### 2016-08-20 a1375b1
mozilla prefs:intercept example
