### 2016-08-20 e4c5d81
mozilla prefs: removed from shell unit tests
