### 2016-08-20 3c7f21d
mozilla prefs: added to plugins/README.md
