### 2016-08-20 4901c9b
mozilla prefs: ignore variables and functions (for now)
