+++ src/plugins/prefs/prefs.h
@@ -13,12 +13,8 @@
 #include <kdbplugin.h>
 
 
-int elektraPrefsOpen (Plugin * handle, Key * errorKey);
-int elektraPrefsClose (Plugin * handle, Key * errorKey);
 int elektraPrefsGet (Plugin * handle, KeySet * ks, Key * parentKey);
 int elektraPrefsSet (Plugin * handle, KeySet * ks, Key * parentKey);
-int elektraPrefsError (Plugin * handle, KeySet * ks, Key * parentKey);
-int elektraPrefsCheckConfig (Key * errorKey, KeySet * conf);
 
 Plugin * ELEKTRA_PLUGIN_EXPORT (prefs);
 
