### 2016-08-20 f4f2f1e
crypto-plugin: add GPG key for unit tests
GPG Key ID: DDEBEF9EE2DC931701338212DAF635B17F230E8D
