### 2016-08-20 d1e58ef
intercept: refactoring
