### 2016-08-20 fdb4035
doc: move some plugins to other category
remove outdated "New Plugins" information, unit test will fail anyway
