### 2016-08-21 ef23225
mozilla prefs: drop \"preferences\" part of keyname
