### 2016-08-21 e119bb8
mozilla prefs: intercept test setup script
