### 2016-08-21 34ec768
mozilla prefs: updated README
