### 2016-08-21 905d3c3
crypto-plugin: begin work on unit tests
Adapt unit tests to new GPG module.

Open Issues:

- Call the checkconf() function
