+++ src/plugins/crypto/gpg.h
@@ -15,6 +15,7 @@
 
 #define ELEKTRA_CRYPTO_PARAM_GPG_BIN "/gpg/bin"
 #define ELEKTRA_CRYPTO_PARAM_GPG_KEY "/gpg/key"
+#define ELEKTRA_CRYPTO_PARAM_GPG_UNIT_TEST "/gpg/unit_test"
 #define ELEKTRA_CRYPTO_DEFAULT_GPG_BIN "/usr/bin/gpg2"
 
 int elektraCryptoGpgEncryptMasterPassword (KeySet * conf, Key * errorKey, Key * msgKey);
