+++ src/plugins/crypto/testmod_crypto_openssl.c
@@ -18,8 +18,8 @@ int main (int argc, char ** argv)
 
 	init (argc, argv);
 
+	test_gpg ();
 	test_init (PLUGIN_NAME);
-	test_config_errors (PLUGIN_NAME);
 	test_crypto_operations (PLUGIN_NAME);
 
 	printf ("\n" PLUGIN_NAME " RESULTS: %d test(s) done. %d error(s).\n", nbTest, nbError);
