+++ src/plugins/crypto/test_key.h
@@ -1,167 +1,93 @@
-unsigned char test_key_asc[] = {
-  0x2d, 0x2d, 0x2d, 0x2d, 0x2d, 0x42, 0x45, 0x47, 0x49, 0x4e, 0x20, 0x50,
-  0x47, 0x50, 0x20, 0x50, 0x52, 0x49, 0x56, 0x41, 0x54, 0x45, 0x20, 0x4b,
-  0x45, 0x59, 0x20, 0x42, 0x4c, 0x4f, 0x43, 0x4b, 0x2d, 0x2d, 0x2d, 0x2d,
-  0x2d, 0x0a, 0x56, 0x65, 0x72, 0x73, 0x69, 0x6f, 0x6e, 0x3a, 0x20, 0x47,
-  0x6e, 0x75, 0x50, 0x47, 0x20, 0x76, 0x31, 0x0a, 0x0a, 0x6c, 0x51, 0x48,
-  0x59, 0x42, 0x46, 0x65, 0x34, 0x6b, 0x4d, 0x38, 0x42, 0x42, 0x41, 0x43,
-  0x7a, 0x79, 0x49, 0x2f, 0x7a, 0x55, 0x51, 0x4a, 0x64, 0x43, 0x2f, 0x6b,
-  0x44, 0x4b, 0x49, 0x49, 0x2f, 0x6a, 0x65, 0x54, 0x4b, 0x62, 0x38, 0x65,
-  0x6b, 0x38, 0x7a, 0x6c, 0x67, 0x58, 0x4a, 0x7a, 0x6c, 0x77, 0x36, 0x4b,
-  0x70, 0x4e, 0x6e, 0x4a, 0x66, 0x47, 0x52, 0x51, 0x57, 0x75, 0x6e, 0x2f,
-  0x6c, 0x0a, 0x4e, 0x67, 0x6c, 0x61, 0x55, 0x54, 0x77, 0x31, 0x5a, 0x4d,
-  0x70, 0x6b, 0x45, 0x4f, 0x78, 0x71, 0x2f, 0x35, 0x4e, 0x59, 0x44, 0x4a,
-  0x42, 0x4d, 0x35, 0x59, 0x56, 0x31, 0x33, 0x74, 0x6d, 0x71, 0x68, 0x57,
-  0x46, 0x4f, 0x39, 0x2f, 0x78, 0x2f, 0x6b, 0x6e, 0x34, 0x5a, 0x62, 0x67,
-  0x6a, 0x2b, 0x7a, 0x34, 0x77, 0x2f, 0x41, 0x55, 0x70, 0x4a, 0x51, 0x66,
-  0x74, 0x56, 0x75, 0x7a, 0x46, 0x66, 0x0a, 0x46, 0x63, 0x68, 0x50, 0x36,
-  0x71, 0x47, 0x6a, 0x69, 0x56, 0x74, 0x67, 0x64, 0x5a, 0x71, 0x6b, 0x64,
-  0x6f, 0x64, 0x47, 0x35, 0x7a, 0x41, 0x63, 0x54, 0x7a, 0x65, 0x59, 0x53,
-  0x6b, 0x4a, 0x61, 0x38, 0x39, 0x63, 0x43, 0x44, 0x6d, 0x58, 0x66, 0x4e,
-  0x31, 0x33, 0x6f, 0x37, 0x4a, 0x79, 0x70, 0x6a, 0x63, 0x44, 0x65, 0x68,
-  0x6b, 0x57, 0x69, 0x4a, 0x77, 0x41, 0x52, 0x41, 0x51, 0x41, 0x42, 0x0a,
[…]
