### 2016-08-21 6b6f16f
crypto-plugin: begin work on unit tests
Adapt unit tests to new GPG module.

Open Issues:

- Call the checkconf() function
