### 2016-08-21 483943c
doc: explain exported symbols
