### 2016-08-22 45213d8
crypto-plugin: finish unit tests
