### 2016-08-22 4184dda
crypto-plugin: finish unit tests
