### 2016-08-22 b43da41
ruby: add first test cases
add ruby tests to cmake and
some simple test cases for Kdb::Key
