### 2016-08-22 c3f59c1
crypto-plugin: mark test_key.h as generated
