+++ src/plugins/crypto/test_key.h
@@ -1,3 +1,8 @@
+/*
+ * this file has been generated by:
+ * xxd -i test_key.asc
+ */
+
 static const kdb_octet_t test_key_asc[] = {
 	0x2d, 0x2d, 0x2d, 0x2d, 0x2d, 0x42, 0x45, 0x47, 0x49, 0x4e, 0x20, 0x50, 0x47, 0x50, 0x20, 0x50, 0x52, 0x49, 0x56, 0x41, 0x54, 0x45,
 	0x20, 0x4b, 0x45, 0x59, 0x20, 0x42, 0x4c, 0x4f, 0x43, 0x4b, 0x2d, 0x2d, 0x2d, 0x2d, 0x2d, 0x0a, 0x56, 0x65, 0x72, 0x73, 0x69, 0x6f,
