### 2016-08-22 41f61f7
crypto-plugin: improve error descriptions in gpg.c
- Check existence of gpg binary
- Check if gpg binary has permission to execute
- Add specific error message if gpg reports bad signature
