### 2016-08-23 a944f7b
mozilla prefs: firefox.cfg console logging
