+++ src/plugins/prefs/autoconfig/firefox.cfg
@@ -1,10 +1,17 @@
 //
 
+
+var console = Components.classes["@mozilla.org/consoleservice;1"] .getService(Components.interfaces.nsIConsoleService);
+console.logStringMessage("Elektra autoconfig started");
 const configName = getPref("elektra.config.file");
+console.logStringMessage("Elektra autoconfig: config file: "+configName);
 const serverPort = getPref("elektra.config.reload_trigger_port");
+console.logStringMessage("Elektra autoconfig: Server port: "+serverPort);
 
 function reload()
 {
+
+    console.logStringMessage("Elektra autoconfig: reloading config file");
     let infile = Components.classes["@mozilla.org/file/local;1"].createInstance(Components.interfaces.nsILocalFile);
     infile.initWithPath(configName);
     var outfile = Components.classes["@mozilla.org/file/directory_service;1"].getService(Components.interfaces.nsIProperties).get("ProfD", Components.interfaces.nsIFile);
@@ -28,6 +35,7 @@ onSocketAccepted: function(serverSocket, clientSocket) {
 			  var bytes = sin.available();
 			  var request = "";
 			  request = sin.read(bytes);
+			  console.logStringMessage("Elektra autoconfig: received message: "+request);
 			  if (request.valueOf() == "reload"){
 			      reload();
 			  }
@@ -43,5 +51,7 @@ reload(); 	//read once on startup
 
 var serverSocket = Components.classes["@mozilla.org/network/server-socket;1"].createInstance(Components.interfaces.nsIServerSocket);
 serverSocket.init(serverPort, true, 1); 	//only listen on loopback
+
+console.logStringMessage("Elektra autoconfig: server socket started");
 
 serverSocket.asyncListen(listener);
