### 2016-08-23 d8a5e5f
mozilla prefs: setup scripts
