+++ src/plugins/prefs/autoconfig/autoconfig.js
@@ -1,2 +1,0 @@
-pref("general.config.filename", "firefox.cfg");
-pref("general.config.obscure_value", 0);
