### 2016-08-23 e1b354f
ruby: add ruby wrapper module kdb
the swig created binding module was renamed to 'kdb_' (_kdb is not possible
due to Ruby naming conventions), which is wrapped by a native Ruby module
'kdb'. This allows embedding Ruby code for easier overriding.

The kdb::Key constructor now supports "variable" "keyword" arguments, which
allows very flexible key creation.
