### 2016-08-24 c00e7e2
mozilla prefs: rename plugin, split and update README
