### 2016-08-24 d29f5ef
mozilla prefs: elektrify-open
