+++ src/plugins/mozprefs/autoconfig/README.md
@@ -4,7 +4,7 @@ This file explains how Firefox preferences can be manipulated at runtime using e
 
 ## Basics ##
 
-`scripts/easySetup.sh -s` will configure everything needed in get started.
+`scripts/configure-firefox.sh -s` will configure everything needed in get started.
 
 ### Files ###
 
@@ -17,7 +17,7 @@ This file explains how Firefox preferences can be manipulated at runtime using e
      pref("general.config.obscure_value", 0);
     ```
   - `elektra.cfg` Contains the autoconfig code
-- `~/.mozilla/firefox/<profile>/prefs.js` Contains the user preferences. The easySetup script will append the config for our autoconfig script:
+- `~/.mozilla/firefox/<profile>/prefs.js` Contains the user preferences. The configure-firefox script will append the config for our autoconfig script:
 
 
    ```
@@ -26,16 +26,17 @@ This file explains how Firefox preferences can be manipulated at runtime using e
    ```
 
    Every time the string `reload` is send to localhost:`elektra.config.reload_trigger_port` the autoconfig script will import the preferences from `elektra.config.file` 
-   `elektry.config.file` contains the name of the dummy file. `open` calls to this file will be intercepted and the automatically content generated by `intercept open`
+   
+   `elektry.config.file` contains the name of the dummy file for `intercept open`. `open` calls to this file will be intercepted and the content generated by `intercept open`
 
 ### Setting preferences ###
 
 #### Guided setup ####
 
-Running `scripts/easySetup -c` provides a guided setup for the http_proxy preference.
+Running `scripts/configure-firefox -c` provides a guided setup for the http_proxy preference.
 
 ```
-% scripts/easySetup.sh -c
+% scripts/configure-firefox.sh -c
 Add new preferences
 Config Setup:
 
@@ -81,7 +82,7 @@ This example shows how to manually setup a preferences. It's equivalent to the e
 
 ### Test Setup ###
 
-Running `scripts/easySetup -t` will set up some test values.
+Running `scripts/configure-firefox -t` will set up some test values.
 
 ```
 % kdb export /preload
@@ -115,10 +116,10 @@ user_pref("a.user.t", true);
 ```
 
 ```
-% LD_PRELOAD=/usr/local/lib/libelektraintercept.so firefox-esr "about:config"
+% kdb elektrify-open firefox-esr "about:config"
 ```
 
-![about:config before](.//config_1.jpg)
+![about:config before](./config_1.jpg)
 ```
 % kdb setmeta user/prefs/lock/a/lock/3 type boolean
 % kdb set user/prefs/lock/a/lock/3 true
@@ -135,7 +136,7 @@ user_pref("a.user.t", true);
 f = false
 t = true
 ```
-![about:config after](./autoconfig/config_2.jpg)
+![about:config after](./config_2.jpg)
 
 
 
