### 2016-08-24 0d60a16
mozilla prefs: updated configure-firefox.sh
