### 2016-08-24 06044d5
mozilla prefs: configure-firefox-sh dont create path variable twice
