### 2016-08-24 e5e4bdd
mozilla prefs: config tools
