+++ scripts/ffconfig/CMakeLists.txt
@@ -1,0 +1,4 @@
+install (PROGRAMS writeConfigFiles DESTINATION "${TARGET_TOOL_EXEC_FOLDER}/ffconfig/")
+install (PROGRAMS setupConfig DESTINATION "${TARGET_TOOL_EXEC_FOLDER}/ffconfig/")
+install (PROGRAMS setupProxy DESTINATION "${TARGET_TOOL_EXEC_FOLDER}/ffconfig/")
+install (PROGRAMS setupHomepage DESTINATION "${TARGET_TOOL_EXEC_FOLDER}/ffconfig/")
