+++ scripts/CMakeLists.txt
@@ -63,6 +63,12 @@ configure_file(
 	"${CMAKE_CURRENT_BINARY_DIR}/list-tools")
 install (PROGRAMS "${CMAKE_CURRENT_BINARY_DIR}/list-tools" DESTINATION ${TARGET_TOOL_EXEC_FOLDER})
 
+configure_file(
+	"${CMAKE_CURRENT_SOURCE_DIR}/configure-firefox.in"
+	"${CMAKE_CURRENT_BINARY_DIR}/configure-firefox" @ONLY)
+install (PROGRAMS "${CMAKE_CURRENT_BINARY_DIR}/configure-firefox" DESTINATION ${TARGET_TOOL_EXEC_FOLDER})
+add_subdirectory (ffconfig)
+
 ADD_CUSTOM_TARGET(source-package
 		COMMAND "${CMAKE_CURRENT_BINARY_DIR}/make-source-package"
 		COMMENT "Make a elektra source package by using git"
