### 2016-08-24 7c20322
ruby: follow ruby naming conventions
- use swig autorename feature
- define "Key::isXxx" methods as predicates (Key.is_xxx?)
- getter setter for name
