+++ src/bindings/swig/ruby/tests/testruby_key.rb
@@ -10,27 +10,27 @@ class KdbKeyTestCases < Test::Unit::TestCase
       k = Kdb::Key.new
       assert_not_nil k
       assert_instance_of Kdb::Key, k
-      assert_false k.isValid
+      assert_false k.is_valid?
 
       kname = "user/tmp/k1"
       k = Kdb::Key.new kname
       assert_not_nil k
-      assert_equal kname, k.getName
-      assert_true k.isValid
+      assert_equal kname, k.name
+      assert_true k.is_valid?
 
-      #puts "hello world - " + k.getName
+      #puts "hello world - " + k.name
     end
   end
 
   def test_key_new_valid_names
     assert_nothing_raised do
-      assert_true Kdb::Key.new("/cascading/tmp/ks1").isValid
-      assert_true Kdb::Key.new("spec/tmp/ks1").isValid
-      assert_true Kdb::Key.new("proc/tmp/ks1").isValid
-      assert_true Kdb::Key.new("dir/tmp/ks1").isValid
-      assert_true Kdb::Key.new("user/tmp/ks1").isValid
-      assert_true Kdb::Key.new("system/tmp/ks1").isValid
-      assert_false Kdb::Key.new("invalidname").isValid
+      assert_true Kdb::Key.new("/cascading/tmp/ks1").is_valid?
+      assert_true Kdb::Key.new("spec/tmp/ks1").is_valid?
+      assert_true Kdb::Key.new("proc/tmp/ks1").is_valid?
+      assert_true Kdb::Key.new("dir/tmp/ks1").is_valid?
+      assert_true Kdb::Key.new("user/tmp/ks1").is_valid?
+      assert_true Kdb::Key.new("system/tmp/ks1").is_valid?
+      assert_false Kdb::Key.new("invalidname").is_valid?
     end
   end
 
@@ -38,74 +38,74 @@ class KdbKeyTestCases < Test::Unit::TestCase
     assert_nothing_raised do
       name = "user/tmp/ks1"
       k = Kdb::Key.new name
-      assert_true k.isValid
-      assert_equal name, k.getName
-      assert_equal "", k.getString # TODO: should we return nil here
+      assert_true k.is_valid?
+      assert_equal name, k.name
+      assert_equal "", k.get_string # TODO: should we return nil here
 
       v = "val"
       k = Kdb::Key.new name, value: v
-      assert_true k.isValid
-      assert_equal name, k.getName
-      assert_equal v, k.getString
+      assert_true k.is_valid?
+      assert_equal name, k.name
+      assert_equal v, k.get_string
 
       k = Kdb::Key.new name, value: v, owner: "me", comment: "ccc"
-      assert_true k.isValid
-      assert_equal name, k.getName
-      assert_equal v, k.getString
-      assert_equal "me", k.getMeta("owner")
-      assert_equal "me", k.getMeta(:owner)
-      assert_equal "ccc", k.getMeta("comment")
-      assert_equal "ccc", k.getMeta(:comment)
+      assert_true k.is_valid?
+      assert_equal name, k.name
+      assert_equal v, k.get_string
+      assert_equal "me", k.get_meta("owner")
+      assert_equal "me", k.get_meta(:owner)
+      assert_equal "ccc", k.get_meta("comment")
+      assert_equal "ccc", k.get_meta(:comment)
 
       k = Kdb::Key.new(name, 
                        owner: "me", 
                        comment: "ccc",
                        flags: Kdb::KEY_BINARY)
-      assert_true k.isValid
-      assert_true k.isBinary
-      assert_equal name, k.getName
-      assert_equal "me", k.getMeta("owner")
[…]
