### 2016-08-24 86d9b40
ruby: add [] alias for metadata handling
