### 2016-08-24 f77873f
doc: also mention firefox
