### 2016-08-24 1f5dcc5
mozilla prefs: minor fixes
