### 2016-08-24 1037d90
crypto-plugin: add test-case to detect segfault
If the value does not fit into a single AES cipher block, a segfault occurs in gcrypt and OpenSSL variant
