### 2016-08-25 9cbc412
crypto-plugin: fix segfault in crypto_openssl
The segfault occured in the encryption method because of a faulty data partitioning.
