+++ src/plugins/crypto/openssl_operations.c
@@ -266,19 +266,15 @@ void elektraCryptoOpenSSLHandleDestroy (elektraCryptoHandle * handle)
 
 int elektraCryptoOpenSSLEncrypt (elektraCryptoHandle * handle, Key * k, Key * errorKey)
 {
-	// NOTE to prevent memory overflows in libcrypto the buffer holding the encrypted content
-	//      is one block bigger than the inupt buffer.
+	// NOTE to prevent memory overflows in libcrypto the buffer holding the encrypted content (cipherBuffer) is 2 cipher blocks long.
 	kdb_octet_t cipherBuffer[2 * ELEKTRA_CRYPTO_SSL_BLOCKSIZE];
-	kdb_octet_t contentBuffer[ELEKTRA_CRYPTO_SSL_BLOCKSIZE] = { 0 };
-	kdb_octet_t * output;
+	kdb_octet_t headerBuffer[ELEKTRA_CRYPTO_SSL_BLOCKSIZE];
 	int written = 0;
-	size_t outputLen;
 	BIO * encrypted;
-	const kdb_octet_t * value = (kdb_octet_t *)keyValue (k);
 
 	// prepare the crypto header data
-	const kdb_unsigned_long_t contentLen = keyGetValueSize (k);
 	kdb_octet_t flags;
+	const size_t contentLen = keyGetValueSize (k);
 	const size_t headerLen = sizeof (flags) + sizeof (contentLen);
 
 	switch (keyIsString (k))
@@ -305,9 +301,9 @@ int elektraCryptoOpenSSLEncrypt (elektraCryptoHandle * handle, Key * k, Key * er
 	}
 
 	// encrypt the header data
-	memcpy (contentBuffer, &flags, sizeof (flags));
-	memcpy (contentBuffer + sizeof (flags), &contentLen, sizeof (contentLen));
-	EVP_EncryptUpdate (&(handle->encrypt), cipherBuffer, &written, contentBuffer, headerLen);
+	memcpy (headerBuffer, &flags, sizeof (flags));
+	memcpy (headerBuffer + sizeof (flags), &contentLen, sizeof (contentLen));
+	EVP_EncryptUpdate (&(handle->encrypt), cipherBuffer, &written, headerBuffer, headerLen);
 	if (written > 0)
 	{
 		BIO_write (encrypted, cipherBuffer, written);
@@ -318,18 +314,21 @@ int elektraCryptoOpenSSLEncrypt (elektraCryptoHandle * handle, Key * k, Key * er
 		goto error;
 	}
 
-	// encrypt content block by block (i = start of the current block)
-	for (kdb_unsigned_long_t i = 0; i < contentLen; i += ELEKTRA_CRYPTO_SSL_BLOCKSIZE)
-	{
-		// load content partition into the content buffer
-		kdb_unsigned_long_t partitionLen = ELEKTRA_CRYPTO_SSL_BLOCKSIZE;
-		if ((i + 1) * ELEKTRA_CRYPTO_SSL_BLOCKSIZE > contentLen)
-		{
-			partitionLen = contentLen - (i * ELEKTRA_CRYPTO_SSL_BLOCKSIZE);
-		}
-		memcpy (contentBuffer, (value + i), partitionLen);
+	// encrypt content block by block
+	kdb_octet_t * content = (kdb_octet_t *)keyValue (k);
+	size_t processed = 0;
 
-		EVP_EncryptUpdate (&(handle->encrypt), cipherBuffer, &written, contentBuffer, partitionLen);
+	while (processed < contentLen)
+	{
+		kdb_unsigned_long_t partitionLen = ELEKTRA_CRYPTO_SSL_BLOCKSIZE;
+		// the last partition may not fill an entire cipher block
[…]
