### 2016-08-25 a333fa6
ruby: no kdb.rb any more, generate 'Kdb' directly
We left the approach of having generating a 'kdb_' module by SWIG and
use a Ruby native 'kdb' module which includes the generated one.
The main problem with this approach: all Keys created by the 'kdb_'
module (e.g. meta data keys) would have type 'Kdb_::Key' and thus
is missing the Ruby implemented additions of 'Kdb::Key'.

So we have left this path and took the harder way of implementing all
adoptions directly with C/SWIG.

Luckily we have found a really good solution for the va_list problem:
 * treat the va_list argument as Ruby Hash
 * 'mis
[…]
