### 2016-08-25 0638c41
crypto-plugin: fix segfault in crypto_gcrypt
Faulty data partitioning has been replaced by in-place encryption and decryption provided by gcrypt.
