+++ src/plugins/crypto/gcrypt_operations.c
@@ -264,15 +264,11 @@ error:
 
 int elektraCryptoGcryEncrypt (elektraCryptoHandle * handle, Key * k, Key * errorKey)
 {
-	const kdb_octet_t * value = (kdb_octet_t *)keyValue (k);
 	size_t outputLen;
 	gcry_error_t gcry_err;
 
-	kdb_octet_t * output;
-	kdb_octet_t cipherBuffer[ELEKTRA_CRYPTO_GCRY_BLOCKSIZE];
-	kdb_octet_t contentBuffer[ELEKTRA_CRYPTO_GCRY_BLOCKSIZE] = { 0 };
-
 	// prepare the crypto header data
+	const kdb_octet_t * content = keyValue (k);
 	const kdb_unsigned_long_t contentLen = keyGetValueSize (k);
 	kdb_octet_t flags;
 
@@ -300,69 +296,50 @@ int elektraCryptoGcryEncrypt (elektraCryptoHandle * handle, Key * k, Key * error
 		outputLen = (contentLen / ELEKTRA_CRYPTO_GCRY_BLOCKSIZE) + 2;
 	}
 	outputLen *= ELEKTRA_CRYPTO_GCRY_BLOCKSIZE;
-	output = elektraMalloc (outputLen);
-	if (output == NULL)
+	kdb_octet_t * output = elektraMalloc (outputLen);
+	if (!output)
 	{
 		ELEKTRA_SET_ERROR (87, errorKey, "Memory allocation failed");
 		return (-1);
 	}
 
-	// encrypt the header (1st block)
-	memcpy (contentBuffer, &flags, sizeof (flags));
-	memcpy (contentBuffer + sizeof (flags), &contentLen, sizeof (contentLen));
-	gcry_err = gcry_cipher_encrypt (*handle, cipherBuffer, ELEKTRA_CRYPTO_GCRY_BLOCKSIZE, contentBuffer, ELEKTRA_CRYPTO_GCRY_BLOCKSIZE);
+	// encrypt the header (1st block) using gcrypt's in-place encryption
+	memcpy (output, &flags, sizeof (flags));
+	memcpy (output + sizeof (flags), &contentLen, sizeof (contentLen));
+	gcry_err = gcry_cipher_encrypt (*handle, output, ELEKTRA_CRYPTO_GCRY_BLOCKSIZE, NULL, 0);
 	if (gcry_err != 0)
 	{
 		ELEKTRA_SET_ERRORF (ELEKTRA_ERROR_CRYPTO_ENCRYPT_FAIL, errorKey, "Encryption failed because: %s", gcry_strerror (gcry_err));
+		memset (output, 0, outputLen);
 		elektraFree (output);
 		return (-1);
 	}
-	memcpy (output, cipherBuffer, ELEKTRA_CRYPTO_GCRY_BLOCKSIZE);
 
-	// encrypt content block by block (i = start of the current block)
-	for (kdb_unsigned_long_t i = 0; i < contentLen; i += ELEKTRA_CRYPTO_GCRY_BLOCKSIZE)
+	// encrypt the value using gcrypt's in-place encryption
+	kdb_octet_t * dataOut = output + ELEKTRA_CRYPTO_GCRY_BLOCKSIZE;
+	const size_t dataLen = outputLen - ELEKTRA_CRYPTO_GCRY_BLOCKSIZE;
+	memcpy (dataOut, content, contentLen);
+	gcry_err = gcry_cipher_encrypt (*handle, dataOut, dataLen, NULL, 0);
+	if (gcry_err != 0)
 	{
-		// load content partition into the content buffer
[…]
