### 2016-08-25 f789eab
extra install path on darwin
