### 2016-08-25 eae0ad5
extra install path on darwin
