### 2016-08-25 71ff1be
crypto-plugin: fix crypto header validation in crypto_openssl
