### 2016-08-25 a86bb1f
crypto-plugin: add further unit tests
- Add value filling up a single cipher block
- Add value filling up two cipher blocks exactly
