### 2016-08-25 cf3248d
crypto-plugin: lookup gpg bin in PATH
