+++ src/plugins/crypto/gpg.c
@@ -41,23 +41,141 @@ static int inTestMode (KeySet * conf)
 }
 
 /**
- * @brief lookup the path to the gpg binary in conf.
- * @param conf KeySet holding the plugin configuration.
- * @returns the path to the gpg binary to use. This value must not be modified!
+ * @brief concatenates dir and file.
+ * @param errorKey holds an error description in case of failure.
+ * @param dir contains the path to a directory
+ * @param file contains a file name
+ * @returns an allocated string containing "dir/file" which must be freed by the caller or NULL in case of error.
  */
-static char * getGpgBinary (KeySet * conf)
+static char * genGpgCandidate (Key * errorKey, char * dir, const char * file)
 {
+	const size_t resultLen = strlen (dir) + strlen (file) + 2;
+	char * result = elektraMalloc (resultLen);
+	if (!result)
+	{
+		ELEKTRA_SET_ERROR (87, errorKey, "Memory allocation failed");
+		return NULL;
+	}
+	snprintf (result, resultLen, "%s/%s", dir, file);
+	return result;
+}
+
+/**
+ * @brief lookup binary file bin in the PATH environment variable.
+ * @param errorKey holds an error description in case of failure.
+ * @param bin the binary file to look for
+ * @param result holds an allocated string containing the full path to the binary file or NULL in case of error. Must be freed by the caller.
+ * @retval -1 if an error occured. See errorKey for a description.
+ * @retval 0 if the binary could not be found within PATH.
+ * @retval 1 if the binary was found and the full path was stored in result.
+ */
+static int searchPathForBin (Key * errorKey, const char * bin, char ** result)
+{
+	*result = NULL;
+
+	const char * envPath = getenv ("PATH");
+	if (envPath)
+	{
+		const size_t envPathLen = strlen (envPath) + 1;
+		char * dir;
+
+		char * path = elektraMalloc (envPathLen);
+		if (!path)
+		{
+			ELEKTRA_SET_ERROR (87, errorKey, "Memory allocation failed");
+			return -1;
+		}
+		memcpy (path, envPath, envPathLen);
+		// save start of path as strsep() modifies path while splitting it up
+		char * pathBegin = path;
+		while ((dir = strsep (&path, ":")) != NULL)
+		{
+			char * candidate = genGpgCandidate (errorKey, dir, bin);
+			if (!candidate)
+			{
+				elektraFree (pathBegin);
+				return -1;
+			}
+			if (access (candidate, X_OK) == 0)
+			{
+				*result = candidate;
+				elektraFree (pathBegin);
+				return 1;
+			}
+			elektraFree (candidate);
+		}
+		elektraFree (pathBegin);
+	}
+	return 0;
+}
+
+/**
+ * @brief lookup the path to the gpg binary in conf.
+ * @param gpgBin holds allocated path to the gpg binary to be used or NULL in case of an error. Must bee freed by the caller.
+ * @param conf KeySet holding the plugin configuration.
+ * @param errorKey holds an error description if something goes wrong.
+ * @retval 1 on success.
[…]
