+++ src/error/CMakeLists.txt
@@ -10,7 +10,7 @@ find_util (elektra-export-errors EXE_ERR_LOC EXE_ERR_ARG)
 set (BINARY_INCLUDE_DIR "${PROJECT_BINARY_DIR}/src/include")
 add_custom_command (
 		OUTPUT ${BINARY_INCLUDE_DIR}/kdberrors.h
-		DEPENDS elektra-export-errors
+		DEPENDS elektra-export-errors ${CMAKE_CURRENT_SOURCE_DIR}/specification
 		COMMAND ${EXE_ERR_LOC}
 		ARGS ${EXE_ERR_ARG} ${CMAKE_SOURCE_DIR}/src/error/specification ${BINARY_INCLUDE_DIR}/kdberrors.h
 		)
