### 2016-08-28 2061397
Build: Use shorter version of filepath
