### 2016-08-29 67aa77d
crypto-plugin: rename error macros
Remove '_FAULT' suffix from crypto error macro names
