### 2016-08-29 c051e4f
crypto-plugin: collect common files in variable in CMakeLists.txt
The following files are stored in CRYPTO_COMMON_FILES:

- helper.c
- helper.h
- gpg.c
- gpg.h
- crypto.c
- crypto.h
