### 2016-08-30 a697eae
crypto-plugin: update README.md
Specify GPG dependency
