### 2016-08-30 9e18629
crypto-plugin: improve error handling in gpg.c
Redirect stderr output of gpg to pipe, append it to errorKey.

Run gpg in batch mode, so no interactions (especially no reads from stdin) take place.
