### 2016-08-30 2801e13
crypto-plugin: fix clang compiler warning in botan_operations.cpp
Replace old cast with reinterpret_cast
