### 2016-08-30 5c98992
conditionals: fixed README example
