### 2016-08-30 8ded447
intercept_open: renamed config path
