### 2016-08-31 c7d179d
getenv.cpp: added no config path /elektra/intercept/getenv + keep /env as fallback
