+++ src/libs/getenv/src/getenv.cpp
@@ -71,6 +71,10 @@ extern "C" {
 Key * elektraParentKey;
 KeySet * elektraConfig;
 KDB * elektraRepo;
+Key * elektraFallbackParentKey;
+KeySet * elektraFallbackConfig;
+KDB * elektraFallbackRepo;
+
 } // extern "C"
 
 namespace
@@ -199,7 +203,7 @@ void addOverride (string kv)
 	getline (ss, v);
 	LOG << "add override " << k << " with " << v << endl;
 
-	string fullName = "proc/env/override/";
+	string fullName = "proc/elektra/intercept/getenv/override/";
 	fullName += k;
 	ksAppendKey (elektraConfig, keyNew (fullName.c_str (), KEY_VALUE, v.c_str (), KEY_END));
 }
@@ -213,7 +217,7 @@ void addOption (string kv)
 	getline (ss, v);
 	LOG << "add option " << k << " with " << v << endl;
 
-	string fullName = "proc/env/option/";
+	string fullName = "proc/elektra/intercept/getenv/option/";
 	fullName += k;
 	ksAppendKey (elektraConfig, keyNew (fullName.c_str (), KEY_VALUE, v.c_str (), KEY_END));
 }
@@ -227,7 +231,7 @@ void addLayer (string kv)
 	getline (ss, v);
 	LOG << "add layer " << k << " with " << v << endl;
 
-	string fullName = "proc/env/layer/";
+	string fullName = "proc/elektra/intercept/getenv/layer/";
 	fullName += k;
 	ksAppendKey (elektraConfig, keyNew (fullName.c_str (), KEY_VALUE, v.c_str (), KEY_END));
 }
@@ -238,8 +242,8 @@ void giveName (string name)
 	std::string basename = ::basename (n);
 	elektraFree (n);
 	LOG << "give name " << name << ", basename: " << basename << std::endl;
-	ksAppendKey (elektraConfig, keyNew ("proc/env/layer/name", KEY_VALUE, name.c_str (), KEY_END));
-	ksAppendKey (elektraConfig, keyNew ("proc/env/layer/basename", KEY_VALUE, basename.c_str (), KEY_END));
+	ksAppendKey (elektraConfig, keyNew ("proc/elektra/intercept/getenv/layer/name", KEY_VALUE, name.c_str (), KEY_END));
+	ksAppendKey (elektraConfig, keyNew ("proc/elektra/intercept/getenv/layer/basename", KEY_VALUE, basename.c_str (), KEY_END));
 }
 
 void parseArgs (int * argc, char ** argv)
@@ -299,7 +303,7 @@ void addEnvironment (string kv)
 	getline (ss, v);
 	LOG << "add option " << k << " with " << v << endl;
 
-	string fullName = "proc/env/option/";
+	string fullName = "proc/elektra/intercept/getenv/option/";
 	fullName += k;
 	ksAppendKey (elektraConfig, keyNew (fullName.c_str (), KEY_VALUE, v.c_str (), KEY_END));
 }
@@ -328,7 +332,25 @@ void addLayers ()
 	Key * c;
 	KeySet * lookupConfig = ksDup (elektraConfig);
 	ksRewind (elektraConfig);
-	std::string prefix = "/env/layer/";
+	std::string fallbackPrefix = "/env";
+	while ((c = ksNext (elektraConfig)))
+	{
+		std::string fullName = keyName (c);
+		size_t pos = fullName.find ('/');
+		if (pos != string::npos && fullName.substr (pos, fallbackPrefix.size ()) == fallbackPrefix)
+		{
+			std::string cascadingName = fullName.substr (fullName.find ('/'));
[…]
