### 2016-08-31 4467975
test_getenv.cpp: adjusted tests to new config path
