### 2016-08-31 7b1970b
getenv test_context.cpp: adjusted tests to new config
