+++ src/libs/getenv/tests/test_context.cpp
@@ -21,8 +21,8 @@ TEST (Context, Exist)
 {
 	Key * k;
 	elektraOpen (nullptr, nullptr);
-	ksAppendKey (elektraConfig, k = keyNew ("user/env/override/does-exist", KEY_VALUE, "hello", KEY_END));
-	Key * f = elektraLookupWithContext ("user/env/override/does-exist");
+	ksAppendKey (elektraConfig, k = keyNew ("user/elektra/intercept/getenv/override/does-exist", KEY_VALUE, "hello", KEY_END));
+	Key * f = elektraLookupWithContext ("user/elektra/intercept/getenv/override/does-exist");
 	EXPECT_EQ (k, f);
 }
 
@@ -31,10 +31,10 @@ TEST (Context, ExistWithContext)
 {
 	Key * k;
 	elektraOpen (nullptr, nullptr);
-	ksAppendKey (elektraConfig, k = keyNew ("user/env/override/does-exist", KEY_VALUE, "hello", KEY_END));
-	ksAppendKey (elektraConfig, keyNew ("spec/env/override/does-exist", KEY_META, "context", "user/env/override/does-exist", KEY_VALUE,
+	ksAppendKey (elektraConfig, k = keyNew ("user/elektra/intercept/getenv/override/does-exist", KEY_VALUE, "hello", KEY_END));
+	ksAppendKey (elektraConfig, keyNew ("spec/elektra/intercept/getenv/override/does-exist", KEY_META, "context", "user/elektra/intercept/getenv/override/does-exist", KEY_VALUE,
 					    "hello", KEY_END));
-	Key * f = elektraLookupWithContext ("/env/override/does-exist");
+	Key * f = elektraLookupWithContext ("/elektra/intercept/getenv/override/does-exist");
 	EXPECT_EQ (k, f);
 }
 
@@ -43,10 +43,10 @@ TEST (Context, ExistWithContextCascading)
 {
 	Key * k;
 	elektraOpen (nullptr, nullptr);
-	ksAppendKey (elektraConfig, k = keyNew ("user/env/override/does-exist-too", KEY_VALUE, "hello", KEY_END));
-	ksAppendKey (elektraConfig, keyNew ("spec/env/override/does-exist", KEY_META, "context", "/env/override/does-exist-too", KEY_VALUE,
+	ksAppendKey (elektraConfig, k = keyNew ("user/elektra/intercept/getenv/override/does-exist-too", KEY_VALUE, "hello", KEY_END));
+	ksAppendKey (elektraConfig, keyNew ("spec/elektra/intercept/getenv/override/does-exist", KEY_META, "context", "/elektra/intercept/getenv/override/does-exist-too", KEY_VALUE,
 					    "hello", KEY_END));
-	Key * f = elektraLookupWithContext ("/env/override/does-exist");
+	Key * f = elektraLookupWithContext ("/elektra/intercept/getenv/override/does-exist");
 	EXPECT_EQ (k, f);
 }
 
@@ -59,14 +59,14 @@ TEST (Context, ExistWithContextOverrideCascading)
 {
 	Key * k;
 	elektraOpen (nullptr, nullptr);
-	ksAppendKey (elektraConfig, keyNew ("user/env/layer/layer", KEY_VALUE, "layer", KEY_END));
+	ksAppendKey (elektraConfig, keyNew ("user/elektra/intercept/getenv/layer/layer", KEY_VALUE, "layer", KEY_END));
 	addLayers ();
 
-	ksAppendKey (elektraConfig, keyNew ("user/env/override/does-exist-too", KEY_VALUE, "wrong", KEY_END));
[…]
