### 2016-08-31 d7cbe72
test_getenv.cpp: added fallback config tests
