### 2016-08-31 d4ad744
blockresolver: README.md
