### 2016-08-31 dcda87e
blockresolver: removed unused functions + added to plugins/README.md
