+++ src/plugins/blockresolver/blockresolver.h
@@ -13,12 +13,9 @@
 #include <kdbplugin.h>
 
 
-int elektraBlockresolverOpen (Plugin * handle, Key * errorKey);
 int elektraBlockresolverClose (Plugin * handle, Key * errorKey);
 int elektraBlockresolverGet (Plugin * handle, KeySet * ks, Key * parentKey);
 int elektraBlockresolverSet (Plugin * handle, KeySet * ks, Key * parentKey);
-int elektraBlockresolverError (Plugin * handle, KeySet * ks, Key * parentKey);
-int elektraBlockresolverCheckConfig (Key * errorKey, KeySet * conf);
 
 Plugin * ELEKTRA_PLUGIN_EXPORT (blockresolver);
 
