### 2016-08-31 67286c0
blockresolver: fixed buffer size
