### 2016-08-31 88a44d3
blockresolver: added unittest
