### 2016-08-31 db7caf0
blockresolver: remove temporary file
