+++ src/libs/getenv/tests/test_context.cpp
@@ -32,8 +32,8 @@ TEST (Context, ExistWithContext)
 	Key * k;
 	elektraOpen (nullptr, nullptr);
 	ksAppendKey (elektraConfig, k = keyNew ("user/elektra/intercept/getenv/override/does-exist", KEY_VALUE, "hello", KEY_END));
-	ksAppendKey (elektraConfig, keyNew ("spec/elektra/intercept/getenv/override/does-exist", KEY_META, "context", "user/elektra/intercept/getenv/override/does-exist", KEY_VALUE,
-					    "hello", KEY_END));
+	ksAppendKey (elektraConfig, keyNew ("spec/elektra/intercept/getenv/override/does-exist", KEY_META, "context",
+					    "user/elektra/intercept/getenv/override/does-exist", KEY_VALUE, "hello", KEY_END));
 	Key * f = elektraLookupWithContext ("/elektra/intercept/getenv/override/does-exist");
 	EXPECT_EQ (k, f);
 }
@@ -44,8 +44,8 @@ TEST (Context, ExistWithContextCascading)
 	Key * k;
 	elektraOpen (nullptr, nullptr);
 	ksAppendKey (elektraConfig, k = keyNew ("user/elektra/intercept/getenv/override/does-exist-too", KEY_VALUE, "hello", KEY_END));
-	ksAppendKey (elektraConfig, keyNew ("spec/elektra/intercept/getenv/override/does-exist", KEY_META, "context", "/elektra/intercept/getenv/override/does-exist-too", KEY_VALUE,
-					    "hello", KEY_END));
+	ksAppendKey (elektraConfig, keyNew ("spec/elektra/intercept/getenv/override/does-exist", KEY_META, "context",
+					    "/elektra/intercept/getenv/override/does-exist-too", KEY_VALUE, "hello", KEY_END));
 	Key * f = elektraLookupWithContext ("/elektra/intercept/getenv/override/does-exist");
 	EXPECT_EQ (k, f);
 }
@@ -63,9 +63,11 @@ TEST (Context, ExistWithContextOverrideCascading)
 	addLayers ();
 
 	ksAppendKey (elektraConfig, keyNew ("user/elektra/intercept/getenv/override/does-exist-too", KEY_VALUE, "wrong", KEY_END));
-	ksAppendKey (elektraConfig, k = keyNew ("user/elektra/intercept/getenv/override/layer/does-exist-too", KEY_VALUE, "correct", KEY_END));
-	ksAppendKey (elektraConfig, keyNew ("spec/elektra/intercept/getenv/override/does-exist", KEY_META, "context", "/elektra/intercept/getenv/override/%layer%/does-exist-too",
-					    KEY_META, "override/#0", "/elektra/intercept/getenv/override/does-exist-too", KEY_VALUE, "hello", KEY_END));
+	ksAppendKey (elektraConfig,
+		     k = keyNew ("user/elektra/intercept/getenv/override/layer/does-exist-too", KEY_VALUE, "correct", KEY_END));
+	ksAppendKey (elektraConfig, keyNew ("spec/elektra/intercept/getenv/override/does-exist", KEY_META, "context",
+					    "/elektra/intercept/getenv/override/%layer%/does-exist-too", KEY_META, "override/#0",
+					    "/elektra/intercept/getenv/override/does-exist-too", KEY_VALUE, "hello", KEY_END));
 	Key * f = elektraLookupWithContext ("/elektra/intercept/getenv/override/does-exist");
 	EXPECT_EQ (k, f);
 }
@@ -79,9 +81,11 @@ TEST (Context, ExistWithContextOverrideCascadingSystem)
 	addLayers ();
 
[…]
