### 2016-08-31 4a3caca
ruby: streamline kdb.i and supress some warnings
do not include swig features/library files which aren't used.
including 'stl.i' pulls in 'rubyiterator.swg' which isn't really
needed, therefore this only bloats our bindings.

suppress 'unused-parameter' warnings of gcc: argv list of wrapper
functions for methods without arguments will be never touched, however
Ruby C-API requires them.
