### 2016-08-31 fb561ed
global hooks: keyset helper function
