### 2016-08-31 351139a
adjusted examples/spec to new conditionals syntax
