+++ examples/spec/locationtracker.ini
@@ -34,5 +34,5 @@ description = GPS polling time in seconds
 mountpoint = gps.conf
 
 [gps/status]
-assign/condition = (battery/level > 'low') ?  ('active') : ('inactive')
+assign/condition = (../../battery/level > 'low') ?  ('active') : ('inactive')
 description = current status of GPS
