### 2016-08-31 6a4fae2
ruby: add Key name comperators
