### 2016-08-31 130de4c
conditionals: now with topological sort
