### 2016-08-31 1791d2a
create initial version of the elektra web proposal
