### 2016-09-01 db66c58
first steps and keyset more in depth explained
