+++ doc/tutorials/python-kdb.md
@@ -6,13 +6,56 @@ When programming in python it is possible to access the kdb database, changing v
 
 ## First Steps
 
-In order to being able to use kdb, obviously, you at first need to `import kdb`. You need access to an python-object of kdb. This is accomlished by calling `kdb.KDB()` and saving this to a variable because later on this object will be needed for various operations.
+In order to being able to use kdb, obviously, you at first need to `import kdb`. You need access to an python-object of kdb. This is accomlished by calling `kdb.KDB()` and saving this to a variable because later on this object will be needed for various operations. 
+The easiest way to do this would be:
+```
+import kdb
+with kdb.KDB() as k:
+	print 'Hello world! I have a kdb-instance! :D'
+	# do all kinds of operations explained below
+```
 
 ## Keyset
 
-A keyset is basically a list of the keys that lie within the specified range. At first it is necessairy to create a new keyset. When simply calling `kdb.KeySet()` the keyset is of undefined size. It is possible to specify a certain (maximum) size for a keyset. To load keys into to keyset from the database you simply call the method `get` provided by the kdb-object.
+A keyset is basically a list of the keys that lie within the specified range. At first it is necessary to create a new keyset. When simply calling `kdb.KeySet()` the keyset is of size 0. There is no restriction to the keyest's size. It is possible to specify a certain (maximum) size for a keyset. To load keys into to keyset from the database you simply call the method `get` provided by the kdb-object.
+
+```
+import kdb
+with kdb.KDB() as k:
+	# create empty keyset
+	ks = kdb.KeySet()
+	print len(ks) # should be 0
+	# load an existing keyset
+	k.get(ks, 'path/to/keys')
+```
 
 It is also possible to iterate completely normally over a keyset and use len and reversed, dup and deepcopy. The elements of a keyset can be accessed by indexes and a keyset can be sliced. Another way of accessing a key is by the key-name (`keyset_name['path/to/keys/key_name']`). If the key-name does not exist within the keyset, a KeyError exception is thrown.
+
+An example that shows how to load an existing keyset and then access every key and value of the loaded keyset:
+
+```
+import kdb
+with kdb.KDB() as k:
+	ks = kdb.KeySet()
+	#loading an existing keyset
+	k.get(ks, 'path/to/keys')
+	for i in ks:
+		# print for every key in the keyset the key and the value
+		print 'key: ' + i + 'value: ' + ks[i]
+```
+
+Here an example of how you can easily check if a key exists:
+
+```
+import kdb
+with kdb.KDB() as k:
+	ks = kdb.KeySet()
+	k.get(ks, 'path/to/keys')
+	try:
+                print 'The value to the key user/sw/pk/key_name is ' + ks["user/sw/pk/key_name"] + '!'
+        except KeyError:
+                print 'The key does not exist!'
+```
 
 If you have changed anything in the keyset and want those changes to be saved to the database, you need to call `set` which is just like `get` provided by the kdb-object.
 
@@ -22,8 +65,10 @@ import kdb
 with kdb.KDB() as k:
 	ks = kdb.KeySet()
 	k.get(ks, 'path/to/keys')
-	... # any number of operations that manipulate the keyset
-	k.set(ks, 'path/to/keys') # by changing the path here it is for instance possible to set the keyset of one user identical to another users
+	# ... any number of operations that manipulate the keyset as explained below
+	# setting and by doing so saving the new keyset
+	k.set(ks, 'path/to/keys') 
+	# by changing the path here it is for instance possible to set the keyset of one user identical to another users
 ```
 
 ## Operations with keys
