+++ doc/webui/README.md
@@ -7,8 +7,8 @@ _a web user interface (Web UI) to remotely manage multiple elektra instances_
 
 Elektra web consists of multiple components:
 
- * (multiple) servers running an elektra daemon (elektrad)
- * a single cluster management server to communicate with the elektra daemons (clusterd)
+ * (multiple) servers running an elektra daemon (`elektrad`)
+ * a single cluster management server to communicate with the elektra daemons (`clusterd`)
  * a client (web browser) that accesses the Web UI on the cluster management server
 
 ![https://cdn.rawgit.com/omnidan/libelektra/http-api-proposal/doc/webui/network_structure.png](https://cdn.rawgit.com/omnidan/libelektra/http-api-proposal/doc/webui/network_structure.png)
@@ -31,20 +31,28 @@ The configuration view of elektra web is similar to the tree view of the
 
 ![https://cdn.rawgit.com/omnidan/libelektra/http-api-proposal/doc/webui/daemon_structure.png](https://cdn.rawgit.com/omnidan/libelektra/http-api-proposal/doc/webui/daemon_structure.png)
 
-TODO: change second elektrad to clusterd
-
-To access single instances, each elektra daemon (elektrad) provides a RESTful
+To access single instances, each elektra daemon (`elektrad`) provides a RESTful
 HTTP API:
 
- * TODO
+ * **GET /version** - get `elektrad` version
+ * **GET /kdb/:path** - get `path` configuration (similar to `kdb get path`)
+ * **POST /kdb/:path** - edit `path` configuration (similar to `kdb set path`)
 
-The cluster management server (clusterd) also provides a RESTful HTTP API.
+The cluster management server (`clusterd`) also provides a RESTful HTTP API.
 Single instances can be configured as follows:
 
- * TODO
+ * **GET /instances** - get a list of all instances
+ * **POST /instances** - create a new instance
+ * **GET /instances/:id** - get information about a single instance
+ * **POST /instances/:id** - edit a single instance
+ * **GET /instances/:id/kdb** - get full configuration of an instance
+ * **GET /instances/:id/kdb/:path** - get `path` configuration of an instance (similar to `kdb get path`)
+ * **POST /instances/:id/kdb/:path** - edit `path` configuration of an instance (similar to `kdb set path`)
+ * **GET /instances/:id/version** - get `elektrad` version of an instance
 
 It is also possible to create and manage groups of multiple elektra instances (clusters):
 
+ * **GET /version** - get `clusterd` version
  * **GET /clusters** - get a list of all clusters
  * **POST /clusters** - create a new cluster
  * **GET /clusters/:id** - get information about a single cluster
@@ -52,5 +60,7 @@ It is also possible to create and manage groups of multiple elektra instances (c
  * **GET /clusters/:id/kdb** - get full configuration of a cluster
  * **GET /clusters/:id/kdb/:path** - get `path` configuration of a cluster (similar to `kdb get path`)
  * **POST /clusters/:id/kdb/:path** - edit `path` configuration of a cluster (similar to `kdb set path`)
+ * **GET /clusters/:id/version** - get list of `elektrad` versions of all instances in the cluster
 
-TODO: blueprint
+Cluster configuration is stored on the cluster management server and persisted
+to all instances.
