+++ src/plugins/gitresolver/README.md
@@ -5,7 +5,7 @@
 - infos/provides = resolver
 - infos/recommends =
 - infos/placements = rollback getresolver setresolver commit
-- infos/status = recommended productive maintained reviewed conformant compatible coverage specific unittest shelltest tested nodep libc configurable final preview 
+- infos/status = recommended productive maintained reviewed conformant compatible coverage specific unittest shelltest tested libc configurable final preview 
 - infos/metadata =
 - infos/description =
 
