+++ src/plugins/timeofday/README.md
@@ -4,7 +4,7 @@
 - infos/provides = tracing
 - infos/needs =
 - infos/placements = pregetstorage postgetstorage presetstorage precommit postcommit prerollback postrollback
-- infos/status = maintained tested
+- infos/status = maintained tested nodep
 - infos/description = Prints timestamps during execution of backend
 
 ## Introduction ##
