+++ src/plugins/uname/README.md
@@ -4,7 +4,7 @@
 - infos/provides = storage info
 - infos/needs =
 - infos/placements = getstorage setstorage
-- infos/status = maintained unittest shelltest concept
+- infos/status = maintained unittest shelltest concept nodep
 - infos/description = Includes uname information into the key database.
 
 ## Introduction ##
