### 2016-09-01 20039ac
crypto-plugin: exclude spec namespace from operation
