### 2016-09-01 de07c47
crypto-plugin: fix clang compiler warning
in gcrypt_operations.c a VLA has been folded into a constant array
