### 2016-09-03 939aaa4
crypto-plugin: fix compile error under OpenBSD
Fix missing path detection of libgcrypt include files under OpenBSD 6.0
