### 2016-09-03 2e0b9d5
crypto-plugin: fix clang compiler warning
in openssl_operations.c a VLA has been folded into a constant array
