### 2016-09-05 1e4e3ca
use PUT, standardize /instances and /clusters API
