### 2016-09-05 5fc13ee
add DELETE API
