### 2016-09-05 d4dce2a
add info about response codes
