### 2016-09-05 60bdfae
decision: api documentation - added issue references + small formulation change
