### 2016-09-05 3fdb735
Ruby: add exception handling
Instruct Swig to use exception classes as exceptions and
pass them correctly.
