### 2016-09-05 d4a53c1
Ruby: be more Ruby on "name" methods
