### 2016-09-05 8cd561d
crypto-plugin: improve in-code documentation
- consistent use of metakey
- add missing method documentation
- explain guards and mutexes
- replace magic number in gpg.c
