+++ src/plugins/crypto/openssl_operations.c
@@ -25,6 +25,11 @@
 
 #define KEY_BUFFER_SIZE (ELEKTRA_CRYPTO_SSL_KEYSIZE + ELEKTRA_CRYPTO_SSL_BLOCKSIZE)
 
+/*
+ * Protects all calls to OpenSSL (libcrypto.so).
+ *
+ * This is required because we can not setup the multi-threading capabilities of OpenSSL directly from within Elektra.
+ */
 static pthread_mutex_t mutex_ssl = PTHREAD_MUTEX_INITIALIZER;
 
 /**
@@ -110,7 +115,7 @@ static int getKeyIvForDecryption (KeySet * config, Key * errorKey, Key * k, Key 
 	const Key * salt = keyGetMeta (k, ELEKTRA_CRYPTO_META_SALT);
 	if (!salt)
 	{
-		ELEKTRA_SET_ERRORF (ELEKTRA_ERROR_CRYPTO_CONFIG_FAULT, errorKey, "missing salt as meta-key %s for key %s",
+		ELEKTRA_SET_ERRORF (ELEKTRA_ERROR_CRYPTO_CONFIG_FAULT, errorKey, "missing salt as metakey %s for key %s",
 				    ELEKTRA_CRYPTO_META_SALT, keyName (k));
 		return -1;
 	}
@@ -481,6 +486,11 @@ error:
 	return (-1);
 }
 
+/**
+ * @brief create a random sequence of characters with given length.
+ * @param length the number of random bytes to be generated.
+ * @returns allocated buffer holding length bytes. Must be freed by the caller.
+  */
 char * elektraCryptoOpenSSLCreateRandomString (const kdb_unsigned_short_t length)
 {
 	kdb_octet_t * buffer = elektraMalloc (length);
