### 2016-09-05 3dadcef
crypto-plugin: fix inconsistent style in return values
replaces (-1) with -1
