### 2016-09-05 fa57d55
crypto-plugin: rename ELEKTRA_CRYPTO_PARAM_MASTER_PWD to ELEKTRA_CRYPTO_PARAM_MASTER_PASSWORD
