+++ src/plugins/crypto/crypto.h
@@ -32,8 +32,8 @@ enum ElektraCryptoOperation
 #define ELEKTRA_CRYPTO_DEFAULT_SALT_LEN (17)
 
 // plugin configuration parameters
-#define ELEKTRA_CRYPTO_PARAM_MASTER_PWD_LEN "/crypto/masterpasswordlength"
-#define ELEKTRA_CRYPTO_PARAM_MASTER_PWD "/crypto/masterpassword"
+#define ELEKTRA_CRYPTO_PARAM_MASTER_PASSWORD_LEN "/crypto/masterpasswordlength"
+#define ELEKTRA_CRYPTO_PARAM_MASTER_PASSWORD "/crypto/masterpassword"
 #define ELEKTRA_CRYPTO_PARAM_KEY_PATH "/crypto/key"
 #define ELEKTRA_CRYPTO_PARAM_IV_PATH "/crypto/iv"
 #define ELEKTRA_CRYPTO_PARAM_SHUTDOWN "/shutdown"
