### 2016-09-05 93a40db
crypto-plugin: remove unused configuration parameter from crypto.h
