### 2016-09-05 96f0b09
Ruby: add Key.clone|dup
