+++ src/bindings/swig/ruby/tests/testruby_key.rb
@@ -156,6 +156,13 @@ class KdbKeyTestCases < Test::Unit::TestCase
     end
   end
 
+  def test_key_get_invalid_meta
+    k = Kdb::Key.new "user/asdf"
+
+    assert_equal "", k["does_not_exist"]
+
+  end
+
 
   def test_key_get_name 
     assert_nothing_raised do
@@ -268,6 +275,53 @@ class KdbKeyTestCases < Test::Unit::TestCase
       assert_true k2.is_direct_below? k1
 
     end
+  end
+
+  def test_Key_cloning
+    k = Kdb::Key.new "user/key1", value: "hello", meta1: "mv"
+    
+    assert_true k.is_valid?
+    assert_equal "user/key1", k.name
+    assert_equal "hello", k.value
+    assert_equal "mv", k["meta1"]
+
+    kc = k.clone
+
+    assert_true k.__id__ != kc.__id__
+    assert_instance_of Kdb::Key, kc
+    assert_equal "user/key1", kc.name
+    assert_equal "hello", kc.value
+    assert_equal "mv", kc["meta1"]
+
+    kc.name = "user/key2"
+    kc.value = "world"
+    kc["meta1"] = "vm"
+    kc["meta2"] = "vm2"
+
+    assert_equal "user/key1", k.name
+    assert_equal "hello", k.value
+    assert_equal "mv", k["meta1"]
+    assert_false k.has_meta? "meta2"
+
+    assert_equal "user/key2", kc.name
+    assert_equal "world", kc.value
+    assert_equal "vm", kc["meta1"]
+    assert_equal "vm2", kc["meta2"]
+
+    kc2 = kc.dup
+
+    assert_true kc.__id__ != kc2.__id__
+    assert_true k.__id__ != kc2.__id__
+
+    kc2.name = "user/key3"
+    kc2.value = "wonderful"
+
+    assert_equal "user/key2", kc.name
+    assert_equal "world", kc.value
+
+    assert_equal "user/key3", kc2.name
+    assert_equal "wonderful", kc2.value
+
   end
 
   def test_meta_iterator
