### 2016-09-05 9754ff2
crypto-plugin: improve loops for encryption and decryption
