### 2016-09-05 9e90061
Markdown Link Converter: add external link checker
rename stuff + fix non shell operation
