### 2016-09-05 d6c3267
Ruby: start with KeySet
implemented:
- basic constructor
- each method
- append
