### 2016-09-05 fa403d5
global plugins: fixed call function
Calls the correct kdb operation on plugins (Get,Set,Error), was
previously only kdbGet.
