### 2016-09-06 36f60a6
Ruby: add KeySet.each tests + Enumerable
add KeySet.each test cases in combinations with Enumerable methods.
Also add spaceship operator for Key, which allows sorting the Ruby way.
