### 2016-09-06 0651202
crypto-plugin: prepare bin-2-string facility
will replace elektraCryptoNormalizeRandomString() method
