+++ src/plugins/crypto/helper.h
@@ -14,6 +14,8 @@
 #include <kdbtypes.h>
 
 kdb_unsigned_long_t elektraCryptoGetIterationCount (KeySet * config);
+kdb_octet_t * elektraCryptoHex2Bin (Key * errorKey, const char * hexBuffer);
+char * elektraCryptoBin2Hex (Key * errorKey, const kdb_octet_t * buffer, const size_t length);
 void elektraCryptoNormalizeRandomString (kdb_octet_t * buffer, const kdb_unsigned_short_t length);
 
 #endif
