+++ src/plugins/crypto/testmod_crypto_botan.c
@@ -18,6 +18,7 @@ int main (int argc, char ** argv)
 
 	init (argc, argv);
 
+	test_hex ();
 	test_gpg ();
 	test_init (PLUGIN_NAME);
 	test_incomplete_config (PLUGIN_NAME);
