### 2016-09-07 5ee18ce
shell_recorder: fix shellcheck warnigns
