### 2016-09-07 d68e1dd
getenv: fixed invalid layer prefix
