### 2016-09-07 1897ab8
filecheck: fixed invalid format string
