+++ src/plugins/filecheck/filecheck.c
@@ -247,7 +247,7 @@ static long checkFile (Key * parentKey, const char * filename, checkStruct * che
 			le_ret = validateLineEnding (line, &(checkConf->validLE), 0);
 			if (le_ret)
 			{
-				ELEKTRA_SET_ERRORF (137, parentKey, "invalid lineending at position %lu", bytesRead + le_ret);
+				ELEKTRA_SET_ERRORF (137, parentKey, "invalid lineending at position %zd", bytesRead + le_ret);
 				retVal = -1;
 				break;
 			}
@@ -257,7 +257,7 @@ static long checkFile (Key * parentKey, const char * filename, checkStruct * che
 			null_ret = checkNull (line, bytesRead);
 			if (null_ret)
 			{
-				ELEKTRA_SET_ERRORF (137, parentKey, "found null-byte at position %lu", bytesRead + null_ret);
+				ELEKTRA_SET_ERRORF (137, parentKey, "found null-byte at position %zd", bytesRead + null_ret);
 				retVal = -1;
 				break;
 			}
@@ -267,7 +267,7 @@ static long checkFile (Key * parentKey, const char * filename, checkStruct * che
 			iconv_ret = validateEncoding (line, conv, bytesRead);
 			if (iconv_ret)
 			{
-				ELEKTRA_SET_ERRORF (137, parentKey, "invalid encoding at position %lu", bytesRead + iconv_ret);
+				ELEKTRA_SET_ERRORF (137, parentKey, "invalid encoding at position %zd", bytesRead + iconv_ret);
 				retVal = -1;
 				break;
 			}
@@ -288,7 +288,7 @@ static long checkFile (Key * parentKey, const char * filename, checkStruct * che
 			unprintable_ret = checkUnprintable (line);
 			if (unprintable_ret)
 			{
-				ELEKTRA_SET_ERRORF (137, parentKey, "unprintable character at position %lu", bytesRead + unprintable_ret);
+				ELEKTRA_SET_ERRORF (137, parentKey, "unprintable character at position %zd", bytesRead + unprintable_ret);
 				retVal = -1;
 				break;
 			}
