### 2016-09-07 db18b1a
Ruby: add additional KeySet constructors
add KeySet constructors which allow passing a initial Key, KeySet or
Array of Keys.
