### 2016-09-07 c6f21a2
Ruby: add KeySet equality operator
