### 2016-09-07 a1e1070
conditionals: fix #651 *BSD regex
