+++ src/plugins/conditionals/conditionals.c
@@ -21,15 +21,16 @@
 #include <stdio.h>
 #include <stdlib.h>
 #include <string.h>
+#include <sys/types.h>
 
 #include "conditionals.h"
 
 #define EPSILON 0.00001
 
 #if defined(__APPLE__)
-#define REGEX_FLAGS_CONDITION (REG_EXTENDED | REG_NEWLINE | REG_ENHANCED | REG_MINIMAL)
+#define REGEX_FLAGS_CONDITION (REG_EXTENDED | REG_ENHANCED | REG_MINIMAL)
 #else
-#define REGEX_FLAGS_CONDITION (REG_EXTENDED | REG_NEWLINE)
+#define REGEX_FLAGS_CONDITION (REG_EXTENDED)
 #endif
 
 typedef enum {
@@ -610,11 +611,12 @@ static CondResult parseCondition (Key * key, const char * condition, const Key *
 static CondResult parseConditionString (const Key * meta, const Key * suffixList, Key * parentKey, Key * key, KeySet * ks, Operation op)
 {
 	const char * conditionString = keyString (meta);
-	const char * regexString = "((\\(((.*)?)\\))[[:space:]]*\\?[[:space:]]*(\\(((.*)?)\\)))($|([[:space:]]*:[[:space:]]*(\\((.*)\\))))";
-
-	regex_t regex;
+	const char * regexString1 = "(\\(((.*)?)\\))[[:space:]]*\\?";
+	const char * regexString2 = "\\?[[:space:]]*(\\(((.*)?)\\))";
+	const char * regexString3 = ":[[:space:]]*(\\(((.*)?)\\))";
+	regex_t regex1, regex2, regex3;
 	CondResult ret;
-	if ((ret = regcomp (&regex, regexString, REGEX_FLAGS_CONDITION)))
+	if ((ret = regcomp (&regex1, regexString1, REGEX_FLAGS_CONDITION)))
 	{
 		ELEKTRA_SET_ERROR (87, parentKey, "Couldn't compile regex: most likely out of memory"); // the regex compiles so the only
 		// possible error would be out of
@@ -622,23 +624,45 @@ static CondResult parseConditionString (const Key * meta, const Key * suffixList
 		ksDel (ks);
 		return ERROR;
 	}
-	int subMatches = 12;
+	if ((ret = regcomp (&regex2, regexString2, REGEX_FLAGS_CONDITION)))
+	{
+		ELEKTRA_SET_ERROR (87, parentKey, "Couldn't compile regex: most likely out of memory"); // the regex compiles so the only
+		// possible error would be out of
+		// memory
+		regfree (&regex1);
+		ksDel (ks);
+		return ERROR;
+	}
+	if ((ret = regcomp (&regex3, regexString3, REGEX_FLAGS_CONDITION)))
+	{
+		ELEKTRA_SET_ERROR (87, parentKey, "Couldn't compile regex: most likely out of memory"); // the regex compiles so the only
+		// possible error would be out of
+		// memory
+		regfree (&regex1);
+		regfree (&regex2);
+		ksDel (ks);
+		return ERROR;
+	}
+	int subMatches = 6;
 	regmatch_t m[subMatches];
-	char * ptr = (char *)conditionString;
-	int nomatch = regexec (&regex, ptr, subMatches, m, 0);
+	int nomatch = regexec (&regex1, conditionString, subMatches, m, 0);
 	if (nomatch)
 	{
 		ELEKTRA_SET_ERRORF (134, parentKey, "Invalid syntax: \"%s\". Check kdb info conditionals for additional information",
 				    conditionString);
-		regfree (&regex);
[…]
