### 2016-09-07 52c357d
mathcheck: fix #653 + many other things
