### 2016-09-07 551137d
conditionals: additional unittest
