+++ src/plugins/boolean/testmod_boolean.c
@@ -1,0 +1,29 @@
+/*
+ * *
+ * @file
+ *
+ * @brief Tests for boolean plugin
+ *
+ * @copyright BSD License (see doc/COPYING or http://www.libelektra.org)
+ *
+ */
+
+#include <stdlib.h>
+#include <string.h>
+
+#include <kdbconfig.h>
+
+#include <tests_plugin.h>
+
+
+int main (int argc, char ** argv)
+{
+	printf ("BOOLEAN     TESTS\n");
+	printf ("==================\n\n");
+
+	init (argc, argv);
+
+	printf ("\ntestmod_boolean RESULTS: %d test(s) done. %d error(s).\n", nbTest, nbError);
+
+	return nbError;
+}
