+++ src/plugins/boolean/boolean.c
@@ -1,0 +1,385 @@
+/**
+ * @file
+ *
+ * @brief Source for boolean plugin
+ *
+ * @copyright BSD License (see doc/COPYING or http://www.libelektra.org)
+ *
+ */
+
+#include "boolean.h"
+
+#include <ctype.h>
+#include <kdberrors.h>
+#include <kdbhelper.h>
+#include <stdio.h>
+#include <stdlib.h>
+#include <string.h>
+
+
+#define DEFAULT_TRUE_VALUE "1"
+#define DEFAULT_FALSE_VALUE "0"
+
+typedef enum {
+	TRUE = 1 << 0,
+	FALSE = 1 << 1,
+	WARNING = 1 << 2,
+} InvalidAction;
+
+typedef struct
+{
+	char * true;
+	char * false;
+	InvalidAction invalid;
+	char ** trueValues;
+	char ** falseValues;
+} BoolData;
+
+int elektraBooleanClose (Plugin * handle ELEKTRA_UNUSED, Key * errorKey ELEKTRA_UNUSED)
+{
+	// free all plugin resources and shut it down
+	// this function is optional
+	BoolData * data = elektraPluginGetData (handle);
+	if (!data) return 1;
+	if (data->trueValues)
+	{
+		char ** ptr = (data->trueValues);
+		while (*ptr)
+		{
+			elektraFree (*ptr);
+			++(ptr);
+		}
+		elektraFree (data->trueValues);
+	}
+	if (data->falseValues)
+	{
+		char ** ptr = (data->falseValues);
+		while (*ptr)
+		{
+			elektraFree (*ptr);
+			++(ptr);
+		}
+		elektraFree (data->falseValues);
+	}
+	elektraFree (data);
+	elektraPluginSetData (handle, NULL);
+	return 1; // success
+}
+
+static int stricmp (const char * s1, const char * s2)
+{
+
+	if (!(*s1) && !(*s2))
+	{
+		return 0;
+	}
+	else if (!(*s1))
+	{
+		return (-1);
+	}
+	else if (!(*s2))
+	{
+		return 1;
+	}
+	else
+	{
+		char * p1 = (char *)s1;
+		char * p2 = (char *)s2;
+		while (1)
+		{
+			if (!(*p1) && !(*p2))
+			{
+				return 0;
+			}
+			else if (!(*p1))
+			{
+				return (-1);
+			}
+			else if (!(*p2))
+			{
+				return 1;
+			}
+			else
+			{
+				if (toupper (*p1) == toupper (*p2))
+				{
+					++p1;
+					++p2;
+				}
+				else
+				{
+					return (toupper (*p1) - toupper (*p2));
+				}
+			}
+		}
+	}
+}
+
+static int isTrue (const char * value, const char ** trueValues)
+{
+	char ** ptr = (char **)trueValues;
+	int retVal = 0;
+	while (*ptr)
+	{
+		if (!stricmp (value, *ptr))
+		{
+			retVal = 1;
+			break;
+		}
+		++ptr;
+	}
+	return retVal;
+}
+
+static int isFalse (const char * value, const char ** falseValues)
+{
+	char ** ptr = (char **)falseValues;
+	int retVal = 0;
+	while (*ptr)
+	{
+		if (!stricmp (value, *ptr))
+		{
+			retVal = 1;
+			break;
+		}
+		++ptr;
+	}
+	return retVal;
+}
+
+static void normalize (Key * key, Key * parentKey, BoolData * data)
+{
+	const char * defaultTrueValues[] = {
+		"TRUE", "1", "ON", "ENABLE", "ENABLED", "YES", NULL,
+	};
+	const char * defaultFalseValues[] = {
[…]
