+++ src/plugins/README.md
@@ -152,6 +152,7 @@ Transformations:
 
 - [keytometa](keytometa/) transforms keys to metadata
 - [rename](rename/) renames keys according to different rules
+- [boolean](boolean/) canonicalizes boolean keys 
 
 Doing other stuff:
 
