### 2016-09-07 91aabbc
error2warning kdbSet
