+++ src/libs/elektra/kdb.c
@@ -1003,7 +1003,32 @@ static void elektraSetRollback (Split * split, Key * parentKey)
 		}
 	}
 }
-
+static int copyError (Key * dest, Key * src)
+{
+	keyRewindMeta (src);
+	Key * metaKey = (Key *)keyGetMeta (src, "error");
+	if (!metaKey) return 0;
+	keySetMeta (dest, keyName (metaKey), keyString (metaKey));
+	while ((metaKey = (Key *)keyNextMeta (src)) != NULL)
+	{
+		if (strncmp (keyName (metaKey), "error/", 6)) break;
+		keySetMeta (dest, keyName (metaKey), keyString (metaKey));
+	}
+	return 1;
+}
+static void clearError (Key * key)
+{
+	keySetMeta (key, "error", 0);
+	keySetMeta (key, "error/number", 0);
+	keySetMeta (key, "error/description", 0);
+	keySetMeta (key, "error/reason", 0);
+	keySetMeta (key, "error/ingroup", 0);
+	keySetMeta (key, "error/module", 0);
+	keySetMeta (key, "error/file", 0);
+	keySetMeta (key, "error/line", 0);
+	keySetMeta (key, "error/configfile", 0);
+	keySetMeta (key, "error/mountpoint", 0);
+}
 
 /** @brief Set keys in an atomic and universal way.
  *
@@ -1082,10 +1107,14 @@ int kdbSet (KDB * handle, KeySet * ks, Key * parentKey)
 	{
 		return -1;
 	}
+	Key * oldError = keyNew (keyName (parentKey), KEY_END);
+	int hasError = copyError (oldError, parentKey);
 
 	if (ns == KEY_NS_META)
 	{
+		clearError (parentKey); // clear previous error to set new one
 		ELEKTRA_SET_ERRORF (104, parentKey, "metakey with name \"%s\" passed to kdbSet", keyName (parentKey));
+		keyDel (oldError);
 		return -1;
 	}
 
@@ -1096,7 +1125,9 @@ int kdbSet (KDB * handle, KeySet * ks, Key * parentKey)
 
 	if (!handle || !ks)
 	{
+		clearError (parentKey); // clear previous error to set new one
 		ELEKTRA_SET_ERROR (37, parentKey, "handle or ks null pointer");
+		keyDel (oldError);
 		return -1;
 	}
 
@@ -1110,6 +1141,7 @@ int kdbSet (KDB * handle, KeySet * ks, Key * parentKey)
 
 	if (splitBuildup (split, handle, parentKey) == -1)
 	{
+		clearError (parentKey); // clear previous error to set new one
 		ELEKTRA_SET_ERROR (38, parentKey, "error in splitBuildup");
 		goto error;
 	}
@@ -1118,6 +1150,7 @@ int kdbSet (KDB * handle, KeySet * ks, Key * parentKey)
 	int syncstate = splitDivide (split, handle, ks);
 	if (syncstate == -1)
 	{
+		clearError (parentKey); // clear previous error to set new one
 		ELEKTRA_SET_ERROR (8, parentKey, keyName (ksCurrent (ks)));
 		goto error;
 	}
@@ -1130,6 +1163,7 @@ int kdbSet (KDB * handle, KeySet * ks, Key * parentKey)
 	{
 		/* No update is needed */
 		keySetName (parentKey, keyName (initialParent));
+		if (syncstate < 0) clearError (parentKey); // clear previous error to set new one
 		if (syncstate == -1)
 		{
 			ELEKTRA_SET_ERROR (107, parentKey, "Assert failed: invalid namespace");
[…]
