### 2016-09-07 98c557f
shell_recorder: fixed leftover shellcheck warnings
