### 2016-09-08 e964daf
global plugins: simplify call
