### 2016-09-08 47ba065
crypto-plugin: replace elektraCryptoNormalizeRandomString() with hex encoding
elektraCryptoNormalizeRandomString() was reducing the entropy of the master password and the salt used by the PBKDF2 function.
Now the master password and the salt are stored in hex-encoded form.
