+++ src/plugins/crypto/openssl_operations.c
@@ -46,6 +46,7 @@ static int getKeyIvForEncryption (KeySet * config, Key * errorKey, Key * k, Key 
 {
 	kdb_octet_t salt[ELEKTRA_CRYPTO_DEFAULT_SALT_LEN];
 	kdb_octet_t keyBuffer[KEY_BUFFER_SIZE];
+	char * saltHexString = NULL;
 
 	// generate the salt
 	pthread_mutex_lock (&mutex_ssl);
@@ -57,24 +58,17 @@ static int getKeyIvForEncryption (KeySet * config, Key * errorKey, Key * k, Key 
 		return -1;
 	}
 	pthread_mutex_unlock (&mutex_ssl);
-	elektraCryptoNormalizeRandomString (salt, sizeof (salt));
-	keySetMeta (k, ELEKTRA_CRYPTO_META_SALT, (char *)salt);
+	saltHexString = elektraCryptoBin2Hex (errorKey, salt, sizeof (salt));
+	if (!saltHexString) return -1; // error set by elektraCryptoBin2Hex()
+	keySetMeta (k, ELEKTRA_CRYPTO_META_SALT, saltHexString);
+	elektraFree (saltHexString);
 
 	// read iteration count
 	const kdb_unsigned_long_t iterations = elektraCryptoGetIterationCount (config);
 
 	// receive master password from the configuration
-	Key * master = ksLookupByName (config, ELEKTRA_CRYPTO_PARAM_MASTER_PASSWORD, 0);
-	if (!master)
-	{
-		ELEKTRA_SET_ERRORF (ELEKTRA_ERROR_CRYPTO_CONFIG_FAULT, errorKey, "missing %s", ELEKTRA_CRYPTO_PARAM_MASTER_PASSWORD);
-		return -1;
-	}
-	Key * msg = keyDup (master);
-	if (elektraCryptoGpgDecryptMasterPassword (config, errorKey, msg) != 1)
-	{
-		goto error;
-	}
+	Key * msg = elektraCryptoGetMasterPassword (errorKey, config);
+	if (!msg) goto error; // error set by elektraCryptoGetMasterPassword()
 
 	// generate/derive the cryptographic key and the IV
 	pthread_mutex_lock (&mutex_ssl);
@@ -93,7 +87,7 @@ static int getKeyIvForEncryption (KeySet * config, Key * errorKey, Key * k, Key 
 	return 1;
 
 error:
-	keyDel (msg);
+	if (msg) keyDel (msg);
 	return -1;
 }
 
@@ -110,6 +104,8 @@ error:
 static int getKeyIvForDecryption (KeySet * config, Key * errorKey, Key * k, Key * cKey, Key * cIv)
 {
 	kdb_octet_t keyBuffer[KEY_BUFFER_SIZE];
+	kdb_octet_t * saltBuffer = NULL;
+	size_t saltBufferLen = 0;
 
 	// get the salt
 	const Key * salt = keyGetMeta (k, ELEKTRA_CRYPTO_META_SALT);
@@ -119,27 +115,20 @@ static int getKeyIvForDecryption (KeySet * config, Key * errorKey, Key * k, Key 
 				    ELEKTRA_CRYPTO_META_SALT, keyName (k));
 		return -1;
 	}
+	elektraCryptoHex2Bin (errorKey, keyString (salt), &saltBuffer, &saltBufferLen);
+	if (!saltBuffer) return -1; // error set by elektraCryptoHex2Bin()
 
 	// get the iteration count
 	const kdb_unsigned_long_t iterations = elektraCryptoGetIterationCount (config);
 
 	// receive master password from the configuration
[…]
