+++ src/plugins/crypto/botan_operations.cpp
@@ -41,6 +41,7 @@ extern "C" {
 static int getKeyIvForEncryption (KeySet * config, Key * errorKey, Key * k, SymmetricKey ** cKey, InitializationVector ** cIv)
 {
 	byte salt[ELEKTRA_CRYPTO_DEFAULT_SALT_LEN];
+	char * saltHexString = NULL;
 	const size_t requiredKeyBytes = ELEKTRA_CRYPTO_BOTAN_KEYSIZE + ELEKTRA_CRYPTO_BOTAN_BLOCKSIZE;
 	Key * msg = NULL;
 
@@ -48,31 +49,24 @@ static int getKeyIvForEncryption (KeySet * config, Key * errorKey, Key * k, Symm
 	{
 		// generate the salt
 		AutoSeeded_RNG rng;
-		rng.randomize (salt, ELEKTRA_CRYPTO_DEFAULT_SALT_LEN - 1);
-		elektraCryptoNormalizeRandomString (salt, sizeof (salt));
-		keySetMeta (k, ELEKTRA_CRYPTO_META_SALT, reinterpret_cast<char *> (salt));
+		rng.randomize (salt, sizeof (salt));
+		saltHexString = elektraCryptoBin2Hex (errorKey, salt, sizeof (salt));
+		if (!saltHexString) return -1; // error set by elektraCryptoBin2Hex()
+		keySetMeta (k, ELEKTRA_CRYPTO_META_SALT, saltHexString);
+		elektraFree (saltHexString);
 
 		// read iteration count
 		const kdb_unsigned_long_t iterations = elektraCryptoGetIterationCount (config);
 
 		// receive master password from the configuration
-		Key * master = ksLookupByName (config, ELEKTRA_CRYPTO_PARAM_MASTER_PASSWORD, 0);
-		if (!master)
-		{
-			ELEKTRA_SET_ERRORF (ELEKTRA_ERROR_CRYPTO_CONFIG_FAULT, errorKey, "missing %s",
-					    ELEKTRA_CRYPTO_PARAM_MASTER_PASSWORD);
-			return -1;
-		}
-		msg = keyDup (master);
-		if (elektraCryptoGpgDecryptMasterPassword (config, errorKey, msg) != 1)
-		{
-			keyDel (msg);
-			return -1;
-		}
+		msg = elektraCryptoGetMasterPassword (errorKey, config);
+		if (!msg) return -1; // error set by elektraCryptoGetMasterPassword()
 
 		// generate/derive the cryptographic key and the IV
 		PBKDF * pbkdf = get_pbkdf ("PBKDF2(SHA-256)");
-		OctetString derived = pbkdf->derive_key (requiredKeyBytes, std::string (keyString (msg)), salt, sizeof (salt), iterations);
+		OctetString derived = pbkdf->derive_key (
+			requiredKeyBytes, std::string (reinterpret_cast<const char *> (keyValue (msg)), keyGetValueSize (msg)), salt,
+			sizeof (salt), iterations);
 
 		*cKey = new SymmetricKey (derived.begin (), ELEKTRA_CRYPTO_BOTAN_KEYSIZE);
 		*cIv = new InitializationVector (derived.begin () + ELEKTRA_CRYPTO_BOTAN_KEYSIZE, ELEKTRA_CRYPTO_BOTAN_BLOCKSIZE);
@@ -102,6 +96,8 @@ static int getKeyIvForEncryption (KeySet * config, Key * errorKey, Key * k, Symm
 static int getKeyIvForDecryption (KeySet * config, Key * errorKey, Key * k, SymmetricKey ** cKey, InitializationVector ** cIv)
[…]
