+++ src/plugins/crypto/helper.h
@@ -13,9 +13,9 @@
 #include <kdb.h>
 #include <kdbtypes.h>
 
+Key * elektraCryptoGetMasterPassword (Key * errorKey, KeySet * config);
 kdb_unsigned_long_t elektraCryptoGetIterationCount (KeySet * config);
-kdb_octet_t * elektraCryptoHex2Bin (Key * errorKey, const char * hexBuffer);
+void elektraCryptoHex2Bin (Key * errorKey, const char * hexBuffer, kdb_octet_t ** output, size_t * outputLen);
 char * elektraCryptoBin2Hex (Key * errorKey, const kdb_octet_t * buffer, const size_t length);
-void elektraCryptoNormalizeRandomString (kdb_octet_t * buffer, const kdb_unsigned_short_t length);
 
 #endif

[…]
