### 2016-09-08 fc0c624
crypto-plugin: encode salt into crypto payload
Metakey crypto/salt is removed during encryption and the salt is encoded into the crypto payload.
