+++ src/plugins/crypto/helper.h
@@ -13,9 +13,11 @@
 #include <kdb.h>
 #include <kdbtypes.h>
 
+int elektraCryptoGetSaltFromMetaKey (Key * errorKey, Key * k, kdb_octet_t ** salt, kdb_unsigned_long_t * saltLen);
+int elektraCryptoGetSaltFromCryptoPayload (Key * errorKey, Key * k, kdb_octet_t ** salt, kdb_unsigned_long_t * saltLen);
 Key * elektraCryptoGetMasterPassword (Key * errorKey, KeySet * config);
 kdb_unsigned_long_t elektraCryptoGetIterationCount (KeySet * config);
-void elektraCryptoHex2Bin (Key * errorKey, const char * hexBuffer, kdb_octet_t ** output, size_t * outputLen);
+void elektraCryptoHex2Bin (Key * errorKey, const char * hexBuffer, kdb_octet_t ** output, kdb_unsigned_long_t * outputLen);
 char * elektraCryptoBin2Hex (Key * errorKey, const kdb_octet_t * buffer, const size_t length);
 
 #endif
