### 2016-09-09 f91bc9f
crypto-plugin: replace MIN macro in test_internals.h
Now implemented as inline function
