### 2016-09-09 dfc3239
crypto-plugin: improve error and warning messages
