### 2016-09-09 41d5cac
crypto-plugin: refactor test cases
