### 2016-09-09 f80eb18
crypto-plugin: fix compile error in test_internals.h
