### 2016-09-09 42e35f4
few more additions to keyset
