+++ doc/tutorials/python-kdb.md
@@ -6,7 +6,7 @@ When programming in python it is possible to access the kdb database, changing v
 
 ## First Steps
 
-In order to being able to use kdb, obviously, you at first need to `import kdb`. You need access to an python-object of kdb. This is accomlished by calling `kdb.KDB()` and saving this to a variable because later on this object will be needed for various operations. 
+In order to being able to use kdb, obviously, you at first need to `import kdb`. You need access to an python-object of kdb. This is accomplished by calling `kdb.KDB()` and saving this to a variable because later on this object will be needed for various operations. 
 The easiest way to do this would be:
 ```
 import kdb
@@ -17,7 +17,7 @@ with kdb.KDB() as k:
 
 ## Keyset
 
-A keyset is basically a list of the keys that lie within the specified range. At first it is necessary to create a new keyset. When simply calling `kdb.KeySet()` the keyset is of size 0. There is no restriction to the keyest's size. It is possible to specify a certain (maximum) size for a keyset. To load keys into to keyset from the database you simply call the method `get` provided by the kdb-object.
+A keyset is basically a list of the keys that lie within the specified range of the databse. When creating an empty keyset this range is obviously zero. It is possible to load the whole database into a keyset but in a lot of cases this is not needed and you can specify which keys exactly you need (which I mean with specified range). At first it is necessary to create a new keyset. When simply calling `kdb.KeySet()` the keyset is of size 0. There is no restriction to the keyset's size. It is possible to specify a certain (maximum) size for a keyset. To load keys into to keyset from the database you simply call the method `get` provided by the kdb-object.
 
 ```
 import kdb
@@ -26,10 +26,10 @@ with kdb.KDB() as k:
 	ks = kdb.KeySet()
 	print len(ks) # should be 0
 	# load an existing keyset
-	k.get(ks, 'path/to/keys')
+	k.get(ks, '/path/to/keys')
 ```
 
-It is also possible to iterate completely normally over a keyset and use len and reversed, dup and deepcopy. The elements of a keyset can be accessed by indexes and a keyset can be sliced. Another way of accessing a key is by the key-name (`keyset_name['path/to/keys/key_name']`). If the key-name does not exist within the keyset, a KeyError exception is thrown.
+It is also possible to iterate as expected over a keyset and use len, reversed and copying. The elements of a keyset can be accessed by indexes and a keyset can be sliced. Another way of accessing a key is by the key-name (`keyset_name['/path/to/keys/key_name']`). If the key-name does not exist within the keyset, a KeyError exception is thrown.
 
 An example that shows how to load an existing keyset and then access every key and value of the loaded keyset:
 
@@ -38,10 +38,15 @@ import kdb
 with kdb.KDB() as k:
 	ks = kdb.KeySet()
 	#loading an existing keyset
-	k.get(ks, 'path/to/keys')
-	for i in ks:
+	k.get(ks, '/path/to/keys')
+	# if you for some reason want to loop through the keyset last key first use: for i in reversed(ks):
+	for i in ks: 
 		# print for every key in the keyset the key and the value
-		print 'key: ' + i + 'value: ' + ks[i]
+		print 'key: '
+		print i
+		print 'value: '
+		print ks[i]
+	
 ```
 
 Here an example of how you can easily check if a key exists:
@@ -50,11 +55,45 @@ Here an example of how you can easily check if a key exists:
 import kdb
[…]
