### 2016-09-11 29d1752
Update elektra-backends.md
fix typo in elektra-backends.md
