+++ doc/help/elektra-backends.md
@@ -12,7 +12,7 @@ Since Elektra 0.8 a backend is composed of many plugins.
 
 ## MULTIPLE PLUGINS
 
-Its clear that too many features in one backend is
+It's clear that too many features in one backend is
 problematic.  It introduces unwanted external dependences and leads to
 less portable backends.  Many different aspects clutter the code making
 the backends unmaintainable.  Features of other backends cannot be taken
