### 2016-09-11 d3642ed
fcrypt: first prototype
Working file encryption. Decryption implemented but failing.
