+++ src/plugins/fcrypt/fcrypt.c
@@ -1,0 +1,220 @@
+/**
+ * @file
+ *
+ * @brief filter plugin providing cryptographic operations
+ *
+ * @copyright BSD License (see doc/COPYING or http://www.libelektra.org)
+ *
+ */
+
+#ifndef HAVE_KDBCONFIG
+#include "kdbconfig.h"
+#endif
+#include "../crypto/gpg.h"
+#include "fcrypt.h"
+#include <kdb.h>
+#include <kdberrors.h>
+#include <kdbtypes.h>
+#include <stdlib.h>
+#include <string.h>
+
+/**
+ * @brief Allocates a new string holding the name of the encrypted file.
+ * @param file holds the path to the original file
+ * @returns an allocated string holding the name of the encrypted file. Must be freed by the caller.
+ */
+static char * getEncryptedFileName (const char * file)
+{
+	const size_t newFileAllocated = strlen (file) + 5;
+	char * newFile = elektraMalloc (newFileAllocated);
+	if (!newFile) return NULL;
+	snprintf (newFile, newFileAllocated, "%s.gpg", file);
+	return newFile;
+}
+
+/**
+ * @brief Read number of total GPG recipient keys from the plugin configuration.
+ * @param config holds the plugin configuration
+ * @returns the number of GPG recipient keys.
+ */
+static size_t getRecipientCount (KeySet * config)
+{
+	Key * k;
+	size_t recipientCount = 0;
+	Key * root = ksLookupByName (config, ELEKTRA_CRYPTO_PARAM_GPG_KEY, 0);
+
+	if (!root) return 0;
+
+	// toplevel
+	if (strlen (keyString (root)) > 0)
+	{
+		recipientCount++;
+	}
+
+	ksRewind (config);
+	while ((k = ksNext (config)) != 0)
+	{
+		if (keyIsBelow (k, root))
+		{
+			recipientCount++;
+		}
+	}
+	return recipientCount;
+}
+
+/**
+ * @brief establish the Elektra plugin contract and decrypt the file provded at parentKey using GPG.
+ * @retval 1 on success
+ * @retval -1 on failure
+ */
+int ELEKTRA_PLUGIN_FUNCTION (ELEKTRA_PLUGIN_NAME, get) (Plugin * handle, KeySet * ks ELEKTRA_UNUSED, Key * parentKey)
+{
+	// Publish module configuration to Elektra (establish the contract)
+	if (!strcmp (keyName (parentKey), "system/elektra/modules/" ELEKTRA_PLUGIN_NAME))
+	{
+		KeySet * moduleConfig = ksNew (30,
+#include "contract.h"
+					       KS_END);
+		ksAppend (ks, moduleConfig);
+		ksDel (moduleConfig);
+		return 1;
+	}
+
+	// regular decryption business
+	KeySet * pluginConfig = elektraPluginGetConfig (handle);
+
+	// decrypt the file with gpg - prepare argument vector for gpg call
+	static const int argc = 8;
+	char * argv[8] = { NULL, "--batch", "--yes", "-o", NULL, "-d", NULL, NULL };
+	char * encryptedFile = getEncryptedFileName (keyString (parentKey));
+
+	if (!encryptedFile)
+	{
+		ELEKTRA_SET_ERROR (87, parentKey, "Memory allocation failed");
+		return -1;
+	}
+
+	// safely discarding const from keyString() return value
+	argv[4] = (char *)keyString (parentKey);
+	argv[6] = encryptedFile;
+
[…]
