+++ src/plugins/fcrypt/CMakeLists.txt
@@ -1,0 +1,17 @@
+include (LibAddPlugin)
+
+add_plugin (fcrypt
+	SOURCES
+		../crypto/helper.h
+		../crypto/helper.c
+		../crypto/gpg.h
+		../crypto/gpg.c
+		fcrypt.h
+		fcrypt.c
+	COMPILE_DEFINITIONS
+		ELEKTRA_PLUGIN_NAME=\"fcrypt\"
+)
+
+if (ADDTESTING_PHASE)
+	add_plugintest (fcrypt)
+endif ()
