### 2016-09-12 3be1cd7
Ruby: return Nil instead a Null-Key
For all methods that return a Key, check if key is null and
return Nil instead of a Null-key (key.is_null? == true)
