### 2016-09-12 43f0469
Ruby: add KeySet.pop test cases
