### 2016-09-12 a0c80df
Ruby: impl dup, clone, cut
