### 2016-09-12 aa665fd
Ruby: add common helper methods to KeySet
