### 2016-09-12 74c78ce
build improvement: use object lib for kdb tool
