### 2016-09-12 8393814
build improvement: object library for kdb tool
