### 2016-09-12 5abb149
Ruby: enhance Key meta data iterator
