### 2016-09-12 1369e86
fcrypt: progress on file encryption prototype
See #914
