+++ src/plugins/fcrypt/fcrypt.c
@@ -18,17 +18,19 @@
 #include <stdlib.h>
 #include <string.h>
 
+
 /**
- * @brief Allocates a new string holding the name of the encrypted file.
+ * @brief Allocates a new string holding the name of the temporary file.
  * @param file holds the path to the original file
  * @returns an allocated string holding the name of the encrypted file. Must be freed by the caller.
  */
-static char * getEncryptedFileName (const char * file)
+static char * getTemporaryFileName (const char * file)
 {
-	const size_t newFileAllocated = strlen (file) + 5;
+	const size_t newFileAllocated = strlen (file) + 7;
 	char * newFile = elektraMalloc (newFileAllocated);
 	if (!newFile) return NULL;
-	snprintf (newFile, newFileAllocated, "%s.gpg", file);
+	snprintf (newFile, newFileAllocated, "%sXXXXXX", file);
+	mktemp (newFile);
 	return newFile;
 }
 
@@ -80,27 +82,43 @@ int ELEKTRA_PLUGIN_FUNCTION (ELEKTRA_PLUGIN_NAME, get) (Plugin * handle, KeySet 
 		return 1;
 	}
 
+	// TODO encrypt if this is a `postgetstorage` call
+
 	// regular decryption business
 	KeySet * pluginConfig = elektraPluginGetConfig (handle);
 
 	// decrypt the file with gpg - prepare argument vector for gpg call
 	static const int argc = 8;
 	char * argv[8] = { NULL, "--batch", "--yes", "-o", NULL, "-d", NULL, NULL };
-	char * encryptedFile = getEncryptedFileName (keyString (parentKey));
-
-	if (!encryptedFile)
+	char * tmpFile = getTemporaryFileName (keyString (parentKey));
+	if (!tmpFile)
 	{
 		ELEKTRA_SET_ERROR (87, parentKey, "Memory allocation failed");
 		return -1;
 	}
 
+	argv[4] = tmpFile;
 	// safely discarding const from keyString() return value
-	argv[4] = (char *)keyString (parentKey);
-	argv[6] = encryptedFile;
+	argv[6] = (char *)keyString (parentKey);
+
+	// NOTE the decryption process works like this:
+	// gpg2 --batch --yes -o tmpfile -d configFile
+	// mv tmpfile configFile
 
 	int result = elektraCryptoGpgCall (pluginConfig, parentKey, NULL, argv, argc);
 
-	elektraFree (encryptedFile);
+	if (result == 1)
+	{
+		// encryption successful, overwrite the original file with the encrypted data
+		if (rename (tmpFile, keyString (parentKey)) != 0)
+		{
+			ELEKTRA_SET_ERRORF (ELEKTRA_ERROR_FCRYPT_RENAME, parentKey, "Renaming file %s to %s failed.", tmpFile,
+					    keyString (parentKey));
+			result = -1;
+		}
+	}
+
+	elektraFree (tmpFile);
 	return result;
 }
 
@@ -112,19 +130,19 @@ int ELEKTRA_PLUGIN_FUNCTION (ELEKTRA_PLUGIN_NAME, get) (Plugin * handle, KeySet 
 int ELEKTRA_PLUGIN_FUNCTION (ELEKTRA_PLUGIN_NAME, set) (Plugin * handle, KeySet * ks ELEKTRA_UNUSED, Key * parentKey)
 {
 	KeySet * pluginConfig = elektraPluginGetConfig (handle);
[…]
