### 2016-09-14 5282d9b
global plugins: elektraGlobalGet
Add the elektraGlobal{Get,Set,Error} functions.
This seemed more practical than using macros everywhere.
