### 2016-09-14 f86546c
doc: add text about labels
