### 2016-09-14 ed1b451
REST service: use cases
