+++ doc/rest_api/snippet_sharing/use_cases/UC_create_snippet.txt
@@ -1,0 +1,17 @@
+Summary:
+----------------------------
+Title: create configuration snippet
+Scope: Entry Management
+Level: User Goal
+Actors: Authenticated User
+Brief: User can create a new database entry based on a set of specified inputs, one of them being a configuration snippet.
+
+Scenarios:
+----------------------------
+Precondition: Authenticated
+Main success scenario: All inputs validate and the configuration snippet could be stored in the Elektra database successfully. The user retrieves a success response.
+Alternative scenario: Wrong inputs lead to an error message and ask the user to correct them, then he can try again. It is also possible that the supplied configuration snippet is of an unsupported format. In either case the user retrieves a response containing error information.
+Error scenario: Technical problems prevent the storage process from completing. The user is informed through an error response.
+Postcondition: Configuration snippet is stored in the database successfully and can be found through search.
+Non-functional Constraints: 
+	- Essential functionality of the service
