### 2016-09-14 0aacbd4
adding an install script for rc files (Fixes #844)
