### 2016-09-14 43e1d58
Ruby: add Kdb::KDB class with tests
