### 2016-09-14 6572cc9
install script as well
