### 2016-09-14 40e394d
Ruby: ignore certain compiler warnings
