### 2016-09-15 68bb9d1
REST service: use-cases updated
