+++ doc/rest_api/snippet_sharing/use_cases/UC_create_snippet.txt
@@ -10,7 +10,7 @@ Scenarios:
 ----------------------------
 Precondition: Authenticated
 Main success scenario: All inputs validate and the configuration snippet could be stored in the Elektra database successfully. The user retrieves a success response.
-Alternative scenario: Wrong inputs lead to an error message and ask the user to correct them, then he can try again. It is also possible that the supplied configuration snippet is of an unsupported format. In either case the user retrieves a response containing error information.
+Alternative scenario: Wrong inputs lead to an error message and ask the user to correct them, then they can try again. It is also possible that the supplied configuration snippet is of an unsupported format. In either case the user retrieves a response containing error information.
 Error scenario: Technical problems prevent the storage process from completing. The user is informed through an error response.
 Postcondition: Configuration snippet is stored in the database successfully and can be found through search.
 Non-functional Constraints: 
