### 2016-09-15 1e33d3a
conditionals: fix regex flags
