### 2016-09-15 e9ae5ab
boolean: unittests
