### 2016-09-15 ca99302
boolean: fixed default value + README.md
