### 2016-09-15 3b466c6
crypto-plugin: add /usr/bin/gpg as second fallback
If the PATH environment variable is empty we check if /usr/bin/gpg2 or /usr/bin/gpg is available as a fallback.

Fixes #940.
