### 2016-09-15 4bd39c6
mathcheck: added unit tests
