### 2016-09-15 fe677ce
getenv: updated README.md
