+++ src/libs/getenv/README.md
@@ -14,14 +14,14 @@ elektrify-getenv(1) -- elektrify the environment of applications
 
     kdb elektrify-getenv curl --elektra-version
     kdb elektrify-getenv curl http://www.libelektra.org
-    kdb set system/env/override/HTTP_PROXY http://proxy.hogege.com:8000/
+    kdb set system/elektra/intercept/getenv/override/HTTP_PROXY http://proxy.hogege.com:8000/
     kdb elektrify-getenv curl http://www.libelektra.org
 
 By using `elektrify-getenv` the last curl invocation will use a different http proxy.
 Or you can also reload while the application is running:
 
     ELEKTRA_RELOAD_TIMEOUT=100 kdb elektrify-getenv firefox
-    kdb set system/env/override/http_proxy http://www.example.com
+    kdb set system/elektra/intercept/getenv/override/http_proxy http://www.example.com
 
 
 ## DESCRIPTION
@@ -56,17 +56,17 @@ To do so, getenv(3) will lookup multiple sources next to searching in the enviro
 1. Given commandline parameters will always be preferred (see [OPTIONS](#OPTIONS) below).
 
    E.g. `elektrify-getenv <app> --elektra:HOME=/path/to/home`
-2. Then `/env/override/<key>` will be looked up, where <key> is the parameter to `getenv`.
+2. Then `/elektra/intercept/getenv/override/<key>` will be looked up, where <key> is the parameter to `getenv`.
    If found, the key will be returned, if it is a null keys, `getenv` will return `NULL`.
 
-   E.g. `kdb set user/env/override/HOME /path/to/home`
+   E.g. `kdb set user/elektra/intercept/getenv/override/HOME /path/to/home`
 3. Then environment will be requested.
 
    E.g. `HOME=/path/to/home elektrify-getenv <application>`
-4. Then `/env/fallback/<key>` will be looked up.
+4. Then `/elektra/intercept/getenv/fallback/<key>` will be looked up.
    If found, the key will be returned, if it is a null keys, `getenv` will return `NULL`.
 
-   E.g. `kdb set user/env/fallback/HOME /path/to/home`
+   E.g. `kdb set user/elektra/intercept/getenv/fallback/HOME /path/to/home`
 
 
 
@@ -87,16 +87,16 @@ the application will be called with `<application> -V -L`.
    Outputs this help.
  * `--elektra-version`:
    Gives version information.
- * `--elektra-debug=file`, `ELEKTRA_DEBUG` or `/env/option/debug`:
+ * `--elektra-debug=file`, `ELEKTRA_DEBUG` or `/elektra/intercept/getenv/option/debug`:
    Trace all getenv(3) calls to a file.
-   stderr if no file is given, e.g. `kdb set user/env/option/debug ""`.
+   stderr if no file is given, e.g. `kdb set user/elektra/intercept/getenv/option/debug ""`.
    Note that null values (no forth argument), will disable debug messages.
    See examples below.
- * `--elektra-clearenv`, `ELEKTRA_CLEARENV` or `/env/option/clearenv`:
+ * `--elektra-clearenv`, `ELEKTRA_CLEARENV` or `/elektra/intercept/getenv/option/clearenv`:
    Call clearenv(3) before entering main.
    This is a recommended security feature.
    Elektra itself, if configured that way, will still be able to use the environment.
- * `--elektra-reload-timeout=time_in_ms`, `ELEKTRA_RELOAD_TIMEOUT` or `/env/option/reload_timeout`:
+ * `--elektra-reload-timeout=time_in_ms`, `ELEKTRA_RELOAD_TIMEOUT` or `/elektra/intercept/getenv/option/reload_timeout`:
    Activate a timeout based feature when a time is given in ms (and is not 0).
 
 Internal Options are available in three different variants:
[…]
