### 2016-09-15 8b0e99e
crypto-plugin: replace strdup call with elektraStrDup
