### 2016-09-15 b8b9bf5
crypto-plugin: improve code structure in getGpgBinary()
