### 2016-09-15 a3f7118
crypto-plugin: ensure return code consistency in isExecutable()
Replaced return value 0 with 1 indicating success to be consistent with the rest of the module.
