### 2016-09-15 63bbc4e
test_getenv: fixed testcase
