### 2016-09-15 1183c53
fcrypt: set include path to crypto plugin dir
