### 2016-09-15 7aa5147
fcrypt: progress on README.md
