### 2016-09-15 541a4c5
fcrypt: track plugin state for kdb get calls
Enables distinction between pregetstorage and postgetstorage.
