+++ src/plugins/fcrypt/fcrypt.c
@@ -18,6 +18,17 @@
 #include <stdlib.h>
 #include <string.h>
 
+enum fcryptGetState
+{
+	PREGETSTORAGE = 0,
+	POSTGETSTORAGE = 1
+};
+
+struct _fcryptState
+{
+	enum fcryptGetState getState;
+};
+typedef struct _fcryptState fcryptState;
 
 /**
  * @brief Allocates a new string holding the name of the temporary file.
@@ -30,7 +41,7 @@ static char * getTemporaryFileName (const char * file)
 	char * newFile = elektraMalloc (newFileAllocated);
 	if (!newFile) return NULL;
 	snprintf (newFile, newFileAllocated, "%sXXXXXX", file);
-	mktemp (newFile);
+	mkstemp (newFile);
 	return newFile;
 }
 
@@ -65,72 +76,14 @@ static size_t getRecipientCount (KeySet * config)
 }
 
 /**
- * @brief establish the Elektra plugin contract and decrypt the file provded at parentKey using GPG.
+ * @brief encrypt the file speicied at parentKey
+ * @param pluginConfig holds the plugin configuration
+ * @pararm parentKey holds the path to the file to be encrpted. Will hold an error description in case of failure.
  * @retval 1 on success
- * @retval -1 on failure
+ * @retval -1 on error, errorKey holds an error description
  */
-int ELEKTRA_PLUGIN_FUNCTION (ELEKTRA_PLUGIN_NAME, get) (Plugin * handle, KeySet * ks ELEKTRA_UNUSED, Key * parentKey)
+static int encrypt (KeySet * pluginConfig, Key * parentKey)
 {
-	// Publish module configuration to Elektra (establish the contract)
-	if (!strcmp (keyName (parentKey), "system/elektra/modules/" ELEKTRA_PLUGIN_NAME))
-	{
-		KeySet * moduleConfig = ksNew (30,
-#include "contract.h"
-					       KS_END);
-		ksAppend (ks, moduleConfig);
-		ksDel (moduleConfig);
-		return 1;
-	}
-
-	// TODO encrypt if this is a `postgetstorage` call
-
-	// regular decryption business
-	KeySet * pluginConfig = elektraPluginGetConfig (handle);
-
-	// decrypt the file with gpg - prepare argument vector for gpg call
-	static const int argc = 8;
-	char * argv[8] = { NULL, "--batch", "--yes", "-o", NULL, "-d", NULL, NULL };
-	char * tmpFile = getTemporaryFileName (keyString (parentKey));
-	if (!tmpFile)
-	{
-		ELEKTRA_SET_ERROR (87, parentKey, "Memory allocation failed");
-		return -1;
-	}
-
-	argv[4] = tmpFile;
-	// safely discarding const from keyString() return value
-	argv[6] = (char *)keyString (parentKey);
-
-	// NOTE the decryption process works like this:
-	// gpg2 --batch --yes -o tmpfile -d configFile
-	// mv tmpfile configFile
-
-	int result = elektraCryptoGpgCall (pluginConfig, parentKey, NULL, argv, argc);
-
-	if (result == 1)
-	{
-		// encryption successful, overwrite the original file with the encrypted data
-		if (rename (tmpFile, keyString (parentKey)) != 0)
-		{
-			ELEKTRA_SET_ERRORF (ELEKTRA_ERROR_FCRYPT_RENAME, parentKey, "Renaming file %s to %s failed.", tmpFile,
[…]
