### 2016-09-15 2f36ba4
fcrypt: add examples to README.md
