+++ src/plugins/fcrypt/README.md
@@ -33,7 +33,19 @@ Please refer to [crypto](../crypto/).
 
 ## Examples ##
 
-TODO
+You can mount the plugin like this:
+
+	kdb mount test.ecf /t fcrypt /gpg/key=DDEBEF9EE2DC931701338212DAF635B17F230E8D
+
+If you create a key under `/t`
+
+	kdb set /t/a "hello world"
+
+you will notice that you can not read the plain text of `test.ecf` because it has been encrypted by GPG.
+
+But you can still access `/t/a` with `kdb get`:
+
+	kdb get /t/a
 
 ## Configuration ##
 
