### 2016-09-15 28bf4fc
fcrypt: add security notice to README.md
