### 2016-09-15 c152df3
fcrypt: add entry to plugins/README.md
