+++ src/plugins/tcl/CMakeLists.txt
@@ -7,7 +7,7 @@ endif ()
 
 add_plugin(tcl
 	CPP
-	INCLUDE_DIRECTORIES
+	INCLUDE_SYSTEM_DIRECTORIES
 		${Boost_INCLUDE_DIR}
 	SOURCES
 		tcl.hpp tcl.cpp
