### 2016-09-15 e96fda5
heredoc of completions
