+++ scripts/install-sh-completion.in
@@ -52,10 +52,22 @@ if ask "Install for zsh?" Y; then
 
 cat <<HERE >>~/.zshrc
 ### generated by kdb install-sh-completion
+fpath=(~/.zsh-comp $fpath)
 autoload -U compinit && compinit
 zmodload -i zsh/complist
 ### end of generated by kdb install-sh-completion
 HERE
+
+    if [ ! -d ~/.zsh-comp ]; then
+        mkdir ~/.zsh-comp
+    fi
+
+cat <<HERE >~/.zsh-comp/_kdb
+@CONTENT_OF_ZSH_COMP@
+HERE
+
+
+
     echo "installed zsh completion configuration"
 fi
 
@@ -69,14 +81,7 @@ if ask "Install for bash?" Y; then
 
 cat <<HERE >>~/.bashrc
 ### generated by kdb install-sh-completion
-# Linux
-if [ -d /etc/bash_completion.d ]; then
-    . /etc/bash_completion.d
-fi
-# OS X
-if [ -d /usr/local/etc/bash_completion ]; then
-    . /usr/local/etc/bash_completion
-fi
+@CONTENT_OF_BASH_COMP@
 ### end of generated by kdb install-sh-completion
 HERE
     echo "installed bash completion configuration"
