### 2016-09-15 5bca952
mathckeck: activate icc again
