### 2016-09-15 ca61d28
tools: remove obsolete TODO
(numbers are used now)
