### 2016-09-16 8b63c4d
error2warning: kdbGet
