+++ src/libs/elektra/kdb.c
@@ -634,6 +634,34 @@ static int elektraGetDoUpdateWithGlobalHooks (KDB * handle, Split * split, KeySe
 	}
 	return 0;
 }
+
+static int copyError (Key * dest, Key * src)
+{
+	keyRewindMeta (src);
+	Key * metaKey = (Key *)keyGetMeta (src, "error");
+	if (!metaKey) return 0;
+	keySetMeta (dest, keyName (metaKey), keyString (metaKey));
+	while ((metaKey = (Key *)keyNextMeta (src)) != NULL)
+	{
+		if (strncmp (keyName (metaKey), "error/", 6)) break;
+		keySetMeta (dest, keyName (metaKey), keyString (metaKey));
+	}
+	return 1;
+}
+static void clearError (Key * key)
+{
+	keySetMeta (key, "error", 0);
+	keySetMeta (key, "error/number", 0);
+	keySetMeta (key, "error/description", 0);
+	keySetMeta (key, "error/reason", 0);
+	keySetMeta (key, "error/ingroup", 0);
+	keySetMeta (key, "error/module", 0);
+	keySetMeta (key, "error/file", 0);
+	keySetMeta (key, "error/line", 0);
+	keySetMeta (key, "error/configfile", 0);
+	keySetMeta (key, "error/mountpoint", 0);
+}
+
 /**
  * @brief Retrieve keys in an atomic and universal way.
  *
@@ -715,8 +743,13 @@ int kdbGet (KDB * handle, KeySet * ks, Key * parentKey)
 		return -1;
 	}
 
+	Key * oldError = keyNew (keyName (parentKey), KEY_END);
+	copyError (oldError, parentKey);
+
 	if (ns == KEY_NS_META)
 	{
+		clearError (parentKey);
+		keyDel (oldError);
 		ELEKTRA_SET_ERRORF (104, parentKey, "metakey with name \"%s\" passed to kdbGet", keyName (parentKey));
 		return -1;
 	}
@@ -735,6 +768,7 @@ int kdbGet (KDB * handle, KeySet * ks, Key * parentKey)
 
 	if (!handle || !ks)
 	{
+		clearError (parentKey);
 		ELEKTRA_SET_ERROR (37, parentKey, "handle or ks null pointer");
 		goto error;
 	}
@@ -744,6 +778,7 @@ int kdbGet (KDB * handle, KeySet * ks, Key * parentKey)
 	}
 	if (splitBuildup (split, handle, parentKey) == -1)
 	{
+		clearError (parentKey);
 		ELEKTRA_SET_ERROR (38, parentKey, "error in splitBuildup");
 		goto error;
 	}
@@ -766,15 +801,21 @@ int kdbGet (KDB * handle, KeySet * ks, Key * parentKey)
 	// Appoint keys (some in the bypass)
 	if (splitAppoint (split, handle, ks) == -1)
 	{
+		clearError (parentKey);
 		ELEKTRA_SET_ERROR (38, parentKey, "error in splitAppoint");
 		goto error;
 	}
 
 	if (handle->globalPlugins[POSTGETSTORAGE] || handle->globalPlugins[POSTGETCLEANUP])
 	{
+		clearError (parentKey);
 		if (elektraGetDoUpdateWithGlobalHooks (NULL, split, NULL, parentKey, initialParent, FIRST) == -1)
 		{
 			goto error;
+		}
+		else
+		{
+			copyError (parentKey, oldError);
 		}
 
 		keySetName (parentKey, keyName (initialParent));
@@ -787,9 +828,14 @@ int kdbGet (KDB * handle, KeySet * ks, Key * parentKey)
 		ksClear (ks);
 		splitMerge (split, ks);
 
+		clearError (parentKey);
[…]
