### 2016-09-16 3822b50
cmake: try to workaround gcc -isystem bug
