+++ cmake/Modules/LibAddPlugin.cmake
@@ -100,7 +100,7 @@ function (add_plugintest testname)
 		set_property(TARGET ${testexename}
 				APPEND PROPERTY INCLUDE_DIRECTORIES
 				${ARG_INCLUDE_DIRECTORIES})
-		if (ARG_INCLUDE_SYSTEM_DIRECTORIES)
+		if (ARG_INCLUDE_SYSTEM_DIRECTORIES AND NOT ARG_INCLUDE_SYSTEM_DIRECTORIES STREQUAL "/usr/include")
 			set_property(TARGET ${testexename}
 					APPEND PROPERTY COMPILE_FLAGS
 					"${CMAKE_INCLUDE_SYSTEM_FLAG_CXX} ${ARG_INCLUDE_SYSTEM_DIRECTORIES}")
@@ -367,7 +367,7 @@ function (add_plugin PLUGIN_SHORT_NAME)
 		${CMAKE_CURRENT_BINARY_DIR} #for readme
 		)
 
-	if (ARG_INCLUDE_SYSTEM_DIRECTORIES)
+	if (ARG_INCLUDE_SYSTEM_DIRECTORIES AND NOT ARG_INCLUDE_SYSTEM_DIRECTORIES STREQUAL "/usr/include")
 		set_property(TARGET ${PLUGIN_OBJS}
 			APPEND PROPERTY COMPILE_FLAGS
 			"${CMAKE_INCLUDE_SYSTEM_FLAG_CXX} ${ARG_INCLUDE_SYSTEM_DIRECTORIES} ${CMAKE_PIC_FLAGS}"
@@ -408,7 +408,7 @@ function (add_plugin PLUGIN_SHORT_NAME)
 			${ARG_INCLUDE_DIRECTORIES}
 			${CMAKE_CURRENT_BINARY_DIR} #for readme
 			)
-		if (ARG_INCLUDE_SYSTEM_DIRECTORIES)
+		if (ARG_INCLUDE_SYSTEM_DIRECTORIES AND NOT ARG_INCLUDE_SYSTEM_DIRECTORIES STREQUAL "/usr/include")
 			set_property(TARGET ${PLUGIN_NAME}
 				APPEND PROPERTY COMPILE_FLAGS
 				"${CMAKE_INCLUDE_SYSTEM_FLAG_CXX} ${ARG_INCLUDE_SYSTEM_DIRECTORIES}"
