### 2016-09-16 b478954
test: fix KeySet syntax
