+++ src/plugins/mozprefs/README.md
@@ -11,8 +11,8 @@
 
 ## Basics ##
 
-This plugin works on [Mozilla preference files](https://developer.mozilla.org/en-US/docs/Mozilla/Preferences/A_brief_guide_to_Mozilla_preferences) and is used in Elektras [Firefox autoconfig script](./autconfig/README.md). 
-
+This plugin works on [Mozilla preference files](https://developer.mozilla.org/en-US/docs/Mozilla/Preferences/A_brief_guide_to_Mozilla_preferences)
+and is used in Elektras [Firefox autoconfig script](autoconfig/README.md).
 
 ### Preference Types ###
 
