### 2016-09-16 8332d17
crypto-plugin: fix static link issues
Make exported function names unique by using the CRYPTO_PLUGIN_FUNCTION macro.
See #951 for full discussion.
