+++ src/plugins/crypto/gcrypt_operations.c
@@ -46,17 +46,17 @@ static int getKeyIvForEncryption (KeySet * config, Key * errorKey, Key * k, Key 
 
 	// generate the salt
 	gcry_create_nonce (salt, sizeof (salt));
-	saltHexString = elektraCryptoBin2Hex (errorKey, salt, sizeof (salt));
-	if (!saltHexString) return -1; // error set by elektraCryptoBin2Hex()
+	saltHexString = CRYPTO_PLUGIN_FUNCTION (bin2hex) (errorKey, salt, sizeof (salt));
+	if (!saltHexString) return -1; // error set by CRYPTO_PLUGIN_FUNCTION(bin2hex)()
 	keySetMeta (k, ELEKTRA_CRYPTO_META_SALT, saltHexString);
 	elektraFree (saltHexString);
 
 	// read iteration count
-	const kdb_unsigned_long_t iterations = elektraCryptoGetIterationCount (errorKey, config);
+	const kdb_unsigned_long_t iterations = CRYPTO_PLUGIN_FUNCTION (getIterationCount) (errorKey, config);
 
 	// receive master password from the configuration
-	Key * msg = elektraCryptoGetMasterPassword (errorKey, config);
-	if (!msg) goto error; // error set by elektraCryptoGetMasterPassword()
+	Key * msg = CRYPTO_PLUGIN_FUNCTION (getMasterPassword) (errorKey, config);
+	if (!msg) goto error; // error set by CRYPTO_PLUGIN_FUNCTION(getMasterPassword)()
 
 	// generate/derive the cryptographic key and the IV
 	if ((gcry_err = gcry_kdf_derive (keyValue (msg), keyGetValueSize (msg), GCRY_KDF_PBKDF2, GCRY_MD_SHA512, salt, sizeof (salt),
@@ -96,17 +96,17 @@ static int getKeyIvForDecryption (KeySet * config, Key * errorKey, Key * k, Key 
 	kdb_unsigned_long_t saltBufferLen = 0;
 
 	// get the salt
-	if (elektraCryptoGetSaltFromCryptoPayload (errorKey, k, &saltBuffer, &saltBufferLen) != 1)
+	if (CRYPTO_PLUGIN_FUNCTION (getSaltFromPayload) (errorKey, k, &saltBuffer, &saltBufferLen) != 1)
 	{
-		return -1; // error set by elektraCryptoGetSaltFromCryptoPayload()
+		return -1; // error set by CRYPTO_PLUGIN_FUNCTION(getSaltFromPayload)()
 	}
 
 	// get the iteration count
-	const kdb_unsigned_long_t iterations = elektraCryptoGetIterationCount (errorKey, config);
+	const kdb_unsigned_long_t iterations = CRYPTO_PLUGIN_FUNCTION (getIterationCount) (errorKey, config);
 
 	// receive master password from the configuration
-	Key * msg = elektraCryptoGetMasterPassword (errorKey, config);
-	if (!msg) goto error; // error set by elektraCryptoGetMasterPassword()
+	Key * msg = CRYPTO_PLUGIN_FUNCTION (getMasterPassword) (errorKey, config);
+	if (!msg) goto error; // error set by CRYPTO_PLUGIN_FUNCTION(getMasterPassword)()
 
 	// derive the cryptographic key and the IV
 	if ((gcry_err = gcry_kdf_derive (keyValue (msg), keyGetValueSize (msg), GCRY_KDF_PBKDF2, GCRY_MD_SHA512, saltBuffer, saltBufferLen,
[…]
