+++ src/plugins/crypto/gpg.h
@@ -10,6 +10,7 @@
 #ifndef ELEKTRA_PLUGIN_CRYPTO_GPG_H
 #define ELEKTRA_PLUGIN_CRYPTO_GPG_H
 
+#include "crypto.h"
 #include <kdb.h>
 #include <kdbtypes.h>
 
@@ -19,8 +20,8 @@
 #define ELEKTRA_CRYPTO_DEFAULT_GPG2_BIN "/usr/bin/gpg2"
 #define ELEKTRA_CRYPTO_DEFAULT_GPG1_BIN "/usr/bin/gpg"
 
-int elektraCryptoGpgEncryptMasterPassword (KeySet * conf, Key * errorKey, Key * msgKey);
-int elektraCryptoGpgDecryptMasterPassword (KeySet * conf, Key * errorKey, Key * msgKey);
-int elektraCryptoGpgCall (KeySet * conf, Key * errorKey, Key * msgKey, char * argv[], size_t argc);
+int CRYPTO_PLUGIN_FUNCTION (gpgEncryptMasterPassword) (KeySet * conf, Key * errorKey, Key * msgKey);
+int CRYPTO_PLUGIN_FUNCTION (gpgDecryptMasterPassword) (KeySet * conf, Key * errorKey, Key * msgKey);
+int CRYPTO_PLUGIN_FUNCTION (gpgCall) (KeySet * conf, Key * errorKey, Key * msgKey, char * argv[], size_t argc);
 
 #endif
