### 2016-09-16 23cde13
doc: add template for NEWS+add script
