+++ doc/todo/NEWS.md
@@ -1,0 +1,68 @@
+== Template for NEWS.md ==
+
+# 0.8.<<VERSION>> Release
+
+<<`scripts/generate-news-entry`>>
+
+
+
+## What is Elektra?
+
+Elektra serves as a universal and secure framework to access configuration
+parameters in a global, hierarchical key database.
+For a small demo see here:
+
+[![asciicast](https://asciinema.org/a/cantr04assr4jkv8v34uz9b8r.png)](https://asciinema.org/a/cantr04assr4jkv8v34uz9b8r)
+
+## Highlights
+
+- <<HIGHLIGHTS>>
+
+
+## <<ONE POINT PER HIGHLIGHT>>
+
+## Documentation
+
+## Compatibility
+
+As always, the ABI and API of kdb.h is fully compatible, i.e. programs
+compiled against an older 0.8 version of Elektra will continue to work
+(ABI) and you will be able to recompile programs without errors (API).
+
+
+## Other
+
+## Get It!
+
+You can download the release from
+[here](http://www.libelektra.org/ftp/elektra/releases/elektra-0.8.<<VERSION>>.tar.gz)
+and also [here on github](https://github.com/ElektraInitiative/ftp/tree/master/releases/elektra-0.8.<<VERSION>>.tar.gz)
+
+<<`scripts/generate-hashsums`>>
+
+This release tarball now is also available
+[signed by me using gpg](http://www.libelektra.org/ftp/elektra/releases/elektra-0.8.<<VERSION>>.tar.gz.gpg)
+
+already built API-Docu can be found [here](http://doc.libelektra.org/api/0.8.<<VERSION>>/html/)
+
+
+## Stay tuned! ##
+
+Subscribe to the
+[RSS feed](http://doc.libelektra.org/news/feed.rss)
+to always get the release notifications.
+
+For any questions and comments, please contact the
+[Mailing List](https://lists.sourceforge.net/lists/listinfo/registry-list)
+the issue tracker [on github](http://git.libelektra.org/issues)
+or by email elektra@markus-raab.org.
+
+[Permalink to this NEWS entry](http://doc.libelektra.org/news/<<UID>>.html)
+
+For more information, see [http://libelektra.org](http://libelektra.org)
+
+Best regards,
+Markus
+
+
+
