### 2016-09-16 a6dc717
small text fixes, PUT /kdb instead of POST
