### 2016-09-16 f975a45
use apiblueprint attributes and data structures
