### 2016-09-16 8c0ef28
remove overriding attributes in example
