### 2016-09-16 61837d1
REST service: use-cases updated (markdown, perspective changed to web gui, added new use cases)
