+++ doc/rest_api/snippet_sharing/use_cases/UC_download_snippet.txt
@@ -1,17 +1,19 @@
-Summary:
-----------------------------
-Title: download configuration snippet in specific format
-Scope: Detailed View
+# Use Case: Download configuration snippet in specific format
+
+## Summary
+
+Title: Download configuration snippet in specific format
+Scope: Entry Details
 Level: User Goal
-Actors: Anonymous User (+ all others)
+Actors: Anonymous user (+ all others)
 Brief: Any user can download the configuration snippets in various formats (XML, JSON, ini, ...).
 
-Scenarios:
-----------------------------
-Precondition: URI of a snippet entry is known
-Main success scenario: User successfully downloads the raw configuration snippet in the selected format. The response output can then be stored directly as configuration file.
-Alternative scenario: The specified entry (URI) cannot be found. The user is informed about the error by a response.
-Error scenario: Technical problems prevent the operation from completing. The user is informed about the error by a response containing further information.
-Postcondition: User retrieves the raw snippet, but the database remains unchanged.
+## Scenarios
+
+Precondition: Entry has been found through search or detailed view of an entry is opened
+Main success scenario: User successfully downloads the raw configuration snippet in the selected format. The snippet can then be stored directly as configuration file.
+Alternative scenario: The specified entry (URI) cannot be found. The user is informed about the error by an error message.
+Error scenario: Technical problems prevent the operation from completing. The user is informed about the error by an error message containing further information.
+Postcondition: The database remains unchanged.
 Non-functional Constraints: 
 	- Essential functionality of the program
