### 2016-09-16 5e519b9
single quotes fixes the problems with heredoc errors
