+++ scripts/install-sh-completion.in
@@ -62,7 +62,7 @@ HERE
         mkdir ~/.zsh-comp
     fi
 
-cat <<HERE >~/.zsh-comp/_kdb
+cat << 'HERE' >~/.zsh-comp/_kdb
 @CONTENT_OF_ZSH_COMP@
 HERE
 
@@ -79,7 +79,7 @@ if ask "Install for bash?" Y; then
         fi
     fi
 
-cat <<HERE >>~/.bashrc
+cat << 'HERE' >>~/.bashrc
 ### generated by kdb install-sh-completion
 @CONTENT_OF_BASH_COMP@
 ### end of generated by kdb install-sh-completion
