### 2016-09-16 ee1ad0f
doc: default-dir of bash-completion
