### 2016-09-16 e6e880e
doc: first draft of release notes
