+++ doc/NEWS.md
@@ -1,3 +1,172 @@
+# 0.8.18 Release
+
+Release did not happen yet but is expected for today!
+
+- guid: 190576e0-9fef-486e-b8da-c4e75be08329
+- author: Markus Raab
+- pubDate: Fri, 16 Sep 2016 23:31:27 +0200
+
+
+
+## What is Elektra?
+
+Elektra serves as a universal and secure framework to access configuration
+parameters in a global, hierarchical key database.
+For a small demo see here:
+
+[![asciicast](https://asciinema.org/a/cantr04assr4jkv8v34uz9b8r.png)](https://asciinema.org/a/cantr04assr4jkv8v34uz9b8r)
+
+## Highlights
+
+- Intercept open syscalls which allows Elektra to dynamically
+  generate config files from its database
+- Experimental version of cryptographic plugins
+- A new zsh completion file (next to the bash completion file)
+- Gitresolver allows to directly read and write config files
+  from git instead of files present in the file system.
+- Survey completed successfully (and debts paid), we now
+  prepare raw data
+
+### Crypto Plugin
+
+A new botan backend was implemented.
+
+Furthermore gpg is used to decrypt a master password, which is used
+by the individual backends.
+
+Thanks to Peter Nirschl.
+
+### Open Interception
+
+When Elektra directly modifies config files which are on the disc,
+and applications read the config files without Elektra, Elektra has
+no control over the access, e.g. we cannot dynamically calculate
+values. To avoid this, we wrote a library that
+intercepts the `open`-call.
+
+Together with the `mozprefs` plugin, we got control over the configuration
+of Firefox and can dynamically change config values with all possibilities
+Elektra provides.
+
+For easy setup, we implemented the script `configure-firefox`.
+
+Thanks to Thomas Waser.
+
+
+
+### zsh completion
+
+Added zsh completion file, thanks to Sebastian Bachmann
+
+## Documentation
+
+- fix kdb-import man page, thanks to Kurt Micheli
+- mark keyIsSystem/keyIsUser as internal
+- fix doxygen reference to example
+- better document that `global-mount` or `gmount` will overwrite
+  previously mounted global plugins
+- fix spelling mistake, thanks to René Schwaiger
+- Wrote tutorial how to use Elektra-python bindings,
+  thanks to Ulrike Schäfer
+
+
+## Quality
+
+- find-tools now pep and pyflakes happy, thanks to Kurt Micheli
+- fix bashism, thanks to Thomas Waser and Kurt Micheli
+- better error message for conditionals plugin, thanks to Thomas Waser
+- better error message for augeas plugin, thanks to Felix Berlakovich
+- Many compilation warnings fixed, thanks to Gabriel Rauter, Thomas Waser
+
+
+## Compatibility
+
+As always, the ABI and API of kdb.h is fully compatible, i.e. programs
+compiled against an older 0.8 version of Elektra will continue to work
+(ABI) and you will be able to recompile programs without errors (API).
+
+### Libtools
+
+Libtools got a new major version (SOVERSION 0 -> 1):
+
+- resolveRecommends was never implemented and was now removed
+- backend/plugin configs are now validated by plugins (needed
+  by gpg plugin, which checks if wrong key IDs are supplied during mount)
+
+## Development
+
+- github descriptions+workflow (showed when creating PRs and issues)
+- many new trigger phases for github, see [doc/GIT](http://git.libelektra.org/tree/master/doc/GIT.md)
+  thanks to Mihael Pranjić
+- valgrind suppressions are great again, thanks to Peter Nirschl
[…]
