### 2016-09-16 3bb4b69
cmake: do not make mac os x installs fail
for #844
