### 2016-09-16 bba764c
Remove to small headline, because ronn does not like it
