### 2016-09-16 cf88aab
doc: check compatibility
