### 2016-09-16 9fec81a
doc: release notes
