### 2016-09-16 4d1101e
doc: fix keyIsBelow
thanks to @tom-wa
