### 2016-09-16 3af25d7
doc: add botan link
thanks to @petermax2
