### 2016-09-16 08b2d5a
tools: allow to mount to system
fix #953
