+++ src/libs/tools/src/backend.cpp
@@ -392,10 +392,6 @@ void Backend::serialize (kdb::KeySet & ret)
 	}
 	else if (mp.at (0) == '/')
 	{
-		Key k ("system" + mp, KEY_END);
-		Key restrictedPath ("system/elektra", KEY_END);
-		if (!k) throw MountpointInvalidException ();
-		if (restrictedPath.isBelow (k)) throw MountpointInvalidException ();
 		ret.append (*Key (backendRootKey.getName () + "/mountpoint", KEY_VALUE, mp.c_str (), KEY_COMMENT,
 				  "The mountpoint says the location where the backend should be mounted.\n"
 				  "This is a cascading mountpoint.\n"
@@ -404,10 +400,6 @@ void Backend::serialize (kdb::KeySet & ret)
 	}
 	else
 	{
-		Key k (mp, KEY_END);
-		Key restrictedPath ("system/elektra", KEY_END);
-		if (!k) throw MountpointInvalidException ();
-		if (restrictedPath.isBelow (k)) throw MountpointInvalidException ();
 		ret.append (*Key (backendRootKey.getName () + "/mountpoint", KEY_VALUE, mp.c_str (), KEY_COMMENT,
 				  "The mountpoint says the location where the backend should be mounted.\n"
 				  "This is a normal mountpoint.\n",
