### 2016-09-16 3511c60
cmake: use Darwin for checks
(checks more specific for Mac OS X)
+ reformat other check
