+++ scripts/CMakeLists.txt
@@ -46,19 +46,19 @@ if (INSTALL_SYSTEM_FILES)
 			DESTINATION ${BASH_COMPLETION_COMPLETIONSDIR}
 			RENAME kdb)
 
-    if (${CMAKE_SYSTEM_NAME} MATCHES "Darwin")
-        # On Mac OS X we need to copy this into another folder.
-        # Because of "System Integrity Protection", we need to 
-        # use a /usr/local path
-        install (FILES kdb_zsh_completion
-                DESTINATION /usr/local/share/zsh/site-functions
-                RENAME _kdb)
-    else()
-        # This Folder is at least the valid folder for Debian
-        install (FILES kdb_zsh_completion
-                DESTINATION /usr/share/zsh/vendor-completions
-                RENAME _kdb)
-    endif()
+	if (${CMAKE_SYSTEM_NAME} MATCHES "Darwin")
+		# On Mac OS X we need to copy this into another folder.
+		# Because of "System Integrity Protection", we need to 
+		# use a /usr/local path
+		install (FILES kdb_zsh_completion
+			DESTINATION /usr/local/share/zsh/site-functions
+			RENAME _kdb)
+	else ()
+		# This Folder is at least the valid folder for Debian
+		install (FILES kdb_zsh_completion
+			DESTINATION /usr/share/zsh/vendor-completions
+			RENAME _kdb)
+	endif ()
 endif()
 
 configure_file(
