### 2016-09-16 762331f
cmake: fix installation of install-sh-completion
