### 2016-09-16 359a207
doc: changes in installed files
