### 2016-09-16 6411e0f
doc: improve NEWS
