### 2016-09-16 4ef143d
intercept open: fixed typo in README.md
