### 2016-09-16 8354c26
doc: final release notes
