+++ doc/NEWS.md
@@ -1,11 +1,8 @@
 # 0.8.18 Release
 
-Release did not happen yet but is expected for today!
-
 - guid: 190576e0-9fef-486e-b8da-c4e75be08329
 - author: Markus Raab
 - pubDate: Fri, 16 Sep 2016 23:31:27 +0200
-
 
 
 ## What is Elektra?
@@ -19,19 +16,20 @@ For a small demo see here:
 ## Highlights
 
 - Intercept open syscalls which allows Elektra to dynamically
-  generate config files from its database
+  generate config files from Elektra's database
 - Experimental version of cryptographic plugins
 - A new zsh completion file (next to the bash completion file)
 - Gitresolver allows to directly read and write config files
   from git instead of files present in the file system.
-- Survey completed successfully (and debts paid), we now
-  prepare raw data
+- Survey completed successfully (and debts paid), we are now
+  preparing raw data.
 
 
 ### Crypto Plugin
 
 Gpg is now used to decrypt a master password, which is used
-by the individual crypto backends.
+by the individual crypto backends. So all necessary parts
+for encryption of decryption of individual keys is present.
 
 Furthermore, a new [botan](https://botan.randombit.net) backend
 was implemented.
@@ -87,8 +85,8 @@ Thanks to Sebastian Bachmann.
 
 ## Documentation
 
-- fix kdb-import man page, thanks to Kurt Micheli
-- mark keyIsSystem/keyIsUser as internal
+- fix `kdb-import` man page, thanks to Kurt Micheli
+- mark `keyIsSystem`/`keyIsUser` as internal
 - fix doxygen reference to example
 - better document that `global-mount` or `gmount` will overwrite
   previously mounted global plugins
@@ -100,14 +98,15 @@ Thanks to Sebastian Bachmann.
 ## Quality
 
 - shell recorder test cases now run during `make test`, thanks to Kurt
-  Micheli and René Schwaiger
-- find-tools now pep and pyflakes happy, thanks to Kurt Micheli
+  Micheli and René Schwaiger (Warning: might remove present keys
+  when it conflicts with their mountpoints)
+- find-tools now is pep and pyflakes happy, thanks to Kurt Micheli
 - fix bashism, thanks to Thomas Waser and Kurt Micheli
 - better error message for conditionals plugin, thanks to Thomas Waser
 - better error message for augeas plugin, thanks to Felix Berlakovich
 - Many compilation warnings fixed, thanks to Gabriel Rauter, Thomas Waser
 - GSettings: fix double free, thanks to Gabriel Rauter
-- Fix external links and add an external link checker,
+- Fix external links and implement an external link checker,
   thanks to Kurt Micheli
 - Fix openwrt/musl warnings with wrong printf format, thanks to
   Thomas Waser
@@ -125,18 +124,19 @@ compiled against an older 0.8 version of Elektra will continue to work
 
 Libtools got a new major version (SOVERSION 0 -> 1):
 
-- resolveRecommends was never implemented and was now removed
 - backend/plugin configs are now validated by plugins (needed
   by gpg plugin, which checks if wrong key IDs are supplied during mount)
+- resolveRecommends was never implemented and was now removed
 
 ### Plugins
 
-The plugins conditionals and mathcheck are incompatible because of changes
-in syntax.
+The plugins conditionals and mathcheck are incompatible in some cases
+because of changes in syntax.
 
 ## Development
 
-- github descriptions+workflow (showed when creating PRs and issues)
[…]
