### 2016-09-16 770e356
doc: add about keyRel2 proposal
