### 2016-09-17 9686e04
fcrypt: adapt to renamed functions in gpg module
Replace elektraCryptoGpgCall with CRYPTO_PLUGIN_FUNCTION (gpgCall)
