### 2016-09-17 5ae8cf6
fcrypt: implement test mode
Preparation for units testing
