### 2016-09-17 9c0d67e
fcrypt: add unit tests
