+++ src/plugins/fcrypt/testmod_fcrypt.c
@@ -15,6 +15,168 @@
 #include <tests_internal.h>
 #include <tests_plugin.h>
 
+#include <gpg.h>
+#include <test_key.h>
+
+#define PLUGIN_NAME "fcrypt"
+#define TEST_KEY_ID "DDEBEF9EE2DC931701338212DAF635B17F230E8D"
+#define TEST_FILE "fcrypt_testfile"
+
+static const kdb_octet_t testContent[] = { 0x01, 0x02, 0xCA, 0xFE, 0xBA, 0xBE, 0x03, 0x04 };
+
+/*
+ * @brief create temporary file name
+ * @returns allocated string holding the path to the temporary file. Must be freed by the caller.
+ */
+static char * getTemporaryFileName ()
+{
+	const size_t newFileAllocated = strlen (TEST_FILE) + 7;
+	char * newFile = elektraMalloc (newFileAllocated);
+	if (!newFile) return NULL;
+	snprintf (newFile, newFileAllocated, "%sXXXXXX", TEST_FILE);
+	mkstemp (newFile);
+	return newFile;
+}
+
+static KeySet * newPluginConfiguration ()
+{
+	return ksNew (2, keyNew (ELEKTRA_CRYPTO_PARAM_GPG_KEY, KEY_VALUE, TEST_KEY_ID, KEY_END),
+		      keyNew (ELEKTRA_CRYPTO_PARAM_GPG_UNIT_TEST, KEY_VALUE, "1", KEY_END), KS_END);
+}
+
+static void writeTestFile (const char * file)
+{
+	FILE * f = fopen (file, "wb");
+	succeed_if (f, "can not write to temporary file");
+	if (!f) return;
+	succeed_if (fwrite (testContent, 1, sizeof (testContent), f) == sizeof (testContent), "test file preparation failed");
+	fclose (f);
+}
+
+/**
+ * @brief read in the content stored in file and compare it to the test vector.
+ * @retval 1 if the file content is equal to the test vector.
+ * @retval -1 if the file content is not equal to the test vector or an error occured.
+ */
+static int isTestFileCorrect (const char * file)
+{
+	int returnValue = -1;
+	kdb_octet_t readBuffer[2 * sizeof (testContent)] = { 0 };
+	size_t readCount = 0;
+
+	FILE * f = fopen (file, "rb");
+	succeed_if (f, "can not read from temporary file");
+	if (!f) return -1;
+
+	readCount = fread (readBuffer, 1, sizeof (readBuffer), f);
+	succeed_if (readCount > 0, "temporary file is empty.");
+
+	if (readCount == sizeof (testContent))
+	{
+		if (memcmp (readBuffer, testContent, sizeof (testContent)) == 0)
+		{
+			returnValue = 1;
+		}
+	}
+
+	fclose (f);
+	return returnValue;
+}
+
+static void test_init ()
+{
+	Plugin * plugin = NULL;
+	Key * parentKey = keyNew ("system", KEY_END);
+	KeySet * modules = ksNew (0, KS_END);
+	KeySet * configKs = newPluginConfiguration ();
+	elektraModulesInit (modules, 0);
+
+	plugin = elektraPluginOpen (PLUGIN_NAME, modules, configKs, 0);
+	succeed_if (plugin != 0, "failed to open the plugin");
+	if (plugin)
+	{
+		succeed_if (!strcmp (plugin->name, PLUGIN_NAME), "got wrong name");
+
+		KeySet * config = elektraPluginGetConfig (plugin);
+		succeed_if (config != 0, "there should be a config");
+
+		succeed_if (plugin->kdbOpen != 0, "no open pointer");
+		succeed_if (plugin->kdbClose != 0, "no close pointer");
[…]
