### 2016-09-17 35dda33
base64: working Base64 encoding
See #897
