+++ src/plugins/base64/testmod_base64.c
@@ -1,0 +1,181 @@
+/**
+ * @file
+ *
+ * @brief test suite for the fcrypt plugin
+ *
+ * @copyright BSD License (see doc/COPYING or http://www.libelektra.org)
+ *
+ */
+
+#include <kdb.h>
+#include <kdbinternal.h>
+#include <stdio.h>
+#include <stdlib.h>
+#include <string.h>
+#include <tests_internal.h>
+#include <tests_plugin.h>
+
+#include "base64.h"
+
+static KeySet * newPluginConfiguration()
+{
+	return ksNew (0, KS_END);
+}
+
+static void test_init ()
+{
+	Plugin * plugin = NULL;
+	Key * parentKey = keyNew ("system", KEY_END);
+	KeySet * modules = ksNew (0, KS_END);
+	KeySet * configKs = newPluginConfiguration ();
+	elektraModulesInit (modules, 0);
+
+	plugin = elektraPluginOpen (ELEKTRA_PLUGIN_NAME, modules, configKs, 0);
+	succeed_if (plugin != 0, "failed to open the plugin");
+	if (plugin)
+	{
+		succeed_if (!strcmp (plugin->name, ELEKTRA_PLUGIN_NAME), "got wrong name");
+
+		KeySet * config = elektraPluginGetConfig (plugin);
+		succeed_if (config != 0, "there should be a config");
+
+		succeed_if (plugin->kdbGet != 0, "no get pointer");
+		succeed_if (plugin->kdbSet != 0, "no set pointer");
+
+		elektraPluginClose (plugin, 0);
+	}
+
+	elektraModulesClose (modules, 0);
+	ksDel (modules);
+	keyDel (parentKey);
+}
+
+static void test_base64_encoding ()
+{
+	// test vectors are defined in RFC4648
+	// see https://www.ietf.org/rfc/rfc4648.txt
+	char * t1 = ELEKTRA_PLUGIN_FUNCTION (ELEKTRA_PLUGIN_NAME, base64Encode) ((kdb_octet_t*)"", 0);
+	succeed_if (t1, "can not allocate string for encoding result");
+	if (t1)
+	{
+		succeed_if (strcmp (t1, "") == 0, "encoding result from test vector 1 does not match");
+		elektraFree (t1);
+	}
+
+	char * t2 = ELEKTRA_PLUGIN_FUNCTION (ELEKTRA_PLUGIN_NAME, base64Encode) ((kdb_octet_t*)"f", 1);
+	succeed_if (t2, "can not allocate string for encoding result");
+	if (t2)
+	{
+		succeed_if (strcmp (t2, "Zg==") == 0, "encoding result from test vector 2 does not match");
+		elektraFree (t2);
+	}
+
+	char * t3 = ELEKTRA_PLUGIN_FUNCTION (ELEKTRA_PLUGIN_NAME, base64Encode) ((kdb_octet_t*)"fo", 2);
+	succeed_if (t3, "can not allocate string for encoding result");
+	if (t3)
+	{
+		succeed_if (strcmp (t3, "Zm8=") == 0, "encoding result from test vector 3 does not match");
+		elektraFree (t3);
+	}
+
+	char * t4 = ELEKTRA_PLUGIN_FUNCTION (ELEKTRA_PLUGIN_NAME, base64Encode) ((kdb_octet_t*)"foo", 3);
+	succeed_if (t4, "can not allocate string for encoding result");
+	if (t4)
+	{
+		succeed_if (strcmp (t4, "Zm9v") == 0, "encoding result from test vector 4 does not match");
+		elektraFree (t4);
+	}
+
+	char * t5 = ELEKTRA_PLUGIN_FUNCTION (ELEKTRA_PLUGIN_NAME, base64Encode) ((kdb_octet_t*)"foob", 4);
[…]
