+++ src/plugins/base64/base64.h
@@ -1,0 +1,27 @@
+/**
+ * @file
+ *
+ * @brief filter plugin for the Base64 encoding
+ *
+ * @copyright BSD License (see doc/COPYING or http://www.libelektra.org)
+ *
+ */
+
+#ifndef ELEKTRA_PLUGIN_BASE64_H
+#define ELEKTRA_PLUGIN_BASE64_H
+
+#include <kdbplugin.h>
+#include <kdbtypes.h>
+#include <stdio.h>
+
+// encoding functions
+char * ELEKTRA_PLUGIN_FUNCTION (ELEKTRA_PLUGIN_NAME, base64Encode) (kdb_octet_t * input, size_t inputLength);
+int ELEKTRA_PLUGIN_FUNCTION (ELEKTRA_PLUGIN_NAME, base64Decode) (const char * input, kdb_octet_t ** output, size_t * outputLength);
+
+// kdb functions
+int ELEKTRA_PLUGIN_FUNCTION (ELEKTRA_PLUGIN_NAME, get) (Plugin * handle, KeySet * ks, Key * parentKey);
+int ELEKTRA_PLUGIN_FUNCTION (ELEKTRA_PLUGIN_NAME, set) (Plugin * handle, KeySet * ks, Key * parentKey);
+
+Plugin * ELEKTRA_PLUGIN_EXPORT (base64);
+
+#endif
