+++ src/plugins/base64/CMakeLists.txt
@@ -1,0 +1,13 @@
+include (LibAddPlugin)
+
+add_plugin (base64
+	SOURCES
+		base64.h
+		base64.c
+	COMPILE_DEFINITIONS
+		ELEKTRA_PLUGIN_NAME=\"base64\"
+)
+
+if (ADDTESTING_PHASE)
+	add_plugintest (base64)
+endif ()

[…]
