### 2016-09-17 7b40f39
base64: working decoding
See #897
