### 2016-09-18 6fd12e8
base64: fix invalid writes during encoding
Used to have wrong output buffer size during encoding.
