### 2016-09-18 0a336ba
base64: optimize decoding operation
Remove unneccessary bit masking
