### 2016-09-18 b66aa79
base64: plugin specification (README.md)
