### 2016-09-18 9696074
base64: implement kdb get/set methods
