### 2016-09-19 3aa6a0b
REST service: api-description updated, use cases updated, borders for use cases defined
