### 2016-09-19 0ae00f2
REST service: technical description, renamed use case files
