### 2016-09-19 5f02b92
REST service: remove UC from PR
