### 2016-09-19 7761db6
use cases for sharing of conf snippets + template
