+++ doc/usecases/snippet_sharing/UC_search_snippet.md
@@ -1,0 +1,19 @@
+# Use Case: Search for configuration snippets
+
+## Summary
+
+Title: Search for configuration snippets
+Scope: Search
+Level: User Goal
+Actors: Anonymous user (+ all others)
+Brief: Any user can use the search to lookup configuration snippets.
+
+## Scenarios
+
+Precondition:
+Main success scenario: User is presented a list of possible candidates that match the entered search parameters.
+Alternative scenario: The search did not find any entry matching the given search parameters, which will result in an empty list being displayed.
+Error scenario: Technical problems prevent the search from succeeding. The user is informed about the issue.
+Postcondition: The search results offer additional functionality like downloading of configuration snippets or detailed view.
+Non-functional Constraints: 
+	- Essential functionality of the program
