### 2016-09-19 7f12434
sharing of conf snippets: implementation details and technical description
