### 2016-09-19 bfa112a
fcrypt: add ordering to README.md
Set ordering constraint to place fcrypt before sync.
