+++ src/plugins/fcrypt/README.md
@@ -5,6 +5,7 @@
 - infos/needs =
 - infos/recommends =
 - infos/placements = pregetstorage postgetstorage precommit
+- infos/ordering = sync
 - infos/status = experimental unfinished discouraged
 - infos/metadata =
 - infos/description = File Encryption
