### 2016-09-19 6f2ae9b
fcrypt: add fsync functionality
As discussed in #945 . This commit has to be reverted after unlimited plugins can be placed.
