+++ src/plugins/fcrypt/README.md
@@ -1,7 +1,7 @@
 - infos = Information about fcrypt plugin is in keys below
 - infos/author = Peter Nirschl <peter.nirschl@gmail.com>
 - infos/licence = BSD
-- infos/provides = filefilter crypto
+- infos/provides = sync filefilter crypto
 - infos/needs =
 - infos/recommends =
 - infos/placements = pregetstorage postgetstorage precommit
