### 2016-09-19 c0c5027
fcrypt: fix broken code formatting
