### 2016-09-19 42227dc
fcrypt: use existing error code for renaming errors
Replace ELEKTRA_ERROR_FCRYPT_RENAME with error number 31
