### 2016-09-19 cae1395
fcrypt: minor improvements
- spelling in comments
- free plugin data at kdb close
- remove stdio.h from fcrypt.h
