### 2016-09-20 b843b49
doc: some todos
