### 2016-09-20 f93e497
logger: make it simpler: do not provide modules
