### 2016-09-20 #44 @arfon
@danielskatz - I've updated the issue metadata (in the original post at the top of this issue now). Once you've found someone to help review this issue you can do:

As there's no 'meta' issue for the reviewer assignment here you *don't* need to do the `@whedon start review` bit.
