### 2016-09-20 #44 @danielskatz
@eschnett, you should now be able to do this review.  You should review the reviewer instructions on http://joss.theoj.org/about.  Let me know if you have any problems
