### 2016-10-01 ElektraInitiative/libelektra#995 @eschnett
Rename license file to `LICENSE`
The license file is currently called `COPYING`. This is confusing, as this is usually only the case when using the GPL as license, not for a BSD license. Also, several automated tools that expect the file name `LICENSE` or `LICENSE.md`.
