### 2016-10-01 ElektraInitiative/libelektra#997 @eschnett
Several self-tests failing
I installed `libelektra` on my system (macOs Sierra), using gcc 6.1, and then ran the self-tests with `make test`. Several tests failed:

```
The following tests FAILED:
	 10 - testtool_mergingkdb (Failed)
	 60 - testscr_check_distribution (Failed)
	 61 - testscr_check_error (Failed)
	 65 - testscr_check_get_set (Failed)
	 68 - testscr_check_kdb_internal_suite (Failed)
	 70 - testscr_check_mount (Failed)
	 73 - testscr_check_real_world (Failed)
	 74 - testscr_check_resolver (Failed)
	 75 - testscr_check_spec (Failed)
	 78 - testshell_globaltest (Failed)
	 79 - testshell_hosts (Failed)
	 80 - testshell_listtest (Failed)
	 81 - testshell_mathcheck (Failed)
	 82 - testshell_profiletest (Failed)
	 83 - testshell_script (Failed)
	112 - testkdb_conflict (Failed)
	113 - testkdb_nested (Failed)
	114 - testkdb_simple (Failed)
Errors while running CTest
```

Is this to be expected?
