### 2016-10-01 1c440ab
doc: rename COPYING to LICENCE.md
confusing because this is usually only the case when using the GPL as license
fix #995
