### 2016-10-01 ElektraInitiative/libelektra#995 @markus2330
Thanks for reporting! You are right, it is confusing and it might be an old artifact caused by changing the licence in the beginning of Elektra.
