### 2016-10-01 9a62b5c
doc: make more clear that test cases might fail
update options to avoid failing test cases
see #997
