### 2016-10-01 ElektraInitiative/libelektra#997 @markus2330
Thank you for reporting! As described in
[doc/TESTING.md](http://git.libelektra.org/blob/master/doc/TESTING.md) some tests write to system paths. The document describes some ways how to get all test cases running.

Please report if you could make them running.
