### 2016-10-01 ElektraInitiative/libelektra#997 @mpranj
You can configure with e.g.
```
cmake -DCMAKE_INSTALL_PREFIX=./install -DKDB_DB_SYSTEM=./kdbsystem -DKDB_DB_SPEC=./spec [...]
```
in your build directory. The tests should then run fine.
