### 2016-10-01 ElektraInitiative/libelektra#997 @eschnett
Thank you; this looks good now:
```
99% tests passed, 1 tests failed out of 114

Label Time Summary:
kdbtests    =  23.57 sec (30 tests)
memleak     =  22.77 sec (27 tests)

Total Test time (real) =  24.83 sec

The following tests FAILED:
	 75 - testscr_check_spec (Failed)
Errors while running CTest
make: *** [test] Error 8
```
