### 2016-10-01 ElektraInitiative/libelektra#997 @mpranj
The remaining failing test is most likely due to the path of `KDB_DB_SPEC` , see #1000.
Thanks for bringing this to light!
