### 2016-10-04 3b3b951
doc: fix link to new licence file
related to  #995
