### 2016-10-27 #44 @eschnett
Several of the self-tests were failing; this was reported and resolved in https://github.com/ElektraInitiative/libelektra/issues/997 .
