### 2016-10-27 #44 @eschnett
The licence was difficult to find, since it was in a file with a non-standard name https://github.com/ElektraInitiative/libelektra/issues/995 .
