### 2016-10-27 #44 @eschnett
Regarding the statement of need: I was initially confused as to the purpose of this library. The documentation speaks of "custom configuration files"; depending on context, this can mean many different things. For example, in one context this can refer to configuring some software ahead of time, and then letting the software executing in batch mode. This doesn't seem to be the use case for Elektra, since it offers features to dynamically updating configuration entries. In other cases, different parts of configurations need to be kept consistent, in a way that can be tested or enforced automatically. Elektra offers no features for this, so this doesn't seem to be part of the use case either. I assume now that Elektra is meant to _dynamically_ configure some software, or several loosely related software installs. It would be good to clarify this, e.g. with an example, in the introduction or the summary.
