### 2016-10-27 #44 @danielskatz
I assume the license issue has been resolved since you checked the box?
