### 2016-10-27 #44 @eschnett
Yes, the licence issue has been remedied.
