### 2016-10-27 #44 @eschnett
It's resolved for the overall documentation, but the paper could be clearer. I didn't check the box there yet.
