### 2016-10-27 ElektraInitiative/libelektra#1055 @eschnett
Tutorial doesn't work
I'm following the tutorial on `https://github.com/ElektraInitiative/libelektra`. I'm using macOS Sierra. I downloaded, configured, built, and installed libelektra without problems.

This is what I typed, and the responses I received:

```
$ kdb export user/hello ini
eschnett@Redshift:~ (15:53:24)
$ kdb export user/hello simpleini
The command export terminated unsuccessfully with the info:
Was not able to load such a plugin!

Maybe you misspelled it, there is no such plugin or the loader has problems.
You might want to try to set LD_LIBRARY_PATH, use kdb-full or kdb-static.
Errors/Warnings during loading were:
1 Warning was issued:
 Warning number: 1
    Description: could not load module, dlopen failed
    Ingroup: modules
    Module: dl
    At: /Users/eschnett/src/libelektra/src/libs/loader/dl.c:86
    Reason: of module: libelektra-simpleini.so, because: dlopen(libelektra-simpleini.so, 2): image not found
    Mountpoint:
    Configfile:
```
- Note that my first command didn't lead to any output. The tutorial shows output `hello_world = ...`.
- The second command fails.
- The error suggests running `kdb-full` or `kdb-static` instead, but no such file was built or installed.
