### 2016-10-27 #44 @eschnett
While testing functionality I've encountered https://github.com/ElektraInitiative/libelektra/issues/1055 .
