### 2016-10-27 ElektraInitiative/libelektra#1055 @markus2330
Thanks for trying the tutorial! The second problem you encountered is that `simpleini` is not portable and does not work with Mac OS X. Thus it is disabled during compilation on your computer and not available at run-time. We plan to implement #701 to have a plugin with identical features on Mac OS X, too.

For the first problem the tutorial seems to be incomplete, can you point me to which tutorial you followed? To get the expected output you would need a `kdb set user/hello/hello_world ...` before.
