### 2016-10-27 #44 @markus2330
Thank you for your effort and reporting helpful issues! I slightly rephrased the paper to clarify that it is about dynamic configuration (at run-time) and added a statement in `doc/WHY.md` about which user groups we target.
