### 2016-10-28 ElektraInitiative/libelektra#1055 @markus2330
Ahh, now I understand! That is an automatic preview of the ascii screencast. The screenshot is not self-contained but needs the already mentioned `kdb set user/hello/hello_world ...` before. (As shown in the ascii screencast [when you click on play](https://asciinema.org/a/cantr04assr4jkv8v34uz9b8r).)

For the next video I will try to get a preview that shows something self-contained. The software asciicast seems to pick a screenshot from somewhere in the middle...
