### 2016-11-01 ElektraInitiative/libelektra#1055 @eschnett
I am looking at the demo, and tried to follow along. This doesn't work for me:
- The first command in the movie doesn't output anything. For me, I get a long list of error messages, ending with
`
IMPORTANT: Make sure you can write to system namespace
           Usually you need to be root for that!
`
- At about second 28, the command `kdb file user/hello/hello_world` outputs a file name. This is a file `hello.ini` in the demo. For me, this is a file `default.ecf`.

I've also looked for actual tutorials that would be easier to follow along than a movie. I've found a link `tutorial` on the main readme. However, this only leads to the page `https://github.com/ElektraInitiative/libelektra/tree/master/doc/tutorials` which holds a list of tutorials without further explanation. None of these seem to do something similar to the demo. I then looked at `validation.md`, but this seems to be an advanced tutorial that expects some background. The commands shown there don't work when I just type them in.

By way of explanation: I'm looking for a "follow-along" tutorial where I can read some text and type in some commands to explore the features offered by Elektra. The tutorials you're offering seem to be different; they probably assume some familiarity with Elektra, and then teach how to perform a particular task.
