### 2016-11-01 #44 @eschnett
libelektra is a large project, and I haven't yet found a way to explore it. I don't want to write (or "elektrify") one of my applications just to test it. Do you have a toy code to play with? Or a tutorial with commands to follow along? As is, I have trouble exploring libelektra's functionality.
