+++ doc/tutorials/README.md
@@ -1,0 +1,36 @@
+Elektra has many different aspects to explore.
+Not all parts are needed by everyone.
+In this document we classify which parts should
+be read by whom.
+
+## General Information
+
+Read this first to get the basic concepts of Elektra.
+
+- [Namespaces](namespaces.md)
+- [Cascading](cascading.md)
+
+## Developers
+
+For these tutorials we assume you want to elektrify your
+application, that means, you want your application
+to participate in the global key database Elektra
+provides.
+
+- [Integration of your Application](application-integration.md)
+- [Plugins Introduction](plugins.md)
+- [Compilation Variants of plugins](compilation-variants.md) (advanced topic)
+- [python-kdb](python-kdb.md)
+
+## System Administrator
+
+For these tutorials we assume that you want to
+work with the configuration of applications
+already somehow integrated with Elektra.
+
+- [export-dump](export-dump.md)
+- [import](import.md)
+- [export](export.md)
+- [merge](merge.md)
+- [elektra-merge-integration](elektra-merge-integration.md)
+- [validation](validation.md)
