### 2016-11-03 31d1448
kdb-mount: improve error text
see #1055
