### 2016-11-04 #44 @eschnett
I have examined libelektra as submitted to the Journal of Open Science. I find that this submission meets or exceeds the requirements of the journal in almost all categories, with two exceptions:

1. The software paper does not "clearly state what problems the software is designed to solve". In fact, this is one of the major issues with libelektra's documentation -- there is much documentation, but there are virtually no introductions. Much of the documentation remains opaque without context.

For example, the documentation speaks of "configuration" without stating exactly what the term means, and without giving examples. The "validation" tutorial does not state what validation is, only that a "configuration" can be "non-validating". The "compilation variants" tutorial begins with discussing "creat[ing] different variants of the same feature", without stating what a "variant" or a "feature" is.

2. I was not able to confirm functionality or performance of this software. libelectra is a large project with many plugins, and it would be beyond the scope of such a review to test all parts of the software. However, I find that even a cursory examination of libelektra's functionality would have required a significant time investment from me, as there are no simple tutorials that would allow getting a first impression of libelektra.

There are only
(a) A video demonstration that seems slightly outdated (commands give different output from me than they do in the movie)
(b) An "application integration tutorial" that would have me modifying an existing application. However, there is no simple example application provided; instead, I would have to engage in a real programming project and learn libelektra's API
(c) A folder with about 15 tutorials that all seem to cover advanced topics such as e.g. "Validation"; none is a "first steps" tutorial.

After skimming much of the documentation, I conclude that libelektra is about configuring applications in the sense of managing user settings in a Gnome or KDE desktop environment for the various graphical applications there. While some settings are application specific, others are more general and apply to multiple applications, or to all of them. Configuration settings are seen as key-value pairs, and managing both the possible keys as well as the user's choices for the values is a complex task. libelektra addresses this. Its features are designed for this use case, and seem to include editing keys and values at run time, accessing keys and values from applications, managing a hierarchical key space, etc. One of the major features of libelektra seems to be that it can encompass existing configuration management software, including existing configuration files in various formats. libelektra claims to unify many existing mechanisms, instead of providing a new one.

I really wish that such a text could be found in libelektra's documentation.

I conclude that this project is not friendly for outsiders, as they have no starting point. I want to emphasize that the feedback I received on the libelektra bugtracker was quick and exceptionally friendly. In this review, will leave the boxes for "functionality" and "performance" in the functionality category, as well as for "statement of need" in the "paper" category unchecked.
