### 2016-11-04 #44 @danielskatz
Thanks @eschnett.  If there are no performance claims made, you can check the performance box.  If there are claims made that you do not find convincing, that's a different story.

@markus2330, can you address these concerns?  It seems like some introductory material is needed in the software paper, and that the software's documentation needs some content to make it more accessible to new users, such as is mentioned in @eschnett's point 2 above.
