### 2016-11-04 #44 @eschnett
@danielskatz Yes, no particular performance claims are made -- I have checked the "performance" box.
