### 2016-11-04 #44 @markus2330
Yes, I can address the concerns. We are currently writing a verification tool that checks that all commands in the tutorials produce the correct output (do not fail) and rewrite the tutorials to have a nice flow and assume less prior knowledge. Furthermore we develop a web page which makes navigation/reading of the tutorials easier.

Are there any time constraints when this should be finished? I am currently at a conference, thus I was not responsive the last days.
