### 2016-11-04 #44 @danielskatz
I don't think there's a real hurry to this, other than that the longer it takes to update, the longer it will be before the paper can be accepted and published.
