### 2016-11-05 42d786f
kdb mount: improve error message
