### 2016-11-05 ElektraInitiative/libelektra#1055 @markus2330
> IMPORTANT: Make sure you can write to system namespace Usually you need to be root for that!

Yes, mounting writes to `/etc/kdb/elektra.ecf` (or whatever `kdb file system/elektra/mountpoints` outputs) and thus you need to be root or follow the instructions given [here](git.libelektra.org/blob/master/doc/TESTING.md). I tried to improve the error message in 42d786f63078fe6294ee02fb18f6addc369224a8.

> At about second 28, the command kdb file user/hello/hello_world outputs a file name. This is a file hello.ini in the demo. For me, this is a file default.ecf.

This is expected because the mounting did not work. `default.ecf` is used when there is no mountpoint.

> which holds a list of tutorials without further explanation

I added a [explanation](https://github.com/ElektraInitiative/libelektra/tree/master/doc/tutorials/README.md).

> they probably assume some familiarity with Elektra

Yes, unfortunately they currently do. We will inform you when we fixed it.
