### 2016-11-22 ElektraInitiative/libelektra#1055 @markus2330
Thanks for your great feedback! We improved the tutorials, now we have tutorials suitable without prior familiarity, in particular: http://tree.libelektra.org/doc/tutorials/mount.md
