### 2016-11-22 #44 @markus2330
Together with our latest release [0.8.19](https://github.com/ElektraInitiative/libelektra/releases) I updated the paper and we created a tutorial that does not assume prior familiarity: http://tree.libelektra.org/doc/tutorials/mount.md
