### 2016-11-22 #44 @danielskatz
@eschnett, back to you for updated review comments...
