### 2016-12-05 #44 @danielskatz
ping: @eschnett
