### 2016-12-12 ElektraInitiative/libelektra#1189 @eschnett
Problem with tutorials
I am looking a libelektra, in particular at the tutorials. I want to go through the tutorials to learn some first steps. The main README points me to https://github.com/ElektraInitiative/libelektra/tree/master/doc/tutorials .

This page suggests to go through three particular tutorials: Namespaces, Cascading, and Mount Configuration Files. Here I provide some feedback on these tutorials:

In the "Namespaces" tutorial, there are graphics that are not displayed by my browser, e.g. the figure after the "kdb set" example, or the figure labelled "At this point the key database should have this structure".

The tutorial requires me to use `sudo`. I would much prefer a tutorial where this was not the case; I don't like evaluating software with root permission. In this case, simply changing the examples should work.

I also see that the tutorials create configuration settings, but they don't clean up. I am left with useless configuration settings. After looking around in my home directory, I concluded that deleting `~/.config/default.ecf` should do the trick.

I did not enter any of the sudo commands, since this seems much too dangerous -- any error I make in the tutorial could destroy my laptop's configuration. For example, playing with `/etc/hosts` could destroy my network settings.

Maybe a "sandbox mode" would be convenient, if a tutorial that changes important system configuration settings is important?

I don't understand why mounting `/etc/hosts` requires `sudo` privileges. Why can't I mount it read-only? Presumably, many applications would run without `sudo` privileges, and still want to access system configuration files. This should be explained in the tutorial.
