### 2016-12-12 #44 @eschnett
@markus2330 The documentation is now much improved, thank you. However, the tutorials themselves are not quite there yet. I've provided some feedback in https://github.com/ElektraInitiative/libelektra/issues/1189 . Since a tutorial is not required by JOSS, this feedback won't affect my review.

@danielskatz I have examined the functionality of libelektra in a cursory manner. The library offers a large API and script interface, and I have looked only at certain basic parts of it. Does this suffice to cover "Functionality: Have the functional claims of the software been confirmed?"?
