### 2016-12-12 ElektraInitiative/libelektra#1189 @markus2330
Thank you for your detailed analysis! We really appreciate your help, but it will take some time to thoroughly fix the issues.

> In the "Namespaces" tutorial, there are graphics that are not displayed by my browser

This will be resolved with the webpage we are developing:
https://web.libelektra.org/tutorials/namespaces
It shows the svg images and also renders everything else nicer.

> The tutorial requires me to use sudo. I would much prefer a tutorial where this was not the case; I don't like evaluating software with root permission. In this case, simply changing the examples should work.

This is actually a larger issue. We are thinking about how to mount as non-root (see #1074) but it has wide-reaching consequences with (ironically) severe consequences for security if done in naïve ways. (Mostly because it allows you to inject code using `LD_PRELOAD`.).

Because `kdb mount` needs root anyway we thought modifying `/etc/hosts` is a realistic example (and goes beyond pointless manipulation in different configuration file syntaxes). Once we have `kdb mount` without root, we certainly will also provide examples that do not need `sudo`.

> I also see that the tutorials create configuration settings, but they don't clean up.

Sorry about that. Our tutorial checker currently does not check if the tutorial correctly undid everything (see #1132).

> I concluded that deleting ~/.config/default.ecf should do the trick.

Exactly. Doing `kdb rm` on the keys will also remove the files.

> I did not enter any of the sudo commands, since this seems much too dangerous -- any error I make in the tutorial could destroy my laptop's configuration. For example, playing with /etc/hosts could destroy my network settings.

We use `kdb set` on hosts regularly and the validation should actually not let you do something harmful. But you are right, there is some danger, e.g. one could remove `localhost`. (We actually could avoid even that, but there might actually be usecases to have non-standard localhosts.)
