### 2016-12-13 #44 @arfon
> @arfon will this suffice to meet the JOSS standards?

Yes I think this is acceptable.

If we're ready to move forward with accepting this submission then @markus2330 could you make an archive of the reviewed software in Zenodo/figshare/other service and update this thread with the DOI of the archive? I can then move forward with accepting the submission.
