### 2016-12-13 #44 @arfon
> Thank you for all your efforts! Can I archive the latest released version (0.8.19) or do I have to use the version when the review process started (0.8.17). In the last round @eschnett looked at 0.8.19.

@markus2330 - please archive the latest released version (`0.8.19`)
