### 2016-12-13 #44 @markus2330
Thank you! Here is the DOI
https://doi.org/10.5281/zenodo.200894
links to
https://zenodo.org/record/200894

I also updated the paper.md (some time ago, but more recent than the pdf linked above)

If you are interested on some "failed" attempts to create a DOI you may continue to read...

Zenodo did not allow to create a DOI for an existing release thus I cloned the repo and created a new release for the existing 0.8.19 tag.

I also uploaded the signed tar on figshare:
https://figshare.com/articles/Elektra_0_8_19/4312394
which has the DOI:
https://dx.doi.org/10.6084/m9.figshare.4312394.v1
but unfortunately the page does not look nice ("sorry, we can't preview this file").

And I also created a DOI of the repository where the .tar.gz of the releases are:
https://doi.org/10.5281/zenodo.200884
but this is 91.6 MB and thus less useful.
